### 2025-10-28 638b94d
migrate setup configuration to pyproject.toml
