### 2025-10-28 7c1ef7a
update Dockerfiles and README for CUDA 13.0 and PyTorch 2.9.0
