+++ .github/workflows/python-package.yml
@@ -14,7 +14,7 @@ jobs:
       fail-fast: false
       matrix:
         python-version: ["3.10", "3.11", "3.12"]
-        torch-version: ["2.5.0", "2.8.0"]
+        torch-version: ["2.5.0", "2.9.0"]
         torch-type: ["stable"]
         include:
         - python-version: "3.12"
