### 2025-10-29 79a4023
refactor: fix deprecation warnings and typo in encoders
- Remove module warning, use lazy imports, fix stacklevels
- Rename calc_pariwise_coo_indices → calc_pairwise_coo_indices
- Zero warnings for normal usage, proper warnings for deprecated code
