+++ torchsparsegradutils/tests/test_encoders.py
@@ -11,19 +11,16 @@ import torch
 import yaml
 
 # Import from the new pairwise_encoder module (recommended)
+from torchsparsegradutils.encoders import PairwiseEncoder
 from torchsparsegradutils.encoders.pairwise_encoder import (
-    PairwiseEncoder,
     _gen_coords,
     _gen_coords_nd,
     _gen_offsets,
     _gen_offsets_nd,
     _trim_nd,
+    calc_pairwise_coo_indices,
     calc_pairwise_coo_indices_nd,
-    calc_pariwise_coo_indices,
 )
-
-# Import from deprecated module for backward compatibility tests
-from torchsparsegradutils.encoders.pairwise_voxel_encoder import PairwiseVoxelEncoder
 from torchsparsegradutils.utils.utils import _sort_coo_indices
 
 if torch.__version__ >= (2,):
@@ -207,7 +204,7 @@ def test_gen_offsets(radius, upper, num_channels, channel_voxel_relation, expect
     assert offsets == expected_shifts
 
 
-# Test the calc_pariwise_coo_indices function:
+# Test the calc_pairwise_coo_indices function:
 
 
 # Test invalid inputs:
@@ -223,18 +220,18 @@ def test_gen_offsets(radius, upper, num_channels, channel_voxel_relation, expect
 )
 def test_pariwise_coo_indices_invalid_inputs(radius, volume_shape, dtype, channel_voxel_relation):
     with pytest.raises(ValueError):
-        calc_pariwise_coo_indices(radius, volume_shape, dtype=dtype, channel_voxel_relation=channel_voxel_relation)
+        calc_pairwise_coo_indices(radius, volume_shape, dtype=dtype, channel_voxel_relation=channel_voxel_relation)
 
 
 # Test dtype and devices:
 def test_pairwise_coo_indices_dtype_and_device(indices_dtype, device):
     radius = 1
     volume_shape = (2, 2, 2, 2)
-    indices_dict = calc_pariwise_coo_indices(
+    indices_dict = calc_pairwise_coo_indices(
         radius, volume_shape, diag=True, upper=False, dtype=indices_dtype, device=device
     )
     for k, indices in indices_dict.items():
-        # NOTE: The calc_pariwise_coo_indices function returns regular tensors, not sparse tensors,
+        # NOTE: The calc_pairwise_coo_indices function returns regular tensors, not sparse tensors,
         # so PyTorch's int32->int64 conversion only happens when creating sparse COO tensors
         assert indices.dtype == indices_dtype
         assert indices.device.type == device.type
@@ -242,7 +239,7 @@ def test_pairwise_coo_indices_dtype_and_device(indices_dtype, device):
 
 # Test upper flag:
 def test_pairwise_coo_indices_upper_lower(radius, volume_shape, diag, upper, channel_voxel_relation):
-    indices_dict = calc_pariwise_coo_indices(radius, volume_shape, diag, upper, channel_voxel_relation)
+    indices_dict = calc_pairwise_coo_indices(radius, volume_shape, diag, upper, channel_voxel_relation)
     for k, indices in indices_dict.items():
         if upper is not None:
             if upper and not diag:
@@ -257,7 +254,7 @@ def test_pairwise_coo_indices_upper_lower(radius, volume_shape, diag, upper, cha
 
 # Test all the indices are unique:
 def test_pairwise_coo_indices_unique(radius, volume_shape, diag, upper, channel_voxel_relation):
-    indices_dict = calc_pariwise_coo_indices(radius, volume_shape, diag, upper, channel_voxel_relation)
[…]
