### 2025-10-29 92a051c
bump: update version to 0.2.1 in configuration files
