### 2025-10-29 d42d1b0
update deployment workflow to use 'python -m build'
