### 2025-10-29 212b3eb
fix PyTorch citation in paper.bib
