+++ paper/paper.bib
@@ -5,8 +5,6 @@
   volume    = {32},
   pages     = {8024--8035},
   year      = {2019}
-  % optional:
-  % note    = {arXiv:1912.01703},
 }
 
 @inproceedings{cupy,
