### 2026-01-12 76556d0
perf: use is_coalesced flag in sparse_eye instead of .coalesce()
Since indices are constructed in sorted order with no duplicates,
pass is_coalesced=True to torch.sparse_coo_tensor() instead of
calling .coalesce() after creation. This avoids unnecessary
index scanning and provides a speedup.
