### 2026-01-12 7cae11b
fix: use packaging.version for PyTorch version comparison
The version checks were comparing torch.__version__ (a string like
"2.3.0") with a tuple (2, 4), which raises TypeError in Python 3.

Changes:
- Use packaging.version.parse() for proper semantic version comparison
  in indexed_matmul.py and test_encoders.py
- Add packaging to dependencies
- Add functional tests that mock torch.__version__ and verify
  NotImplementedError is raised on old PyTorch versions
