+++ torchsparsegradutils/tests/test_indexed_matmul.py
@@ -1,11 +1,6 @@
 import pytest
 import torch
 
-if torch.__version__ < (2, 4):
-    pytest.skip(
-        "Skipping test based on nested tensors since an old version of pytorch is used", allow_module_level=True
-    )
-
 from torchsparsegradutils import gather_mm, segment_mm
 
 # Identify Testing Parameters
@@ -92,3 +87,33 @@ def test_gather_mm(device, value_dtype, index_dtype, shapes):
 
     for i in range(N):
         assert torch.allclose(ab[i, :].squeeze(), a[i, :].squeeze() @ b[idx_b[i], :, :].squeeze(), atol=ATOL, rtol=RTOL)
+
+
+def test_segment_mm_raises_on_old_pytorch():
+    """Test that segment_mm raises NotImplementedError on PyTorch < 2.4."""
+    from unittest.mock import patch
+
+    from torchsparsegradutils import indexed_matmul
+
+    a = torch.randn(10, 4)
+    b = torch.randn(2, 4, 3)
+    seglen_a = torch.tensor([5, 5])
+
+    with patch.object(torch, "__version__", "2.3.0"):
+        with pytest.raises(NotImplementedError, match="PyTorch version is too old"):
+            indexed_matmul.segment_mm(a, b, seglen_a)
+
+
+def test_gather_mm_raises_on_old_pytorch():
+    """Test that gather_mm raises NotImplementedError on PyTorch < 2.4."""
+    from unittest.mock import patch
+
+    from torchsparsegradutils import indexed_matmul
+
+    a = torch.randn(3, 4)
+    b = torch.randn(2, 4, 5)
+    idx_b = torch.tensor([0, 1, 0])
+
+    with patch.object(torch, "__version__", "2.3.0"):
+        with pytest.raises(NotImplementedError, match="PyTorch version is too old"):
+            indexed_matmul.gather_mm(a, b, idx_b)
