### 2026-01-13 bdf08db
perf: replace torch.cat([*B]) with reshape in sparse_triangular_solve
