### 2026-01-13 bdd78d9
Add clarifying comment for is_coalesced flag
