### 2026-01-14 748bd6f
perf: optimize _sort_coo_indices with single torch.sort instead of unique+argsort
