### 2026-01-14 3ff6b2f
Fix flaky test failures (#64)
The flakiness was caused by inconsistent RNG seeding across test files and
one statistical test that needed adjustment for CUDA float32 numerical precision.

Changes:
- Add global seed fixture in conftest.py (seed=42, autouse=True)
- Remove 31 @pytest.mark.flaky(reruns=5) markers from test files
- Remove redundant local seed fixtures from test_linear_cg.py and test_distributions.py

SPECIFIC ISSUE: test_native_rsample_forward on CUDA float32
============================================================
This test uses Nagao's (1973) covariance test which is sensitiv
[…]
