+++ torchsparsegradutils/tests/test_sparse_solve.py
@@ -82,12 +82,12 @@ def solve(request):
     return request.param
 
 
-@pytest.mark.flaky(reruns=5)
 def test_solve_forward_routine(layout, solve, device, value_dtype, index_dtype, shapes):
-
     _, A_shape, B_shape, num_zero = shapes
     n = A_shape[0]
-    A, A_dense = make_spd_sparse(n, layout, value_dtype, index_dtype, device, nz=num_zero)
+    A, A_dense = make_spd_sparse(
+        n, layout, value_dtype, index_dtype, device, nz=num_zero
+    )
     B = torch.rand(*B_shape, dtype=value_dtype, device=device)
 
     X_ref = torch.linalg.solve(A_dense, B)
@@ -96,12 +96,14 @@ def test_solve_forward_routine(layout, solve, device, value_dtype, index_dtype, 
     assert torch.allclose(X_test, X_ref, atol=ATOL)
 
 
-@pytest.mark.flaky(reruns=5)
-def test_solve_backward_routine(layout, solve, device, value_dtype, index_dtype, shapes):
-
+def test_solve_backward_routine(
+    layout, solve, device, value_dtype, index_dtype, shapes
+):
     _, A_shape, B_shape, num_zero = shapes
     n = A_shape[0]
-    As1, A_dense = make_spd_sparse(n, layout, value_dtype, index_dtype, device, nz=num_zero)
+    As1, A_dense = make_spd_sparse(
+        n, layout, value_dtype, index_dtype, device, nz=num_zero
+    )
 
     Ad2 = A_dense.detach().clone().requires_grad_()
     As1.requires_grad_()
@@ -120,7 +122,6 @@ def test_solve_backward_routine(layout, solve, device, value_dtype, index_dtype,
     assert torch.allclose(Bd1.grad, Bd2.grad, atol=ATOL)
 
 
-@pytest.mark.flaky(reruns=5)
 def test_linear_cg_kwargs(device, value_dtype, layout):
     """Test sparse_generic_solve with LinearCGSettings kwargs."""
     from torchsparsegradutils.utils.linear_cg import LinearCGSettings
@@ -131,16 +132,20 @@ def test_linear_cg_kwargs(device, value_dtype, layout):
 
     # Test with custom LinearCGSettings
     settings = LinearCGSettings(
-        cg_tolerance=ATOL, max_cg_iterations=500, terminate_cg_by_size=False, verbose_linalg=False
+        cg_tolerance=ATOL,
+        max_cg_iterations=500,
+        terminate_cg_by_size=False,
+        verbose_linalg=False,
     )
 
     X_ref = torch.linalg.solve(A_dense, B)
-    X_test = sparse_generic_solve(A, B, solve=linear_cg, transpose_solve=linear_cg, settings=settings)
+    X_test = sparse_generic_solve(
+        A, B, solve=linear_cg, transpose_solve=linear_cg, settings=settings
+    )
 
     assert torch.allclose(X_test, X_ref, atol=ATOL)
 
 
-@pytest.mark.flaky(reruns=5)
 def test_bicgstab_kwargs(device, value_dtype, layout):
     """Test sparse_generic_solve with BICGSTABSettings kwargs."""
     from torchsparsegradutils.utils.bicgstab import BICGSTABSettings
@@ -153,12 +158,13 @@ def test_bicgstab_kwargs(device, value_dtype, layout):
     settings = BICGSTABSettings(reltol=ATOL, abstol=1e-8, matvec_max=1000, precon=None)
 
     X_ref = torch.linalg.solve(A_dense, B)
-    X_test = sparse_generic_solve(A, B, solve=bicgstab, transpose_solve=bicgstab, settings=settings)
+    X_test = sparse_generic_solve(
[…]
