### 2026-01-14 92ecab6
fix: Fix flaky test failures and centralize test configuration (#64)
This commit addresses test flakiness and standardizes test infrastructure:

1. RNG SEEDING
=============
- Add global seed fixture in conftest.py (seed=42, autouse=True)
- Remove 31 @pytest.mark.flaky(reruns=5) markers from test files
- Remove redundant local seed fixtures:
  - test_linear_cg.py: removed seed fixture
  - test_distributions.py: removed seed fixture
  - test_minres.py: removed seed fixture and random import
  - test_dist_stats_helpers.py: removed 6 torch.manual_seed(42) calls
  - test_integration_pairwise_spars
[…]
