+++ torchsparsegradutils/tests/test_linear_cg.py
@@ -3,24 +3,12 @@ import random
 
 import pytest
 import torch
+from test_config import Tolerances
 
 from torchsparsegradutils.utils.linear_cg import linear_cg
 
-
-# Autouse fixture to seed RNG and restore state after each test
-@pytest.fixture(autouse=True)
-def seed_restore():
-    unlock = os.getenv("UNLOCK_SEED")
-    if unlock is None or unlock.lower() == "false":
-        rng = torch.get_rng_state()
-        torch.manual_seed(0)
-        if torch.cuda.is_available():
-            torch.cuda.manual_seed_all(0)
-        random.seed(0)
-        yield
-        torch.set_rng_state(rng)
-    else:
-        yield
+# Get iterative solver tolerances for float64 (used throughout this file)
+ATOL, RTOL = Tolerances.iterative(torch.float64)
 
 
 # Test basic CG solve for vectors and matrices
@@ -38,16 +26,16 @@ def test_cg():
     # reference solve
     chol = torch.linalg.cholesky(matrix)
     actual = torch.cholesky_solve(rhs.unsqueeze(1), chol).squeeze()
-    assert torch.allclose(solves, actual, atol=1e-3, rtol=1e-4)
-    assert torch.allclose(solves_init, actual, atol=1e-3, rtol=1e-4)
+    assert torch.allclose(solves, actual, atol=ATOL, rtol=RTOL)
+    assert torch.allclose(solves_init, actual, atol=ATOL, rtol=RTOL)
     # multiple RHS
     rhs_mat = torch.randn(size, 50, dtype=torch.float64)
     solves = linear_cg(matrix.matmul, rhs=rhs_mat, max_iter=size)
     init_mat = torch.randn(size, 50, dtype=torch.float64)
     solves_init = linear_cg(matrix.matmul, rhs=rhs_mat, max_iter=size, initial_guess=init_mat)
     actual_mat = torch.cholesky_solve(rhs_mat, chol)
-    assert torch.allclose(solves, actual_mat, atol=1e-3, rtol=1e-4)
-    assert torch.allclose(solves_init, actual_mat, atol=1e-3, rtol=1e-4)
+    assert torch.allclose(solves, actual_mat, atol=ATOL, rtol=RTOL)
+    assert torch.allclose(solves_init, actual_mat, atol=ATOL, rtol=RTOL)
 
 
 # Test CG with tridiagonal outputs
@@ -58,15 +46,21 @@ def test_cg_with_tridiag():
     matrix.div_(matrix.norm()).add_(torch.eye(size, dtype=torch.float64) * 1e-1)
     rhs = torch.randn(size, 50, dtype=torch.float64)
     solves, t_mats = linear_cg(
-        matrix.matmul, rhs=rhs, n_tridiag=5, max_tridiag_iter=10, max_iter=size, tolerance=0, eps=1e-15
+        matrix.matmul,
+        rhs=rhs,
+        n_tridiag=5,
+        max_tridiag_iter=10,
+        max_iter=size,
+        tolerance=0,
+        eps=1e-15,
     )
     chol = torch.linalg.cholesky(matrix)
     actual = torch.cholesky_solve(rhs, chol)
-    assert torch.allclose(solves, actual, atol=1e-3, rtol=1e-4)
+    assert torch.allclose(solves, actual, atol=ATOL, rtol=RTOL)
     eigs = torch.linalg.eigvalsh(matrix)
     for i in range(5):
         approx = torch.linalg.eigvalsh(t_mats[i])
-        assert torch.allclose(eigs, approx, atol=1e-3, rtol=1e-4)
[…]
