### 2026-01-22 #9046 @crvernon
@theo-barfoot

## Update regarding JOSS scope and paper format changes

### What this means for your submission

For full details, please see:
* Our updated scope definition: https://joss.readthedocs.io/en/latest/submitting.html#scope-and-significance
* An example paper in the new format: https://joss.readthedocs.io/en/latest/example_paper.html

cc:  @abhishektiwari
