### 2026-02-03 #9046 @abhishektiwari
Hi @theo-barfoot Circling back on updated scope. Please review updated scope requirements for JOSS as shared above by EiC in comment. Once paper is updated please let me know so we can move ahead.
