### 2026-02-04 #9046 @theo-barfoot
Apologies for the delay and thank you for chasing this. The paper is now updated according to the new scope.
