### 2026-02-05 #9046 @abhishektiwari
@shailesh1729 @maieroli2010 @lbluque, would any of you be willing to review this submission for Journal of Open Source Software (JOSS)? The submission under consideration is "Sparsity-preserving gradient utility tools for PyTorch".

Thanks
Abhi
