### 2026-02-17 #9046 @abhishektiwari
@idoby @Saran-nns @mirca would any of you be willing to review this submission for Journal of Open Source Software (JOSS)? The submission under consideration is "Sparsity-preserving gradient utility tools for PyTorch".

Thanks
Abhi
