### 2026-02-17 #9046 @idoby
> @idoby @Saran-nns @mirca would any of you be willing to review this submission for Journal of Open Source Software (JOSS)? The submission under consideration is "Sparsity-preserving gradient utility tools for PyTorch".
>
> Thanks
> Abhi

Yeah this is literally what I'm doing right now it looks great
