### 2026-02-17 #10072 @editorialbot
[REVIEW]: Sparsity-preserving gradient utility tools for PyTorch
**Submitting author:** @theo-barfoot (<a href="http://orcid.org/0009-0005-2309-5238">Theodore Barfoot</a>)
**Repository:** https://github.com/cai4cai/torchsparsegradutils
**Branch with paper.md** (empty if default branch):
**Version:** 0.2.0
