### 2026-02-20 #10072 @yewentao256
Hi @theo-barfoot, several issues found currently:

1. unit test error

```bash
(review) [yewentao256@nm-automation-h100-standalone-1-preserve torchsparsegradutils]$ pytest -qsx
==================================== test session starts =====================================
rootdir: /home/yewentao256/torchsparsegradutils
configfile: pyproject.toml
testpaths: torchsparsegradutils/tests
plugins: rerunfailures-15.1
collected 5128 items

torchsparsegradutils/tests/test_bicgstab.py ....
torchsparsegradutils/tests/test_cupy_bindings.py ....F

========================================== FAILURES ==========================================
_________________________ test_t2c_and_c2t_coo[cuda:0-int32-float32] _________________________

device = device(type='cuda', index=0)
cupy_bindings_io = (tensor(indices=tensor([[0, 0, 0, 1, 1, 1, 2, 3],
                       [1, 3, 5, 2, 4, 5, 4, 1]]),
       values=ten...at32), <module 'scipy.sparse' from '/home/yewentao256/review/lib64/python3.12/site-packages/scipy/sparse/__init__.py'>)

    def test_t2c_and_c2t_coo(device, cupy_bindings_io):
        x_t, x_c, xsp = cupy_bindings_io
        # torch to COO sparse Cupy
        x_t_coo = x_t.to_sparse_coo()
>       x_c_coo = tsgucupy.t2c_coo(x_t_coo)
                  ^^^^^^^^^^^^^^^^^^^^^^^^^

torchsparsegradutils/tests/test_cupy_bindings.py:81:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
torchsparsegradutils/cupy/cupy_bindings.py:208: in t2c_coo
    data_c = xp.asarray(x_torch.values())
             ^^^^^^^^^^^^^^^^^^^^^^^^^^^^
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

self = tensor([0.1356, 0.6706, 0.5639, 0.2001, 0.9178, 0.2681, 0.1997, 0.4494],
       device='cuda:0')
dtype = None

    def __array__(self, dtype=None):
        if has_torch_function_unary(self):
            return handle_torch_function(Tensor.__array__, (self,), self, dtype=dtype)
        if dtype is None:
>           return self.numpy()
                   ^^^^^^^^^^^^
E           TypeError: can't convert cuda:0 device type tensor to numpy. Use Tensor.cpu() to copy the tensor to host memory first.

../review/lib64/python3.12/site-packages/torch/_tensor.py:1253: TypeError
```

`uv pip install cupy-cuda12x` could solve but then I met another error

```bash
================================== short test summary info ===================================
FAILED torchsparsegradutils/tests/test_cupy_sparse_solve.py::test_cupy_cg_kwargs[cuda:0-float32-coo] - TypeError: cg() got an unexpected keyword argument 'tol'
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! stopping after 1 failures !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
```

Would recommend making a full requirements.txt file for user and fix all unit tests

2. Referrence: it could be strengthened by adding formal citations/URLs for the specific PyTorch issue(s) mentioned and for upstream code sources referenced in the text/acknowledgements.
3. For the benchmark, I don't have a 4090, would H100 could benefit from it?
4. Typo "distringuishing", please check if there are other typos.
