### 2026-02-20 #10072 @idoby
Hello 👋
Thanks for submitting this package and paper to JOSS, it seems polished and very useful. I can tell you I have two papers under review right now that use sparsity-preserving gradients and a non-backprop equivalent to these gradients.
So it's definitely useful and cutting edge!

I've read the paper and it's pretty good! Currently I think it's missing mostly three things: a link to the GH repo, a link to the official docs at RTD, and some code examples to show off the library in a more hands-on context. Other than these three things, I think the paper is well-written and explains the need and the tradeoffs of the software well.
I think the paper could benefit from including some of the benchmark tests from the docs to illustrate the benefits of using sparse matrices/tensors.

I have checked off every item on my review checklist but would like to also take the time (probably next week) to browse the code of the package itself and maybe play with it a little if I find the time, so please don't end the review. However, I do think this paper can be published at JOSS as soon as the error messages above are fixed by the authors.

Great job! 🚀
