### 2026-03-10 cai4cai/torchsparsegradutils#73 @theo-barfoot
Fix CUDA torch to CuPy sparse conversion (use DLPack)
## Context

Raised during JOSS review ([openjournals/joss-reviews#10072](https://github.com/openjournals/joss-reviews/issues/10072)) by @yewentao256.

## Problem

`test_cupy_bindings.py::test_t2c_and_c2t_coo` fails on CUDA because `t2c_coo` (and likely other conversion helpers in `cupy_bindings.py`) calls `xp.asarray(x_torch.values())` on a CUDA tensor, which internally triggers `Tensor.__array__()` to `.numpy()` and crashes:

```
TypeError: can't convert cuda:0 device type tensor to numpy.
Use Tensor.cpu() to copy the tensor to host memory first.
```

### Reproducing

```bash
pytest -qsx torchsparsegradutils/tests/test_cupy_bindings.py::test_t2c_and_c2t_coo
```

## Proposed Fix

Use the **DLPack** protocol for zero-copy GPU tensor exchange instead of going through NumPy:

```python
# Torch → CuPy (CUDA)
cupy.from_dlpack(tensor)

# CuPy → Torch (CUDA)
torch.from_dlpack(cupy_array)
```

This should be applied in `t2c_coo`, `t2c_csr`, `c2t_coo`, `c2t_csr`, and any other conversion functions in `cupy_bindings.py` where `xp.asarray()` is called on a CUDA tensor.

For CPU tensors the existing `xp.asarray()` path can remain (or DLPack can be used uniformly).

## Acceptance Criteria

- [ ] All `test_cupy_bindings.py` tests pass on CUDA
- [ ] Conversions are zero-copy on GPU where possible
- [ ] CPU conversion path still works
