### 2026-03-10 cai4cai/torchsparsegradutils#74 @theo-barfoot
CuPy solver kwarg compatibility (`tol` vs `rtol`/`atol`) across CuPy versions
## Context

Raised during JOSS review ([openjournals/joss-reviews#10072](https://github.com/openjournals/joss-reviews/issues/10072)) by @yewentao256.

## Problem

`test_cupy_sparse_solve.py::test_cupy_cg_kwargs` fails with:

```
TypeError: cg() got an unexpected keyword argument 'tol'
```

This is a **CuPy API version mismatch**: newer versions of CuPy renamed `tol` to `rtol`/`atol` (following SciPy's convention), while older versions use `tol`. The current kwarg mapping logic in `cupy_sparse_solve.py` (`_wrap_iterative_solver`) assumes specific parameter names per backend but does not account for version differences within CuPy itself.

### Reproducing

```bash
pytest -qsx torchsparsegradutils/tests/test_cupy_sparse_solve.py::test_cupy_cg_kwargs
```

## Proposed Fix

Two options (can combine):

1. **Introspect the solver signature at runtime** to determine which kwargs it accepts, then map accordingly (e.g. `inspect.signature(solver_func)` or try/except).
2. **Pin a supported CuPy version range** in `pyproject.toml` optional extras and document it. Tests should skip gracefully with a clear message when the installed CuPy version is outside the supported range.

The existing `_wrap_iterative_solver` in `cupy_sparse_solve.py` already has backend-aware parameter mapping — this just needs to also be **version-aware** for CuPy.

## Acceptance Criteria

- [ ] `test_cupy_sparse_solve.py` passes (or skips gracefully with a clear message for unsupported CuPy versions)
- [ ] Solver kwarg mapping handles both old (`tol`) and new (`rtol`) CuPy API conventions
- [ ] Supported CuPy version range is documented
