### 2026-03-10 cai4cai/torchsparsegradutils#76 @theo-barfoot
Paper: add repo/docs links, code examples, and benchmark summary
## Context

Raised during JOSS review ([openjournals/joss-reviews#10072](https://github.com/openjournals/joss-reviews/issues/10072)) by @idoby.

## Reviewer Feedback

> Currently I think it's missing mostly three things: a link to the GH repo, a link to the official docs at RTD, and some code examples to show off the library in a more hands-on context. [...] I think the paper could benefit from including some of the benchmark tests from the docs to illustrate the benefits of using sparse matrices/tensors.

## Required Changes

### 1. Add explicit links to the paper body

Add a clearly visible link to:
- The GitHub repository: https://github.com/cai4cai/torchsparsegradutils
- The ReadTheDocs documentation: https://torchsparsegradutils.readthedocs.io

These should appear in the paper text (e.g. in the Summary or a dedicated "Availability" section), not just in JOSS metadata.

### 2. Add short code examples

Include 2–3 brief code snippets (10–20 lines total) demonstrating typical usage, e.g.:
- Sparse matrix multiplication with sparse gradient preservation
- Sparse linear system solve (triangular or generic)
- Sparse multivariate normal sampling

These can be placed in a new "Usage Examples" or "Quick Start" section in the paper.

### 3. Add benchmark summary

Include a brief benchmark summary (1–2 sentences or a small table) showing memory/runtime benefits. The full benchmark scripts are already in the repo — just summarise the key takeaway and reference the scripts/docs for full results.

If mentioning specific hardware (e.g. 4090), note that results may vary by GPU. Reviewer @yewentao256 asked about H100 specifically — consider generalising hardware references or noting which GPUs were tested.

## Acceptance Criteria

- [ ] Paper body contains explicit links to GitHub repo and RTD docs
- [ ] Paper includes 2–3 short code examples
- [ ] Paper includes a brief benchmark summary with reference to full results in docs/repo
