### 2026-03-10 cai4cai/torchsparsegradutils#77 @theo-barfoot
Fix typos in paper and docs and run spell check pass
## Context

Raised during JOSS review ([openjournals/joss-reviews#10072](https://github.com/openjournals/joss-reviews/issues/10072)) by @yewentao256.

## Problem

Reviewer flagged the typo **"distringuishing"** (should be "distinguishing") and recommended checking for other typos.

> Typo "distringuishing", please check if there are other typos.

## Required Changes

- [ ] Fix "distringuishing" → "distinguishing" (locate in paper, docs, or code docstrings)
- [ ] Run a spell check pass across `paper/paper.md`, `README.md`, and RST docs in `docs/source/`
- [ ] Fix any other typos found
