### 2026-03-10 cai4cai/torchsparsegradutils#78 @theo-barfoot
Fix invalid DOI and strengthen bibliography references
## Context

Raised during JOSS review ([openjournals/joss-reviews#10072](https://github.com/openjournals/joss-reviews/issues/10072)) by editorialbot and @yewentao256.

## Problems

### 1. Invalid DOI flagged by editorialbot

> `10.5555/3327757.3327857` is INVALID — `10.5555` is a known broken prefix, replace with `https://dl.acm.org/doi/{doi}` in the `url` field.

This DOI corresponds to the PyTorch NeurIPS 2019 paper. The `10.5555` prefix is a pseudo-DOI from ACM that doesn't resolve. If this DOI is still present in `paper.bib`, remove it and use a URL to the official proceedings page instead.

### 2. Missing DOIs (skipped by editorialbot)

The following references had no DOI and editorialbot couldn't find one automatically:
- PyTorch
- CuPy
- JAX
- Fast Graph Representation Learning with PyTorch Geometric
- Solving sparse linear systems in PyTorch

For each, verify whether a DOI exists and add it if so. For software/blog references where no DOI exists, ensure a `url` field is present.

### 3. Add formal citations for PyTorch issues mentioned in text

@yewentao256 recommended strengthening references by "adding formal citations/URLs for the specific PyTorch issue(s) mentioned and for upstream code sources referenced in the text/acknowledgements."

The paper currently references PyTorch issue #41128 inline. Rather than adding GitHub issues as bib entries, consider:
- Keeping inline references as "PyTorch issue #41128" with a footnote URL
- Adding citations for PyTorch sparse/masked tensor documentation where relevant
- Ensuring upstream code sources acknowledged in the Acknowledgements section (pykrylov, cornellius-gp/linear_operator, pytorch-minimize) have proper citations or URLs

## Acceptance Criteria

- [x] Invalid `10.5555` DOI removed or replaced with a working URL
- [x] All references have either a valid DOI or a `url` field
- [x] PyTorch issue references are clearly linked (footnote or inline URL)
- [x] Upstream code sources in Acknowledgements have proper attribution
