### 2026-03-10 #10072 @theo-barfoot
Hi @yewentao256 and @idoby thank you so much for the detailed review and for running the test suite. Really glad to hear this paper has the potential to get published.

As per the editorial guidance, I have opened issues on my repo to address the issues with the code and paper. I will start working to correct these issues now and update when the fixes are merged.

Thanks again!
