### 2026-04-07 8307a41
Address JOSS review feedback on top of latest main
