+++ paper/paper.md
@@ -27,41 +27,39 @@ bibliography: paper.bib
 
 # Summary
 
-The `torchsparsegradutils` package provides differentiable sparse linear-algebra utilities for PyTorch [@pytorch] that preserve sparsity for returned gradients during backpropagation. While PyTorch directly supports sparse tensors, its default semantics treat sparse layouts as storage optimisations rather than a mathematical structure that results in optimising directly for that sparse subspace. Gradients resulting from PyTorch native functions are often dense and incompatible with end-to-end training of models that require fixed sparsity patterns (e.g., sparse covariance/precision structures).
+The `torchsparsegradutils` package provides differentiable sparse linear-algebra utilities for PyTorch [@pytorch] that preserve sparsity in returned gradients during backpropagation. While PyTorch supports sparse tensors, its default dense-equivalent backward semantics can densify gradients and make it difficult to optimise models with fixed sparsity patterns, such as sparse covariance or precision parameterisations.
 
-To address this limitation, we introduce `torchsparsegradutils`. Key features include: (1) memory-efficient sparse-dense matrix multiplication with sparse gradient preservation, (2) sparse triangular and generic linear system solvers, enabling sparse gradients during backpropagation, and multiple algorithmic backends (BICGSTAB, CG, LSMR, MINRES), (3) cross-platform sparse solver wrappers for CuPy [@cupy] and JAX [@jax], (4) sparse multivariate normal distributions with $\boldsymbol{L}\boldsymbol{L}^T$ and $\boldsymbol{L}\boldsymbol{D}\boldsymbol{L}^T$ sparse covariance and precision matrix parameterisations with reparameterised sampling methods, and (5) specialised encoders for spatial neighbourhood relationships in N-dimensional data.
+The package provides sparse-dense matrix multiplication with sparse-gradient preservation, sparse triangular and generic linear system solvers (including BICGSTAB, CG, LSMR, and MINRES backends), optional CuPy [@cupy] and JAX [@jax] solver wrappers, sparse multivariate normal distributions with $\boldsymbol{L}\boldsymbol{L}^T$ and $\boldsymbol{L}\boldsymbol{D}\boldsymbol{L}^T$ parameterisations, and specialised encoders for spatial neighbourhood relationships in N-dimensional data.
+
+The source code is available on GitHub at [https://github.com/cai4cai/torchsparsegradutils](https://github.com/cai4cai/torchsparsegradutils), with full documentation hosted at [https://torchsparsegradutils.readthedocs.io](https://torchsparsegradutils.readthedocs.io).
 
 # Statement of need
 
 Many scientific machine learning models benefit from representing large linear operators (e.g., neighbourhood couplings, precision matrices, sparse Jacobians) using sparse tensors to reduce memory and compute complexity. In high-dimensional settings, such as volumetric medical imaging, dense covariance or precision parameterisations are typically intractable, motivating sparse end-to-end parameterisations.
 
-However, learning these models with gradient-based optimisation requires backpropagation through sparse linear algebra (matrix products, triangular solves, and linear system solves). PyTorch's default sparse semantics are not designed to preserve user-imposed sparsity structure during differentiation (PyTorch issue #87448), which can lead to memory blow-ups and prevent end-to-end optimisation of sparse probabilistic models.
+However, learning these models with gradient-based optimisation requires backpropagation through sparse linear algebra (matrix products, triangular solves, and linear system solves). PyTorch's default sparse semantics are not designed to preserve user-imposed sparsity structure during differentiation ([PyTorch issue #87448](https://github.com/pytorch/pytorch/issues/87448)), which can lead to memory blow-ups and prevent end-to-end optimisation of sparse probabilistic models.
 
[…]
