### 2026-04-07 cai4cai/torchsparsegradutils#75 @theo-barfoot
Issue addressed in https://github.com/cai4cai/torchsparsegradutils/pull/79
