### 2026-04-07 #10072 @theo-barfoot
Hi @abhishektiwari, @yewentao256, and @idoby,

Apologies for the delay, it's been an intense few weeks doing my PhD viva and starting a new job!

I’ve now addressed the reviewer feedback and pushed the fixes.

Summary of changes:
- Fixed the CUDA Torch->CuPy sparse conversion issue and updated the CuPy solver kwarg handling for `tol` / `rtol` / `atol`
- Improved optional dependency handling and installation docs
- Updated the paper to add explicit GitHub/RTD links, short usage examples, and a brief benchmark summary
- Corrected typos and cleaned up the bibliography / reference formatting

The repo issues opened to track these items were:
[#73](https://github.com/cai4cai/torchsparsegradutils/issues/73), [#74](https://github.com/cai4cai/torchsparsegradutils/issues/74), [#75](https://github.com/cai4cai/torchsparsegradutils/issues/75), [#76](https://github.com/cai4cai/torchsparsegradutils/issues/76), [#77](https://github.com/cai4cai/torchsparsegradutils/issues/77), and [#78](https://github.com/cai4cai/torchsparsegradutils/issues/78)

I also re-ran the targeted tests related to the reported CuPy/JAX issues and they now pass.

I'll ask editorialbot to regenerate the paper PDF so the updated manuscript is visible in the thread.

When you have a chance, I’d be grateful if you could take another look at the updated version.
