### 2026-04-10 #10072 @yewentao256
Still some errors in unit test, I think it is more with the env set up.
`uv venv --python 3.12 --seed --managed-python` with a brand new env.

Then `pip install torchsparsegradutils[all]` shows `WARNING: torchsparsegradutils 0.2.1 does not provide the extra 'all'`

After that, `pytest` gets `= 207 failed, 4215 passed, 47 skipped, 140 warnings, 7 errors, 796 rerun in 210.94s (0:03:30) =`

Tested on H200

@theo-barfoot Could you pin a version and tests that works for the steps mentioned above?
