### 2026-04-15 fb6eb6d
test: stabilize JOSS suite for 0.2.2
