+++ torchsparsegradutils/tests/test_jax_sparse_solve.py
@@ -1,5 +1,6 @@
 import pytest
 import torch
+from test_config import DEVICES, INDEX_DTYPES, VALUE_DTYPES, Tolerances
 
 import torchsparsegradutils.jax as tsgujax
 from torchsparsegradutils.utils import convert_coo_to_csr
@@ -15,23 +16,14 @@ else:
 # 90% of GPU:0 to be preallocated by default even for CPU tests
 # CPU tests show GPU memory usage (~500MB) on all available GPUs
 
-# devices
-DEVICES = [torch.device("cpu")]
-if torch.cuda.is_available():
-    DEVICES.append(torch.device("cuda:0"))
-
 # test parameters
 TEST_DATA = [
     ("vector_1d", (12, 12), (12,), 32),
     ("vector_2d", (12, 12), (12, 1), 32),
     ("multi_rhs", (12, 12), (12, 6), 32),
 ]
-INDEX_DTYPES = [torch.int32, torch.int64]
-VALUE_DTYPES = [torch.float32, torch.float64]
 LAYOUTS = [torch.sparse_coo, torch.sparse_csr]
 SOLVERS = [None, cg, bicgstab]
-
-ATOL = 1e-6
 
 
 def data_id(d):
@@ -86,7 +78,6 @@ def solver(request):
     return request.param
 
 
-@pytest.mark.flaky(reruns=5)
 def test_solve_forward_j4t(layout, device, value_dtype, index_dtype, solver, shapes):
     _, A_shape, B_shape, num_zero = shapes
     n = A_shape[0]
@@ -95,10 +86,10 @@ def test_solve_forward_j4t(layout, device, value_dtype, index_dtype, solver, sha
     X_ref = torch.linalg.solve(A_dense, B)
     X_test = tsgujax.sparse_solve_j4t(A_sp, B, solve=solver)
 
-    assert torch.allclose(X_test, X_ref, atol=ATOL)
+    atol, rtol = Tolerances.iterative(value_dtype)
+    assert torch.allclose(X_test, X_ref, atol=atol, rtol=rtol)
 
 
-@pytest.mark.flaky(reruns=5)
 def test_solve_backward_j4t(layout, device, value_dtype, index_dtype, solver, shapes):
     _, A_shape, B_shape, num_zero = shapes
     n = A_shape[0]
@@ -115,11 +106,11 @@ def test_solve_backward_j4t(layout, device, value_dtype, index_dtype, solver, sh
     X2.backward(grad_out)
     nz = A_sp.grad.to_dense() != 0
     # same tolerances as generic solve
-    assert torch.allclose(A_sp.grad.to_dense()[nz], Ad.grad[nz], atol=ATOL)
-    assert torch.allclose(Bd1.grad, Bd2.grad, atol=ATOL)
+    atol, rtol = Tolerances.iterative(value_dtype)
+    assert torch.allclose(A_sp.grad.to_dense()[nz], Ad.grad[nz], atol=atol, rtol=rtol)
+    assert torch.allclose(Bd1.grad, Bd2.grad, atol=atol, rtol=rtol)
 
 
-@pytest.mark.flaky(reruns=5)
 def test_jax_cg_kwargs(device, value_dtype, layout):
     """Test sparse_solve_j4t with CG solver kwargs."""
     n = 10
@@ -130,10 +121,10 @@ def test_jax_cg_kwargs(device, value_dtype, layout):
     X_ref = torch.linalg.solve(A_dense, B)
     X_test = tsgujax.sparse_solve_j4t(A_sp, B, solve=cg, transpose_solve=cg, tol=1e-6, atol=1e-8, maxiter=500)
 
-    assert torch.allclose(X_test, X_ref, atol=ATOL)
[…]
