+++ torchsparsegradutils/tests/test_sparse_solve.py
@@ -2,14 +2,11 @@
 
 import pytest
 import torch
+from test_config import DEVICES, INDEX_DTYPES, VALUE_DTYPES, Tolerances
 
 from torchsparsegradutils.sparse_solve import sparse_generic_solve
 from torchsparsegradutils.utils import bicgstab, convert_coo_to_csr, linear_cg, minres
 from torchsparsegradutils.utils.random_sparse import make_spd_sparse
-
-DEVICES = [torch.device("cpu")]
-if torch.cuda.is_available():
-    DEVICES.append(torch.device("cuda"))
 
 TEST_DATA = [
     # name  A_shape, B_shape, A_num_zero
@@ -18,12 +15,8 @@ TEST_DATA = [
     ("multi_rhs", (12, 12), (12, 6), 32),
 ]
 
-INDEX_DTYPES = [torch.int32, torch.int64]
-VALUE_DTYPES = [torch.float32, torch.float64]
 LAYOUTS = [torch.sparse_coo, torch.sparse_csr]
 SOLVES = [None, linear_cg, bicgstab, minres]
-
-ATOL = 1e-6
 
 
 # Define Test Names:
@@ -82,9 +75,7 @@ def solve(request):
     return request.param
 
 
-@pytest.mark.flaky(reruns=5)
 def test_solve_forward_routine(layout, solve, device, value_dtype, index_dtype, shapes):
-
     _, A_shape, B_shape, num_zero = shapes
     n = A_shape[0]
     A, A_dense = make_spd_sparse(n, layout, value_dtype, index_dtype, device, nz=num_zero)
@@ -93,12 +84,11 @@ def test_solve_forward_routine(layout, solve, device, value_dtype, index_dtype, 
     X_ref = torch.linalg.solve(A_dense, B)
     X_test = sparse_generic_solve(A, B, solve=solve, transpose_solve=solve)
 
-    assert torch.allclose(X_test, X_ref, atol=ATOL)
+    atol, rtol = Tolerances.iterative(value_dtype)
+    assert torch.allclose(X_test, X_ref, atol=atol, rtol=rtol)
 
 
-@pytest.mark.flaky(reruns=5)
 def test_solve_backward_routine(layout, solve, device, value_dtype, index_dtype, shapes):
-
     _, A_shape, B_shape, num_zero = shapes
     n = A_shape[0]
     As1, A_dense = make_spd_sparse(n, layout, value_dtype, index_dtype, device, nz=num_zero)
@@ -115,12 +105,12 @@ def test_solve_backward_routine(layout, solve, device, value_dtype, index_dtype,
     res_ref.backward(grad_output)
     res_test.backward(grad_output)
 
+    atol, rtol = Tolerances.iterative(value_dtype)
     nz_mask = As1.grad.to_dense() != 0.0
-    assert torch.allclose(As1.grad.to_dense()[nz_mask], Ad2.grad[nz_mask], atol=ATOL)
-    assert torch.allclose(Bd1.grad, Bd2.grad, atol=ATOL)
+    assert torch.allclose(As1.grad.to_dense()[nz_mask], Ad2.grad[nz_mask], atol=atol, rtol=rtol)
+    assert torch.allclose(Bd1.grad, Bd2.grad, atol=atol, rtol=rtol)
 
 
-@pytest.mark.flaky(reruns=5)
 def test_linear_cg_kwargs(device, value_dtype, layout):
     """Test sparse_generic_solve with LinearCGSettings kwargs."""
     from torchsparsegradutils.utils.linear_cg import LinearCGSettings
@@ -129,18 +119,18 @@ def test_linear_cg_kwargs(device, value_dtype, layout):
[…]
