### 2026-04-22 baf1562
test: finalize JOSS validation for 0.2.2
