### 2026-04-22 10e2258
chore: update vulnerable test tooling pins
