+++ requirements-ci.txt
@@ -3,8 +3,8 @@ scipy
 jax[cpu]
 isort==6.0.1
 flake8==7.3.0
-black==25.1.0
+black==26.3.1
 parameterized==0.9.0
-pytest==8.4.1
+pytest==9.0.3
 pytest-rerunfailures==15.1
 pyyaml==6.0.2
