### 2026-04-22 #10072 @theo-barfoot
Hi, thanks for flagging this.

There were a bunch of CUDA only tests  (that don't run on CI) that I used for benchmarking and my own testing, some of these were failing due to OOM, or the tests have never passed but I kept them for my own information. I have now improved the CUDA tests and they should now all pass: `4732 passed, 204 skipped, 329 warnings` (tested on my 4090).

The `torchsparsegradutils[all]` wasn't working as I hadn't made a new release.

These issues should now be fixed!

Thanks so much.
