### 2026-04-28 #10072 @yewentao256
`pytest`
`================= 4732 passed, 204 skipped, 329 warnings in 199.16s (0:03:19) =================`

Tested again on H200, works now, thanks!

Also optionally run the benchmark and LGTM

```bash
python sparse_mm_suite.py
📈 Performance Summary:
==================================================

🏆 Fastest configuration:
   Algorithm: sparse_mm
   Time: 411.3μs
   Memory: 483.5MB
```

Only some referrence issues now:
- SuiteSparse Rothberg/cfd2 matrix should be added to refs
- PyTorch issue #87448 should be added
- pykrylov、cornellius-gp/linear_operator、pytorch-minimize in acknowledgement should be added
