### 2026-05-07 9b1754f
Address paper reference feedback
