+++ paper/paper.bib
@@ -35,6 +35,28 @@
   doi     = {10.1038/s41592-019-0686-2}
 }
 
+@article{davis2011university,
+  title     = {The University of Florida Sparse Matrix Collection},
+  author    = {Davis, Timothy A. and Hu, Yifan},
+  journal   = {ACM Transactions on Mathematical Software},
+  volume    = {38},
+  number    = {1},
+  pages     = {1:1--1:25},
+  year      = {2011},
+  doi       = {10.1145/2049662.2049663},
+  url       = {https://dl.acm.org/doi/10.1145/2049662.2049663}
+}
+
+@misc{rothberg1997cfd2,
+  title        = {{Rothberg/cfd2}: {CFD}, symmetric pressure matrix},
+  author       = {Rothberg, Edward},
+  editor       = {Davis, Timothy A.},
+  year         = {1997},
+  howpublished = {SuiteSparse Matrix Collection},
+  url          = {https://sparse.tamu.edu/Rothberg/cfd2},
+  note         = {Matrix ID 805}
+}
+
 @book{saad2003iterative,
   title     = {Iterative Methods for Sparse Linear Systems},
   author    = {Saad, Yousef},
@@ -60,6 +82,40 @@
   volume    = {31},
   year      = {2018},
   url       = {https://arxiv.org/abs/1809.11165}
+}
+
+@misc{gardner2017linearoperator,
+  title        = {{LinearOperator}: Structured linear algebra for {PyTorch}},
+  author       = {Gardner, Jacob R. and Pleiss, Geoff and GPyTorch Contributors},
+  year         = {2017},
+  publisher    = {GitHub},
+  howpublished = {GitHub repository},
+  url          = {https://github.com/cornellius-gp/linear_operator}
+}
+
+@misc{orban2014pykrylov,
+  title        = {{PyKrylov}: Krylov Methods in Pure {Python}},
+  author       = {Orban, Dominique and PyKrylov Developers},
+  year         = {2014},
+  publisher    = {GitHub},
+  howpublished = {GitHub repository},
+  url          = {https://github.com/PythonOptimizers/pykrylov}
+}
+
+@misc{feinman2021pytorchminimize,
+  title     = {Pytorch-minimize: a library for numerical optimization with autograd},
+  author    = {Feinman, Reuben},
+  publisher = {GitHub},
+  year      = {2021},
+  url       = {https://github.com/rfeinman/pytorch-minimize}
+}
+
+@misc{pytorch_sparse_issue_87448,
+  title        = {Semantics of sparse operations clarification: Sparsity of the gradient with respect to a sparse tensor input},
+  author       = {Vercauteren, Tom},
+  year         = {2022},
+  howpublished = {{PyTorch} GitHub issue \#87448},
+  url          = {https://github.com/pytorch/pytorch/issues/87448}
 }
 
 @misc{flaport2020sparse,
