+++ paper/paper.bib
@@ -36,7 +36,7 @@
 }
 
 @article{davis2011university,
-  title     = {The University of Florida Sparse Matrix Collection},
+  title     = {{The University of Florida Sparse Matrix Collection}},
   author    = {Davis, Timothy A. and Hu, Yifan},
   journal   = {ACM Transactions on Mathematical Software},
   volume    = {38},
@@ -76,7 +76,7 @@
 }
 
 @inproceedings{gpytorch,
-  title     = {{GPyTorch}: Blackbox Matrix-Matrix Gaussian Process Inference with {GPU} Acceleration},
+  title     = {{GPyTorch}: Blackbox Matrix-Matrix {Gaussian Process} Inference with {GPU} Acceleration},
   author    = {Gardner, Jacob R. and Pleiss, Geoff and Bindel, David and Weinberger, Kilian Q. and Wilson, Andrew Gordon},
   booktitle = {Advances in Neural Information Processing Systems},
   volume    = {31},
@@ -103,7 +103,7 @@
 }
 
 @misc{feinman2021pytorchminimize,
-  title     = {Pytorch-minimize: a library for numerical optimization with autograd},
+  title     = {{PyTorch}-minimize: a library for numerical optimization with autograd},
   author    = {Feinman, Reuben},
   publisher = {GitHub},
   year      = {2021},
