### 2026-05-17 3e13b3a
docs: address JOSS notation feedback
