### 2026-07-06 efa515c
docs: update masked tensor section and add references and improve AI disclosure
