+++ paper/paper.bib
@@ -118,6 +118,24 @@
   url          = {https://github.com/pytorch/pytorch/issues/87448}
 }
 
+@misc{pytorch_maskedtensor,
+  title        = {{torch.masked}: MaskedTensor documentation},
+  author       = {{PyTorch Contributors}},
+  year         = {2026},
+  howpublished = {{PyTorch} documentation},
+  url          = {https://docs.pytorch.org/docs/stable/masked.html},
+  note         = {Accessed: 2026-07-01}
+}
+
+@misc{pytorch_maskedtensor_sparsity,
+  title        = {{MaskedTensor} sparsity},
+  author       = {{PyTorch Contributors}},
+  year         = {2026},
+  howpublished = {{PyTorch} tutorial},
+  url          = {https://docs.pytorch.org/tutorials/unstable/maskedtensor_sparsity.html},
+  note         = {Accessed: 2026-07-01}
+}
+
 @misc{flaport2020sparse,
   title        = {Solving sparse linear systems in PyTorch},
   author       = {Laporte, Floris},
