### 2026-07-07 6da61e2
Fix JAX test skip when optional dependency is missing
