+++ torchsparsegradutils/tests/test_jax_bindings.py
@@ -1,18 +1,14 @@
-import jax
 import numpy as np
 import pytest
 import torch
 from test_config import DEVICES
 
-import torchsparsegradutils as tsgu
 import torchsparsegradutils.jax as tsgujax
 
 # skip if JAX unavailable
-pytest.importorskip("jax")
+jax = pytest.importorskip("jax")
 if not tsgujax.have_jax:
     pytest.skip("JAX bindings unavailable, skipping jax tests", allow_module_level=True)
-
-import jax.numpy as jnp
 
 
 def _id_device(d):
