### 2026-07-07 faa7606
Fix higher-order sparse solve autograd for #81
