+++ torchsparsegradutils/tests/test_sparse_solve.py
@@ -42,6 +42,48 @@ def solve_id(solve):
     return "default"
 
 
+def _make_differentiable_tridiag_spd(theta, layout):
+    """Create a small sparse SPD matrix with fixed sparsity and differentiable values."""
+    n = theta.numel()
+    device = theta.device
+
+    diag = theta.square() + 2.0
+    off = -0.1 * torch.sigmoid(theta[:-1])
+
+    rows = torch.cat(
+        [
+            torch.arange(n, device=device),
+            torch.arange(n - 1, device=device),
+            torch.arange(1, n, device=device),
+        ]
+    )
+    cols = torch.cat(
+        [
+            torch.arange(n, device=device),
+            torch.arange(1, n, device=device),
+            torch.arange(n - 1, device=device),
+        ]
+    )
+    values = torch.cat([diag, off, off])
+
+    A = torch.sparse_coo_tensor(torch.stack([rows, cols]), values, (n, n)).coalesce()
+    if layout == torch.sparse_csr:
+        A = A.to_sparse_csr()
+    return A, A.to_dense()
+
+
+def _settings_for_higher_order_solve(solve):
+    if solve is linear_cg:
+        from torchsparsegradutils.utils.linear_cg import LinearCGSettings
+
+        return {"settings": LinearCGSettings(cg_tolerance=1e-10, max_cg_iterations=1000)}
+    if solve is minres:
+        from torchsparsegradutils.utils.minres import MINRESSettings
+
+        return {"settings": MINRESSettings(minres_tolerance=1e-10, max_cg_iterations=1000)}
+    return {}
+
+
 # Define Fixtures
 
 
@@ -260,3 +302,59 @@ def test_kwargs_with_different_solvers_same_matrix():
     # All solutions should be close to each other
     assert torch.allclose(X_cg, X_minres, atol=atol, rtol=rtol)
     assert torch.allclose(X_bicgstab, X_minres, atol=atol, rtol=rtol)
+
+
+@pytest.mark.parametrize("layout", LAYOUTS, ids=[layout_id(d) for d in LAYOUTS])
+@pytest.mark.parametrize("solve", [linear_cg, minres], ids=[solve_id(linear_cg), solve_id(minres)])
+def test_sparse_generic_solve_higher_order_create_graph_no_out_error(layout, solve):
+    device = torch.device("cpu")
+    dtype = torch.float64
+    torch.manual_seed(0)
+
+    theta = torch.randn(8, dtype=dtype, device=device, requires_grad=True)
+    A, _ = _make_differentiable_tridiag_spd(theta, layout)
+    B = torch.randn(8, 2, dtype=dtype, device=device)
+
+    kwargs = _settings_for_higher_order_solve(solve)
+    loss = sparse_generic_solve(A, B, solve=solve, transpose_solve=solve, **kwargs).sum()
+
+    grad_theta = torch.autograd.grad(loss, theta, create_graph=True)[0]
+
+    assert grad_theta.requires_grad
+    assert torch.isfinite(grad_theta).all()
+
+    second = torch.autograd.grad(grad_theta.sum(), theta)[0]
+    assert torch.isfinite(second).all()
+
+
+@pytest.mark.parametrize("layout", LAYOUTS, ids=[layout_id(d) for d in LAYOUTS])
+@pytest.mark.parametrize("solve", [linear_cg, minres], ids=[solve_id(linear_cg), solve_id(minres)])
+def test_sparse_generic_solve_higher_order_matches_dense_reference(layout, solve):
+    device = torch.device("cpu")
+    dtype = torch.float64
[…]
