### 2026-07-08 c581e2d
Add non-symmetric BiCGSTAB higher-order test for #85
