+++ torchsparsegradutils/tests/test_sparse_solve.py
@@ -72,16 +72,103 @@ def _make_differentiable_tridiag_spd(theta, layout):
     return A, A.to_dense()
 
 
-def _settings_for_higher_order_solve(solve):
+def _make_differentiable_nonsymmetric_tridiag(theta, layout):
+    """Create a small fixed-sparsity non-symmetric diagonally dominant matrix."""
+    n = 6
+    device = theta.device
+    dtype = theta.dtype
+
+    if theta.numel() != 3 * n - 2:
+        raise ValueError(f"theta should have length {3 * n - 2}, got {theta.numel()}")
+
+    diag_theta = theta[:n]
+    upper_theta = theta[n : n + n - 1]
+    lower_theta = theta[n + n - 1 :]
+
+    diag = diag_theta.square() + 3.0
+    upper = 0.05 * torch.tanh(upper_theta)
+    lower = -0.08 * torch.sigmoid(lower_theta)
+
+    rows = torch.cat(
+        [
+            torch.arange(n, device=device),
+            torch.arange(n - 1, device=device),
+            torch.arange(1, n, device=device),
+        ]
+    )
+    cols = torch.cat(
+        [
+            torch.arange(n, device=device),
+            torch.arange(1, n, device=device),
+            torch.arange(n - 1, device=device),
+        ]
+    )
+    values = torch.cat([diag, upper, lower]).to(dtype=dtype)
+
+    A = torch.sparse_coo_tensor(torch.stack([rows, cols]), values, (n, n)).coalesce()
+    if layout == torch.sparse_csr:
+        A = A.to_sparse_csr()
+    return A, A.to_dense()
+
+
+def _transpose_sparse_for_test(A):
+    """Return A.T in the same sparse layout family for test-only transpose solves."""
+    At = A.to_sparse_coo().transpose(0, 1).coalesce()
+    if A.layout == torch.sparse_csr:
+        return At.to_sparse_csr()
+    return At
+
+
+def _bicgstab_transpose(A, B, **kwargs):
+    """Solve A.T X = B using BiCGSTAB."""
+    return bicgstab(_transpose_sparse_for_test(A), B, **kwargs)
+
+
+def _bicgstab_higher_order_kwargs(value_dtype):
+    from torchsparsegradutils.utils.bicgstab import BICGSTABSettings
+
+    if value_dtype == torch.float32:
+        return {"settings": BICGSTABSettings(reltol=1e-6, abstol=1e-6, matvec_max=1000)}
+    return {"settings": BICGSTABSettings(reltol=1e-12, abstol=1e-12, matvec_max=1000)}
+
+
+def _higher_order_bicgstab_tolerances(value_dtype):
+    if value_dtype == torch.float32:
+        return {
+            "output": (1e-5, 1e-5),
+            "grad": (5e-5, 5e-5),
+            "hess": (5e-3, 5e-3),
+        }
+    return {
+        "output": (1e-8, 1e-8),
+        "grad": (1e-6, 1e-6),
+        "hess": (5e-5, 5e-5),
+    }
+
+
+def _settings_for_higher_order_solve(solve, value_dtype):
     if solve is linear_cg:
         from torchsparsegradutils.utils.linear_cg import LinearCGSettings
 
-        return {"settings": LinearCGSettings(cg_tolerance=1e-10, max_cg_iterations=1000)}
+        return {"settings": LinearCGSettings(cg_tolerance=1e-5, max_cg_iterations=1000)}
     if solve is minres:
         from torchsparsegradutils.utils.minres import MINRESSettings
 
[…]
