### 2026-07-08 aedb894
Address PR #85 sparse solve test review comments
