### 2026-07-09 6839ddd
Fix higher-order sparse solve gradients and optional JAX test skipping (#85)
Fix higher-order autograd support for sparse_generic_solve by routing the backward transpose solve through sparse_generic_solve rather than directly calling the raw iterative solver.

This avoids differentiating through solver internals when create_graph=True, while preserving the implicit-gradient path used by the custom autograd function.

Also:
- preserve the solve output in the autograd context without detaching it;
- add higher-order regression tests for linear_cg, minres, and BiCGSTAB;
- add a non-symmetric BiCG
[…]
