+++ torchsparsegradutils/tests/test_sparse_solve.py
@@ -42,6 +42,132 @@ def solve_id(solve):
     return "default"
 
 
+def _make_differentiable_tridiag_spd(theta, layout):
+    """Create a small sparse SPD matrix with fixed sparsity and differentiable values."""
+    n = theta.numel()
+    device = theta.device
+
+    diag = theta.square() + 2.0
+    off = -0.1 * torch.sigmoid(theta[:-1])
+
+    rows = torch.cat(
+        [
+            torch.arange(n, device=device),
+            torch.arange(n - 1, device=device),
+            torch.arange(1, n, device=device),
+        ]
+    )
+    cols = torch.cat(
+        [
+            torch.arange(n, device=device),
+            torch.arange(1, n, device=device),
+            torch.arange(n - 1, device=device),
+        ]
+    )
+    values = torch.cat([diag, off, off])
+
+    A = torch.sparse_coo_tensor(torch.stack([rows, cols]), values, (n, n)).coalesce()
+    if layout == torch.sparse_csr:
+        A = A.to_sparse_csr()
+    return A, A.to_dense()
+
+
+def _make_differentiable_nonsymmetric_tridiag(theta, layout):
+    """Create a small fixed-sparsity non-symmetric diagonally dominant matrix."""
+    n = 6
+    device = theta.device
+    dtype = theta.dtype
+
+    if theta.numel() != 3 * n - 2:
+        raise ValueError(f"theta should have length {3 * n - 2}, got {theta.numel()}")
+
+    diag_theta = theta[:n]
+    upper_theta = theta[n : n + n - 1]
+    lower_theta = theta[n + n - 1 :]
+
+    diag = diag_theta.square() + 3.0
+    upper = 0.05 * torch.tanh(upper_theta)
+    lower = -0.08 * torch.sigmoid(lower_theta)
+
+    rows = torch.cat(
+        [
+            torch.arange(n, device=device),
+            torch.arange(n - 1, device=device),
+            torch.arange(1, n, device=device),
+        ]
+    )
+    cols = torch.cat(
+        [
+            torch.arange(n, device=device),
+            torch.arange(1, n, device=device),
+            torch.arange(n - 1, device=device),
+        ]
+    )
+    values = torch.cat([diag, upper, lower]).to(dtype=dtype)
+
+    A = torch.sparse_coo_tensor(torch.stack([rows, cols]), values, (n, n)).coalesce()
+    if layout == torch.sparse_csr:
+        A = A.to_sparse_csr()
+    return A, A.to_dense()
+
+
+def _bicgstab_transpose(A, B, **kwargs):
+    """Solve A.T X = B using BiCGSTAB."""
+    if A.layout == torch.sparse_csr:
+        # A.T currently triggers aten::as_strided for sparse CSR tensors in PyTorch,
+        # so use transpose(...).to_sparse_csr() for the CSR test case.
+        # A.transpose(0, 1) for csr will return csc, hence the to_sparse_csr() call.
+        return bicgstab(A.transpose(0, 1).to_sparse_csr(), B, **kwargs)
+    return bicgstab(A.T, B, **kwargs)
+
+
+def _bicgstab_higher_order_kwargs(value_dtype):
+    from torchsparsegradutils.utils.bicgstab import BICGSTABSettings
+
+    if value_dtype == torch.float32:
+        return {"settings": BICGSTABSettings(reltol=1e-6, abstol=1e-6, matvec_max=1000)}
+    return {"settings": BICGSTABSettings(reltol=1e-12, abstol=1e-12, matvec_max=1000)}
+
+
[…]
