### 2026-07-10 05dbc6f
feat: add sparse_logsumexp reduction
Numerically-stable log-sum-exp over the nonzero values of a 2-D sparse
tensor (COO/CSR/CSC), mirroring torch.logsumexp. Reads CSR/CSC indices
directly (no COO coalesce) and reuses the compressed-axis nnz counts to
avoid a redundant bincount. The max-shift is detached so gradients equal
softmax over the participating entries. An include_zeros flag selects
whether structural zeros contribute exp(0)=1 or are treated as -inf.
