+++ torchsparsegradutils/sparse_logsumexp.py
@@ -1,0 +1,188 @@
+from typing import Sequence, Union
+
+import torch
+from torch import Tensor
+
+__all__ = ["sparse_logsumexp"]
+
+
+def _segment_logsumexp(
+    values: Tensor,
+    segment_ids: Tensor,
+    n_segments: int,
+    n_zeros_per_segment: Union[Tensor, None],
+) -> Tensor:
+    r"""Numerically stable log-sum-exp over scattered segments.
+
+    Reduces ``values`` into ``n_segments`` groups, where ``segment_ids[k]`` gives
+    the group of ``values[k]``. Uses the standard max-shift trick so that
+    :math:`\exp` never overflows.
+
+    Parameters
+    ----------
+    values : Tensor, shape ``(nnz,)``
+        The explicit (nonzero) values to reduce.
+    segment_ids : Tensor, shape ``(nnz,)``
+        Output segment index for each value.
+    n_segments : int
+        Number of output segments.
+    n_zeros_per_segment : Tensor or None, shape ``(n_segments,)``
+        Count of structural zeros contributing ``exp(0) = 1`` to each segment, or
+        ``None`` to ignore structural zeros entirely.
+
+    Returns
+    -------
+    Tensor, shape ``(n_segments,)``
+        Per-segment log-sum-exp. Empty segments (no values and no zeros) are ``-inf``.
+    """
+    device, dtype = values.device, values.dtype
+
+    # Per-segment maximum, used purely as a numerical-stability shift. It is
+    # detached: the log-sum-exp is exactly invariant to the shift, so gradients
+    # must not flow through the max (otherwise autograd tracks a spurious path
+    # through argmax). Empty segments start at -inf.
+    max_val = torch.full((n_segments,), float("-inf"), device=device, dtype=dtype)
+    max_val.scatter_reduce_(0, segment_ids, values, reduce="amax", include_self=True)
+    shift = max_val.detach().clone()
+    shift[shift == float("-inf")] = 0.0  # safe shift for empty segments
+
+    # Shifted sum-exp over the explicit values.
+    sum_exp = torch.zeros(n_segments, device=device, dtype=dtype)
+    sum_exp.scatter_reduce_(0, segment_ids, (values - shift[segment_ids]).exp(), reduce="sum", include_self=True)
+
+    # Structural zeros each contribute exp(0 - shift) = exp(-shift).
+    if n_zeros_per_segment is not None:
+        sum_exp = sum_exp + n_zeros_per_segment.to(dtype) * (-shift).exp()
+
+    # Un-shift. Segments with no contribution at all (sum_exp == 0) stay -inf.
+    result = shift + sum_exp.log()
+    return torch.where(sum_exp == 0.0, torch.full_like(result, float("-inf")), result)
+
+
+def _row_col_val(input: Tensor, nrows: int, ncols: int):
+    """Per-nnz ``(rows, cols, vals, row_nnz, col_nnz)`` for a 2-D sparse tensor.
+
+    CSR/CSC expose their explicit values without duplicate cells, so we read the
+    index arrays directly and only uncompress the compressed axis — avoiding the
+    full sort that ``to_sparse_coo().coalesce()`` would pay for. COO may carry
+    duplicate coordinates, so it is coalesced to merge them first.
+
+    ``row_nnz`` / ``col_nnz`` are the per-row / per-column nonzero counts when the
+    layout yields them for free (the compressed axis' segment lengths), else
+    ``None`` — letting the caller skip a redundant ``bincount`` on that axis.
+    """
+    if input.layout == torch.sparse_csr:
+        vals = input.values()
[…]
