### 2026-07-10 9609754
bench: add sparse_logsumexp benchmark
Compares sparse_logsumexp against a dense torch.logsumexp baseline and a
segment_reduce variant across sizes/densities, on CPU and CUDA, with
graceful handling of dense out-of-memory at large sizes.
