+++ torchsparsegradutils/benchmarks/sparse_logsumexp_rand.py
@@ -1,0 +1,173 @@
+#!/usr/bin/env python3
+"""
+Sparse Log-Sum-Exp Benchmark - Random Matrices
+
+Benchmarks ``sparse_logsumexp`` against a dense ``torch.logsumexp`` baseline on
+randomly generated sparse matrices of various sizes, sparsity patterns, layouts
+and dtypes. The reduction is over ``dim=1`` (one value per row); forward and
+backward time and peak memory are measured. The dense baseline materialises
+``A.to_dense()`` and is the memory ceiling the sparse path exists to avoid.
+"""
+
+import os
+import sys
+
+# Add the parent directory to sys.path to allow importing torchsparsegradutils
+sys.path.insert(0, os.path.join(os.path.dirname(__file__), "..", ".."))
+
+import numpy as np
+import torch
+from benchmark_utils import (
+    measure_op,
+    print_benchmark_header,
+    print_result_row,
+    print_results_table_header,
+    save_benchmark_results,
+)
+from tqdm import tqdm
+
+from torchsparsegradutils import sparse_logsumexp
+from torchsparsegradutils.utils import rand_sparse
+
+REPEATS = 100
+WARMUP_RUNS = 10
+
+# Only run on CUDA
+device = torch.device("cuda")
+assert torch.cuda.is_available(), "This benchmark requires a CUDA GPU"
+
+# problem sizes: (label, N, M, nnz)
+SIZES = [
+    ("small", 2**10, 2**10, 2**12),
+    ("medium", 2**12, 2**12, 2**14),
+    ("large", 2**14, 2**14, 2**16),
+]
+
+INDEX_DTYPES = [torch.int32, torch.int64]
+VALUE_DTYPES = [torch.float32, torch.float64]
+LAYOUTS = [torch.sparse_coo, torch.sparse_csr]
+
+# reduce over dim=1; measure_op takes a binary op, so B is an unused placeholder.
+ALGORITHMS = [
+    ("sparse_logsumexp", lambda A, B: sparse_logsumexp(A, dim=1)),
+    ("dense.logsumexp", lambda A, B: torch.logsumexp(A.to_dense(), dim=1)),
+]
+
+
+def run_sparse_logsumexp_benchmark():
+    """Run the sparse log-sum-exp benchmark suite."""
+
+    print_benchmark_header("Sparse Log-Sum-Exp Benchmark - Random Matrices")
+
+    records = []
+
+    for size_label, N, M, nnz in tqdm(SIZES, desc="Problem sizes"):
+        print(f"\n🔍 Testing size: {size_label} (N={N}, M={M}, nnz={nnz})")
+
+        A_shape = (N, M)
+        B = torch.zeros(1, device=device)  # unused placeholder for measure_op's binary signature
+
+        for idx_dt in tqdm(INDEX_DTYPES, desc="Index dtypes", leave=False):
+            for val_dt in tqdm(VALUE_DTYPES, desc="Value dtypes", leave=False):
+                for layout in tqdm(LAYOUTS, desc="Layouts", leave=False):
+                    layout_name = "coo" if layout == torch.sparse_coo else "csr"
+                    print(f"\n  📊 Configuration: idx_dtype={idx_dt}, val_dtype={val_dt}, layout={layout_name}")
+
+                    print_results_table_header()
+
+                    for alg_name, alg_fn in ALGORITHMS:
+                        try:
+                            print(f"    🧮 Testing {alg_name} ({layout_name})...")
+
+                            A_sparse = rand_sparse(
+                                A_shape, nnz, layout, indices_dtype=idx_dt, values_dtype=val_dt, device=device
+                            )
+
+                            (
+                                t_fwd,
+                                std_fwd,
+                                mem_fwd,
+                                std_mem_fwd,
[…]
