### 2026-07-10 0bb3931
docs: document sparse_logsumexp in API reference and README
