### 2026-07-10 36dbd11
test: cover sparse_logsumexp against dense reference
Parametrized over COO/CSR layouts, devices, dtypes, all dim modes and
both include_zeros settings; cross-checks against torch.logsumexp on the
densified tensor. Adds keepdim, CSC, empty row/column, error-path and
gradient tests, and registers the module for doctest execution.
