+++ torchsparsegradutils/tests/test_sparse_logsumexp.py
@@ -1,0 +1,139 @@
+import pytest
+import torch
+from test_config import DEVICES, SPARSE_LAYOUTS, VALUE_DTYPES, Tolerances
+
+from torchsparsegradutils import sparse_logsumexp
+
+# dim arguments to exercise: per-row, per-column, and full reduction.
+DIMS = [0, 1, [0, 1]]
+
+
+def dtype_id(dtype):
+    return str(dtype).split(".")[-1]
+
+
+def device_id(device):
+    return str(device)
+
+
+def layout_id(layout):
+    return "coo" if layout == torch.sparse_coo else "csr"
+
+
+def dim_id(dim):
+    return f"dim{dim}"
+
+
+def _to_layout(dense, layout):
+    if layout == torch.sparse_coo:
+        return dense.to_sparse_coo()
+    if layout == torch.sparse_csr:
+        return dense.to_sparse_csr()
+    if layout == torch.sparse_csc:
+        return dense.to_sparse_csc()
+    raise ValueError(layout)
+
+
+def _dense_reference(dense, dim, include_zeros):
+    """Dense torch.logsumexp reference honouring the include_zeros semantics."""
+    if not include_zeros:
+        dense = dense.masked_fill(dense == 0, float("-inf"))
+    return torch.logsumexp(dense, dim=tuple(dim) if isinstance(dim, list) else dim)
+
+
+def _assert_close(actual, expected, dtype):
+    """allclose that also treats matching -inf entries as equal."""
+    atol, rtol = Tolerances.direct(dtype)
+    inf_mask = torch.isinf(expected)
+    assert torch.equal(torch.isinf(actual), inf_mask), f"inf pattern mismatch:\n{actual}\n{expected}"
+    assert torch.allclose(actual[~inf_mask], expected[~inf_mask], atol=atol, rtol=rtol), f"{actual}\n!=\n{expected}"
+
+
+@pytest.fixture(params=SPARSE_LAYOUTS, ids=[layout_id(x) for x in SPARSE_LAYOUTS])
+def layout(request):
+    return request.param
+
+
+@pytest.fixture(params=DEVICES, ids=[device_id(d) for d in DEVICES])
+def device(request):
+    return request.param
+
+
+@pytest.fixture(params=VALUE_DTYPES, ids=[dtype_id(d) for d in VALUE_DTYPES])
+def value_dtype(request):
+    return request.param
+
+
+@pytest.fixture(params=DIMS, ids=[dim_id(d) for d in DIMS])
+def dim(request):
+    return request.param
+
+
+@pytest.fixture(params=[True, False], ids=["incl_zeros", "excl_zeros"])
+def include_zeros(request):
+    return request.param
+
+
+def _make_dense(device, value_dtype, seed):
+    """A 5x4 tensor with a deterministic mix of zeros, incl. an all-zero row/col."""
+    g = torch.Generator().manual_seed(seed)
+    dense = torch.randn(5, 4, generator=g, dtype=torch.float64)
+    dense = dense.masked_fill(torch.rand(5, 4, generator=g) < 0.5, 0.0)
+    dense[2] = 0.0  # all-zero row -> empty segment when reducing dim=1
+    dense[:, 3] = 0.0  # all-zero column -> empty segment when reducing dim=0
+    return dense.to(device=device, dtype=value_dtype)
+
+
+def test_matches_dense_reference(layout, device, value_dtype, dim, include_zeros):
+    dense = _make_dense(device, value_dtype, seed=0)
+    sp = _to_layout(dense, layout)
+    out = sparse_logsumexp(sp, dim=dim, include_zeros=include_zeros)
+    ref = _dense_reference(dense, dim, include_zeros)
[…]
