### 2026-07-10 f9e3b9d
feat: support batched 3-D sparse_logsumexp
Extend the reduction to batched 3-D (batch, rows, cols) inputs, matching
the batched convention of sparse_mm and the solvers. Each slice is reduced
independently by folding the batch index into the scatter segment id, so a
single scatter handles the whole batch; the batch axis itself is not
reducible. Batched inputs go through COO (batched CSR/CSC require equal nnz
per slice in PyTorch). Adds batched tests across dims/include_zeros/keepdim.
