+++ torchsparsegradutils/sparse_logsumexp.py
@@ -37,16 +37,12 @@ def _segment_logsumexp(
     """
     device, dtype = values.device, values.dtype
 
-    # Per-segment maximum, used purely as a numerical-stability shift. It is
-    # detached: the log-sum-exp is exactly invariant to the shift, so gradients
-    # must not flow through the max (otherwise autograd tracks a spurious path
-    # through argmax). Empty segments start at -inf.
+    # Per-segment max, detached — a stability shift the result is invariant to.
     max_val = torch.full((n_segments,), float("-inf"), device=device, dtype=dtype)
     max_val.scatter_reduce_(0, segment_ids, values, reduce="amax", include_self=True)
     shift = max_val.detach().clone()
-    shift[shift == float("-inf")] = 0.0  # safe shift for empty segments
+    shift[shift == float("-inf")] = 0.0  # empty segments
 
-    # Shifted sum-exp over the explicit values.
     sum_exp = torch.zeros(n_segments, device=device, dtype=dtype)
     sum_exp.scatter_reduce_(0, segment_ids, (values - shift[segment_ids]).exp(), reduce="sum", include_self=True)
 
@@ -56,7 +52,7 @@ def _segment_logsumexp(
 
     # Un-shift. Segments with no contribution at all (sum_exp == 0) stay -inf.
     result = shift + sum_exp.log()
-    return torch.where(sum_exp == 0.0, torch.full_like(result, float("-inf")), result)
+    return torch.where(sum_exp == 0.0, float("-inf"), result)
 
 
 def _row_col_val(input: Tensor, nrows: int, ncols: int):
@@ -73,17 +69,81 @@ def _row_col_val(input: Tensor, nrows: int, ncols: int):
     """
     if input.layout == torch.sparse_csr:
         vals = input.values()
-        row_nnz = input.crow_indices()[1:] - input.crow_indices()[:-1]
+        crow = input.crow_indices()
+        row_nnz = crow[1:] - crow[:-1]
         rows = torch.repeat_interleave(torch.arange(nrows, device=vals.device), row_nnz)
         return rows, input.col_indices().long(), vals, row_nnz, None
     if input.layout == torch.sparse_csc:
         vals = input.values()
-        col_nnz = input.ccol_indices()[1:] - input.ccol_indices()[:-1]
+        ccol = input.ccol_indices()
+        col_nnz = ccol[1:] - ccol[:-1]
         cols = torch.repeat_interleave(torch.arange(ncols, device=vals.device), col_nnz)
         return input.row_indices().long(), cols, vals, None, col_nnz
     coo = input.coalesce()
     rows, cols = coo.indices()
     return rows, cols, coo.values(), None, None
+
+
+def _logsumexp_2d(input: Tensor, dims, keepdim: bool, include_zeros: bool) -> Tensor:
+    """Reduction for an unbatched 2-D sparse tensor (see :func:`sparse_logsumexp`)."""
+    nrows, ncols = input.shape
+
+    rows, cols, vals, row_nnz, col_nnz = _row_col_val(input, nrows, ncols)
+
+    if dims == [0, 1]:
+        # Reduce everything -> scalar.
+        n_zeros = (
+            torch.tensor([nrows * ncols - vals.numel()], device=vals.device, dtype=vals.dtype)
+            if include_zeros
+            else None
+        )
+        result = _segment_logsumexp(vals, torch.zeros_like(rows), 1, n_zeros).squeeze(0)
+        return result.reshape(1, 1) if keepdim else result
+
+    # dims == [1]: reduce columns -> one value per row.
+    # dims == [0]: reduce rows    -> one value per column.
+    if dims == [1]:
[…]
