### 2026-07-10 12ba1d6
refactor: rename segment->scatter in logsumexp helper
The helper reduces via Tensor.scatter_reduce_ with an arbitrary per-value
output index, so 'segment' (which implies contiguous runs, as in
torch.segment_reduce) was misleading. Rename to _scatter_logsumexp /
scatter_index and note in the docstring that group members need not be
contiguous.
