+++ torchsparsegradutils/sparse_logsumexp.py
@@ -6,51 +6,53 @@ from torch import Tensor
 __all__ = ["sparse_logsumexp"]
 
 
-def _segment_logsumexp(
+def _scatter_logsumexp(
     values: Tensor,
-    segment_ids: Tensor,
-    n_segments: int,
-    n_zeros_per_segment: Union[Tensor, None],
+    scatter_index: Tensor,
+    n_groups: int,
+    n_zeros_per_group: Union[Tensor, None],
 ) -> Tensor:
-    r"""Numerically stable log-sum-exp over scattered segments.
+    r"""Numerically stable log-sum-exp over scatter-reduced groups.
 
-    Reduces ``values`` into ``n_segments`` groups, where ``segment_ids[k]`` gives
-    the group of ``values[k]``. Uses the standard max-shift trick so that
-    :math:`\exp` never overflows.
+    Reduces ``values`` into ``n_groups`` groups, where ``scatter_index[k]`` gives
+    the group of ``values[k]``. This is a *scatter* reduction, not a segmented one:
+    ``scatter_index`` is an arbitrary per-value output index (via
+    ``Tensor.scatter_reduce_``), so a group's members need not be contiguous. Uses
+    the standard max-shift trick so that :math:`\exp` never overflows.
 
     Parameters
     ----------
     values : Tensor, shape ``(nnz,)``
         The explicit (nonzero) values to reduce.
-    segment_ids : Tensor, shape ``(nnz,)``
-        Output segment index for each value.
-    n_segments : int
-        Number of output segments.
-    n_zeros_per_segment : Tensor or None, shape ``(n_segments,)``
-        Count of structural zeros contributing ``exp(0) = 1`` to each segment, or
+    scatter_index : Tensor, shape ``(nnz,)``
+        Output group index for each value.
+    n_groups : int
+        Number of output groups.
+    n_zeros_per_group : Tensor or None, shape ``(n_groups,)``
+        Count of structural zeros contributing ``exp(0) = 1`` to each group, or
         ``None`` to ignore structural zeros entirely.
 
     Returns
     -------
-    Tensor, shape ``(n_segments,)``
-        Per-segment log-sum-exp. Empty segments (no values and no zeros) are ``-inf``.
+    Tensor, shape ``(n_groups,)``
+        Per-group log-sum-exp. Empty groups (no values and no zeros) are ``-inf``.
     """
     device, dtype = values.device, values.dtype
 
-    # Per-segment max, detached — a stability shift the result is invariant to.
-    max_val = torch.full((n_segments,), float("-inf"), device=device, dtype=dtype)
-    max_val.scatter_reduce_(0, segment_ids, values, reduce="amax", include_self=True)
+    # Per-group max, detached — a stability shift the result is invariant to.
+    max_val = torch.full((n_groups,), float("-inf"), device=device, dtype=dtype)
+    max_val.scatter_reduce_(0, scatter_index, values, reduce="amax", include_self=True)
     shift = max_val.detach().clone()
-    shift[shift == float("-inf")] = 0.0  # empty segments
+    shift[shift == float("-inf")] = 0.0  # empty groups
 
-    sum_exp = torch.zeros(n_segments, device=device, dtype=dtype)
-    sum_exp.scatter_reduce_(0, segment_ids, (values - shift[segment_ids]).exp(), reduce="sum", include_self=True)
+    sum_exp = torch.zeros(n_groups, device=device, dtype=dtype)
[…]
