### 2026-07-10 b95ab5c
refactor: use torch.logsumexp directly for scalar reduction
For dim=[0,1] every value falls in one group, so torch.logsumexp does the
reduction directly. The structural-zero mass (n_zeros entries of exp(0)=1)
equals a single entry of value log(n_zeros), appended so the reduction stays
one numerically stable call whose gradient still accounts for the zeros.
