### 2026-07-10 3587e21
docs: note why scatter is used over a segmented reduction for CSR/CSC
