### 2026-07-10 70883db
docs: clarify 'slice' means one batch matrix input[k]
