### 2026-07-10 5cd9da9
bench: cover >1M-dim and SuiteSparse matrices for sparse_logsumexp
Extend the random benchmark to 2**16 and 2**20 (>1M rows/cols) where the dense
baseline OOMs, and add sparse_logsumexp_suitesparse.py running on Rothberg/cfd2
(123k x 123k) and Williams/webbase-1M (1M x 1M). Register both in
benchmark_suite.py and track their result CSVs.
