+++ torchsparsegradutils/benchmarks/sparse_logsumexp_suitesparse.py
@@ -1,12 +1,12 @@
 #!/usr/bin/env python3
 """
-Sparse Log-Sum-Exp Benchmark - Random Matrices
+Sparse Log-Sum-Exp Benchmark - SuiteSparse Collection
 
-Benchmarks ``sparse_logsumexp`` against a dense ``torch.logsumexp`` baseline on
-randomly generated sparse matrices of various sizes, sparsity patterns, layouts
-and dtypes. The reduction is over ``dim=1`` (one value per row); forward and
-backward time and peak memory are measured. The dense baseline materialises
-``A.to_dense()`` and is the memory ceiling the sparse path exists to avoid.
+Benchmarks ``sparse_logsumexp`` against a dense ``torch.logsumexp`` baseline on a
+real matrix from the SuiteSparse Matrix Collection (Rothberg/cfd2, ~123k x 123k).
+The reduction is over ``dim=1``; forward and backward time and peak memory are
+measured. The dense baseline materialises ``A.to_dense()`` and is the memory
+ceiling the sparse path exists to avoid -- at this size it OOMs (recorded as NaN).
 """
 
 import os
@@ -18,6 +18,7 @@ sys.path.insert(0, os.path.join(os.path.dirname(__file__), "..", ".."))
 import numpy as np
 import torch
 from benchmark_utils import (
+    load_mat_from_suitesparse_collection,
     measure_op,
     print_benchmark_header,
     print_result_row,
@@ -27,7 +28,6 @@ from benchmark_utils import (
 from tqdm import tqdm
 
 from torchsparsegradutils import sparse_logsumexp
-from torchsparsegradutils.utils import rand_sparse
 
 REPEATS = 100
 WARMUP_RUNS = 10
@@ -35,15 +35,6 @@ WARMUP_RUNS = 10
 # Only run on CUDA
 device = torch.device("cuda")
 assert torch.cuda.is_available(), "This benchmark requires a CUDA GPU"
-
-# problem sizes: (label, N, M, nnz)
-SIZES = [
-    ("small", 2**10, 2**10, 2**12),
-    ("medium", 2**12, 2**12, 2**14),
-    ("large", 2**14, 2**14, 2**16),
-    ("xlarge", 2**16, 2**16, 2**18),
-    ("million", 2**20, 2**20, 2**22),
-]
 
 INDEX_DTYPES = [torch.int32, torch.int64]
 VALUE_DTYPES = [torch.float32, torch.float64]
@@ -55,35 +46,46 @@ ALGORITHMS = [
     ("dense.logsumexp", lambda A, B: torch.logsumexp(A.to_dense(), dim=1)),
 ]
 
+MATRICES = [
+    ("Rothberg", "cfd2"),  # 123k x 123k
+    ("Williams", "webbase-1M"),  # 1M x 1M
+]
 
-def run_sparse_logsumexp_benchmark():
-    """Run the sparse log-sum-exp benchmark suite."""
 
-    print_benchmark_header("Sparse Log-Sum-Exp Benchmark - Random Matrices")
+def run_sparse_logsumexp_suitesparse_benchmark():
+    """Run the sparse log-sum-exp benchmark suite with SuiteSparse matrices."""
+
+    print_benchmark_header("Sparse Log-Sum-Exp Benchmark - SuiteSparse Collection")
 
     records = []
+    B = torch.zeros(1, device=device)  # unused placeholder for measure_op's binary signature
 
-    for size_label, N, M, nnz in tqdm(SIZES, desc="Problem sizes"):
-        print(f"\n🔍 Testing size: {size_label} (N={N}, M={M}, nnz={nnz})")
-
-        A_shape = (N, M)
[…]
