### 2026-07-10 ad236c6
fix: correct stability shift for structural zeros in sparse_logsumexp
With include_zeros=True the structural zeros (value 0) participate in the
reduction, but the max-shift was computed only from the explicit values. When
every explicit value in a group is negative this overflowed (exp(-shift) -> inf),
and a fully dense group (no structural zeros) hit 0 * inf -> nan. Include 0 in
the per-group max where structural zeros are present, and contribute the zero
term only where the count is nonzero. Adds regression tests.

Reported-by: theo-barfoot
