### 2026-07-10 a82425b
fix: handle +inf values in sparse_logsumexp shift
An explicit +inf made the per-group max +inf, so value - shift computed
inf - inf = nan. Replace any non-finite shift (both +inf and the empty-group
-inf) with 0, so a +inf value correctly yields +inf. Adds regression test.

Reported-by: theo-barfoot
