### 2026-07-10 d73d6bf
fix: validate dim before normalising in sparse_logsumexp
dim was reduced modulo ndim before validation, so out-of-range dims silently
wrapped (2 -> 0), a set() dropped duplicates ([0, 0] -> [0]), and an empty
sequence fell through to a wrong path. Validate raw dims first: IndexError for
out-of-range, RuntimeError for empty or repeated dims, matching torch.logsumexp.
Adds regression tests; valid negative dims are unaffected.

Reported-by: theo-barfoot
