### 2026-07-10 16586cf
test: parametrize sparse_logsumexp tests over index dtype
Add an index_dtype fixture (int32/int64) matching the repo convention and thread
it through the dense-reference test so the CSR/CSC index paths are covered for
both dtypes. COO indices are int64-only in PyTorch, so the fixture varies only
the compressed layouts.

Suggested-by: theo-barfoot
