### 2026-07-10 9ec000c
refactor: guard coalesce with is_coalesced in sparse_logsumexp
Skip the redundant coalesce() call when the COO input is already
coalesced, matching the guard pattern used elsewhere in the repo and at
the CSR/CSC fallback in this same file (line 138).

Suggested-by: tvercaut
