+++ torchsparsegradutils/tests/test_sparse_logsumexp.py
@@ -1,0 +1,240 @@
+import pytest
+import torch
+from test_config import DEVICES, INDEX_DTYPES, SPARSE_LAYOUTS, VALUE_DTYPES, Tolerances
+
+from torchsparsegradutils import sparse_logsumexp
+
+# dim arguments to exercise: per-row, per-column, and full reduction.
+DIMS = [0, 1, [0, 1]]
+
+
+def dtype_id(dtype):
+    return str(dtype).split(".")[-1]
+
+
+def device_id(device):
+    return str(device)
+
+
+def layout_id(layout):
+    return "coo" if layout == torch.sparse_coo else "csr"
+
+
+def dim_id(dim):
+    return f"dim{dim}"
+
+
+def _to_layout(dense, layout, index_dtype=torch.int64):
+    if layout == torch.sparse_coo:
+        coo = dense.to_sparse_coo().coalesce()
+        return torch.sparse_coo_tensor(coo.indices().to(index_dtype), coo.values(), coo.shape).coalesce()
+    if layout == torch.sparse_csr:
+        csr = dense.to_sparse_csr()
+        return torch.sparse_csr_tensor(
+            csr.crow_indices().to(index_dtype), csr.col_indices().to(index_dtype), csr.values(), csr.shape
+        )
+    if layout == torch.sparse_csc:
+        csc = dense.to_sparse_csc()
+        return torch.sparse_csc_tensor(
+            csc.ccol_indices().to(index_dtype), csc.row_indices().to(index_dtype), csc.values(), csc.shape
+        )
+    raise ValueError(layout)
+
+
+def _dense_reference(dense, dim, include_zeros):
+    """Dense torch.logsumexp reference honouring the include_zeros semantics."""
+    if not include_zeros:
+        dense = dense.masked_fill(dense == 0, float("-inf"))
+    return torch.logsumexp(dense, dim=tuple(dim) if isinstance(dim, list) else dim)
+
+
+def _assert_close(actual, expected, dtype):
+    """allclose that also treats matching -inf entries as equal."""
+    atol, rtol = Tolerances.direct(dtype)
+    inf_mask = torch.isinf(expected)
+    assert torch.equal(torch.isinf(actual), inf_mask), f"inf pattern mismatch:\n{actual}\n{expected}"
+    assert torch.allclose(actual[~inf_mask], expected[~inf_mask], atol=atol, rtol=rtol), f"{actual}\n!=\n{expected}"
+
+
+@pytest.fixture(params=SPARSE_LAYOUTS, ids=[layout_id(x) for x in SPARSE_LAYOUTS])
+def layout(request):
+    return request.param
+
+
+@pytest.fixture(params=DEVICES, ids=[device_id(d) for d in DEVICES])
+def device(request):
+    return request.param
+
+
+@pytest.fixture(params=VALUE_DTYPES, ids=[dtype_id(d) for d in VALUE_DTYPES])
+def value_dtype(request):
+    return request.param
+
+
+@pytest.fixture(params=INDEX_DTYPES, ids=[dtype_id(d) for d in INDEX_DTYPES])
+def index_dtype(request):
+    return request.param
+
+
+@pytest.fixture(params=DIMS, ids=[dim_id(d) for d in DIMS])
+def dim(request):
+    return request.param
+
+
+@pytest.fixture(params=[True, False], ids=["incl_zeros", "excl_zeros"])
+def include_zeros(request):
+    return request.param
+
+
+def _make_dense(device, value_dtype, seed):
+    """A 5x4 tensor with a deterministic mix of zeros, incl. an all-zero row/col."""
+    g = torch.Generator().manual_seed(seed)
[…]
