+++ torchsparsegradutils/__init__.py
@@ -1,4 +1,5 @@
 from .indexed_matmul import gather_mm, segment_mm
+from .sparse_logsumexp import sparse_logsumexp
 from .sparse_lstsq import sparse_generic_lstsq
 from .sparse_matmul import sparse_mm
 from .sparse_solve import sparse_generic_solve, sparse_triangular_solve
@@ -10,4 +11,5 @@ __all__ = [
     "sparse_triangular_solve",
     "sparse_generic_solve",
     "sparse_generic_lstsq",
+    "sparse_logsumexp",
 ]
