### 2026-07-11 48b0872
feat: add sparse_bidir_logsumexp (simultaneous row + column LSE)
Computes the column-wise (dim=0) and row-wise (dim=1) log-sum-exp of a 2-D or
batched 3-D sparse tensor in a single traversal, instead of two sparse_logsumexp
passes. Both reductions share one extraction feeding a single batched
_scatter_logsumexp via values.expand(2, nnz) (a view); its backward correctly
sums each nonzero's row+col gradient. Supports COO/CSR/CSC, include_zeros,
keepdim (tuple only), and output_layout in {tuple, padded, nested}; padded is
the native (2, G) buffer.

Suggested-by: tvercaut
