+++ torchsparsegradutils/sparse_logsumexp.py
@@ -1,9 +1,10 @@
-from typing import Sequence, Union
+from typing import Literal, Sequence, Union
 
 import torch
+from packaging.version import parse as parse_version
 from torch import Tensor
 
-__all__ = ["sparse_logsumexp"]
+__all__ = ["sparse_logsumexp", "sparse_bidir_logsumexp"]
 
 
 def _scatter_logsumexp(
@@ -164,6 +165,75 @@ def _logsumexp_batched(input: Tensor, dims, keepdim: bool, include_zeros: bool) 
     return result.unsqueeze(keep_ax) if keepdim else result
 
 
+def _bidir_2d(input: Tensor, include_zeros: bool):
+    """Row- and column-wise log-sum-exp of a 2-D sparse tensor in one batched scatter.
+
+    Both reductions run over the *same* nonzero values, so instead of two passes we
+    stack the two directions into a single batched ``_scatter_logsumexp``:
+    ``batch 0`` scatters by column (reducing rows -> ``col_lse``) and ``batch 1``
+    scatters by row (reducing columns -> ``row_lse``). ``values.expand(2, nnz)`` is a
+    view (no copy); its backward correctly *sums* each nonzero's row and column
+    gradient contributions.
+
+    Returns ``(col_lse (ncols,), row_lse (nrows,), padded (2, G))`` where
+    ``G = max(nrows, ncols)`` and the ``padded`` tail (beyond each axis' length) is
+    ``-inf`` (empty groups).
+    """
+    nrows, ncols = input.shape
+    rows, cols, vals, row_nnz, col_nnz = _row_col_val(input, nrows, ncols)
+    G = max(nrows, ncols)
+
+    batched_vals = vals.expand(2, vals.numel())  # (2, nnz) view, no copy
+    batched_idx = torch.stack([cols, rows])  # batch 0 -> col_lse, batch 1 -> row_lse
+
+    if include_zeros:
+        col_nnz = col_nnz if col_nnz is not None else torch.bincount(cols, minlength=ncols)
+        row_nnz = row_nnz if row_nnz is not None else torch.bincount(rows, minlength=nrows)
+        n_zeros = torch.zeros(2, G, device=vals.device, dtype=vals.dtype)  # padded tail stays 0
+        n_zeros[0, :ncols] = nrows - col_nnz.to(vals.dtype)
+        n_zeros[1, :nrows] = ncols - row_nnz.to(vals.dtype)
+    else:
+        n_zeros = None
+
+    padded = _scatter_logsumexp(batched_vals, batched_idx, G, n_zeros)  # (2, G)
+    return padded[0, :ncols], padded[1, :nrows], padded
+
+
+def _bidir_batched(input: Tensor, include_zeros: bool):
+    """Per-slice row- and column-wise log-sum-exp of a batched 3-D sparse tensor.
+
+    Same single batched scatter as :func:`_bidir_2d`, with the batch index folded into
+    the group index (each slice padded to ``G = max(nrows, ncols)`` groups so both
+    directions share one output). Batched inputs go through COO (batched CSR/CSC
+    require equal nnz per slice in PyTorch).
+
+    Returns ``(col_lse (b, ncols), row_lse (b, nrows), padded (b, 2, G))``.
+    """
+    b, nrows, ncols = input.shape
+    coo = input if (input.layout == torch.sparse_coo and input.is_coalesced()) else input.to_sparse_coo().coalesce()
+    bidx, rows, cols = coo.indices()
+    vals = coo.values()
+    G = max(nrows, ncols)
+
+    batched_vals = vals.expand(2, vals.numel())  # (2, nnz) view, no copy
[…]
