### 2026-07-11 ce582fe
test: cover sparse_bidir_logsumexp
Parametrised over COO/CSR/CSC x value/index dtypes x include_zeros (reusing the
sparse_logsumexp fixtures/helpers): tuple equals the two-call baseline and dense
reference; padded/nested layouts agree and pad to -inf; keepdim shapes and its
rejection for non-tuple layouts; the sparse_logsumexp numerical edge cases (very-negative
values, explicit +inf) exercised on *both* axes; batched shapes; gradient parity
with the two-call baseline; and the error paths.
