+++ torchsparsegradutils/tests/test_sparse_bidir_logsumexp.py
@@ -1,0 +1,171 @@
+import pytest
+import torch
+from packaging.version import parse as parse_version
+from test_config import DEVICES, INDEX_DTYPES, SPARSE_LAYOUTS, VALUE_DTYPES
+
+# Reuse the sibling module's helpers/fixtures so the two suites stay in lock-step.
+from test_sparse_logsumexp import (
+    _assert_close,
+    _dense_reference,
+    _make_batched_dense,
+    _make_dense,
+    _to_layout,
+    device_id,
+    dtype_id,
+    layout_id,
+)
+
+from torchsparsegradutils import sparse_bidir_logsumexp, sparse_logsumexp
+
+_NESTED_OK = parse_version(torch.__version__) >= parse_version("2.4")
+
+
+@pytest.fixture(params=SPARSE_LAYOUTS, ids=[layout_id(x) for x in SPARSE_LAYOUTS])
+def layout(request):
+    return request.param
+
+
+@pytest.fixture(params=DEVICES, ids=[device_id(d) for d in DEVICES])
+def device(request):
+    return request.param
+
+
+@pytest.fixture(params=VALUE_DTYPES, ids=[dtype_id(d) for d in VALUE_DTYPES])
+def value_dtype(request):
+    return request.param
+
+
+@pytest.fixture(params=INDEX_DTYPES, ids=[dtype_id(d) for d in INDEX_DTYPES])
+def index_dtype(request):
+    return request.param
+
+
+@pytest.fixture(params=[True, False], ids=["incl_zeros", "excl_zeros"])
+def include_zeros(request):
+    return request.param
+
+
+def test_matches_two_call_and_dense(layout, device, value_dtype, index_dtype, include_zeros):
+    """(col_lse, row_lse) equals two sparse_logsumexp calls and the dense reference."""
+    dense = _make_dense(device, value_dtype, seed=0)  # 5x4: has an all-zero row and column
+    sp = _to_layout(dense, layout, index_dtype)
+    col_lse, row_lse = sparse_bidir_logsumexp(sp, include_zeros=include_zeros)
+    _assert_close(col_lse, sparse_logsumexp(sp, dim=0, include_zeros=include_zeros), value_dtype)
+    _assert_close(row_lse, sparse_logsumexp(sp, dim=1, include_zeros=include_zeros), value_dtype)
+    _assert_close(col_lse, _dense_reference(dense, 0, include_zeros), value_dtype)
+    _assert_close(row_lse, _dense_reference(dense, 1, include_zeros), value_dtype)
+
+
+def test_output_layouts_agree(layout, device, include_zeros):
+    """padded and nested carry the same values as tuple; padded pads to -inf."""
+    dense = _make_dense(device, torch.float64, seed=1)  # 5x4 -> G=5, col_lse padded by one
+    sp = _to_layout(dense, layout)
+    nrows, ncols = dense.shape
+    G = max(nrows, ncols)
+    col_lse, row_lse = sparse_bidir_logsumexp(sp, include_zeros=include_zeros)
+
+    padded = sparse_bidir_logsumexp(sp, include_zeros=include_zeros, output_layout="padded")
+    assert padded.shape == (2, G)
+    _assert_close(padded[0, :ncols], col_lse, torch.float64)
+    _assert_close(padded[1, :nrows], row_lse, torch.float64)
+    assert torch.isneginf(padded[0, ncols:]).all()  # col_lse tail padded with -inf
+
+    if _NESTED_OK:
+        nested = sparse_bidir_logsumexp(sp, include_zeros=include_zeros, output_layout="nested")
+        parts = nested.unbind()
+        _assert_close(parts[0], col_lse, torch.float64)
[…]
