### 2026-07-11 6b37a6f
bench: add sparse_bidir_logsumexp benchmarks vs two-call baseline
Random-matrix and SuiteSparse (Rothberg/cfd2, Williams/webbase-1M) benchmarks
comparing sparse_bidir_logsumexp against the two-call baseline
(sparse_logsumexp dim=0 + dim=1), mirroring the sparse_logsumexp scripts and
registered in benchmark_suite. Both ops return a single tensor (padded / cat)
so measure_op's backward exercises both reductions. Real >=1M-dim coverage per
the sparse_logsumexp review.
