### 2026-07-11 87c3df7
docs: correct batched equivalence and soften bidir docstring claims
The 'Equivalent to' snippet used dim=0/dim=1, which raises on a batched 3-D
input (batch axis is dim=0); qualify it as 2-D and note the batched dim=1/dim=2
reductions. Also soften the unqualified 'cheaper than' (benchmarks show ~20%
higher peak memory), drop the subjective 'most pythonic', document the batched
tuple shapes, and note that the nested layout is an unbind-only prototype whose
whole-tensor ops (.sum()/.shape) are unsupported.

Suggested-by: Fable
