### 2026-07-11 89e9e81
fix: reject hybrid sparse tensors and gate nested up front
A 3-D hybrid sparse tensor (2 sparse + 1 dense dim) has ndim==3, so it slipped
past the rank check and was misrouted as batched, then crashed on index
unpacking. Add a dense_dim() != 0 guard to both public entry points, matching
the repo's existing dense-dim validation in utils (ValueError, 'zero dense
dimensions'). Also hoist the nested PyTorch>=2.4 version gate ahead of the
reduction so it no longer pays the full forward cost before raising.

Suggested-by: Fable
