### 2026-07-11 0d89343
test: cover batched layouts, wide matrices, and CSC
Close the periphery gaps the review flagged: batched padded/nested now assert
values (not just shapes), so a col/row plane swap in the batched padded assembly
can't pass; a transposed wide (4x5) case exercises the row-side -inf padding that
the tall fixtures never hit; and a forward-only CSC test reaches the free-col_nnz
path that the [coo, csr] fixture set never covered.

Suggested-by: Fable
