### 2026-07-11 a9e446d
test: add batched grad parity, duplicate-COO, and raises matchers
Cover the remaining robustness gaps: a batched gradient-parity case (the
_bidir_batched backward is a distinct path, previously only probed manually); an
uncoalesced COO with repeated coordinates (duplicates must sum before exp, so a
refactor to _values()/_indices() that double-counts would fail); and match=
patterns on the two ValueError raises so an unrelated ValueError can't satisfy
them.

Suggested-by: Fable
