### 2026-07-11 9b14094
bench: make index-dtype labels truthful and fail loudly
COO always stores int64 indices, so an int32 request was silently promoted and
every 'int32' COO row (and the SuiteSparse CSR rows built via to_sparse_csr)
duplicated the int64 measurement under a wrong label. Skip COO+int32 loudly,
rebuild SuiteSparse CSR with the requested index dtype, and assert the built
dtype matches the label. Also hoist matrix generation out of the timing try so a
generation failure surfaces instead of being misattributed to the op.

Suggested-by: Fable
