### 2026-07-13 2b3b273
docs: clarify n_groups is shared across batch slices
_scatter_logsumexp's output is a rectangular (*batch, n_groups), so n_groups
is one value for every slice. Document that a caller whose slices need
different group counts passes the max and pads — the surplus groups come back
-inf and can be sliced off, which is the contract the bidirectional path
relies on with G = max(nrows, ncols).

Addresses tvercaut's review comment on PR #89.
