### 2026-07-13 f75af67
docs: note col_lse/row_lse are views on the padded buffer
_bidir_2d returns the scatter's native output buffer alongside two basic-slice
views into it, so the three returns cost one allocation between them rather
than three.

Addresses tvercaut's review comment on PR #89.
