### 2026-07-13 5972175
perf: only pay the padded transpose when that layout is requested
_bidir_batched returned padded.permute(1, 0, 2).contiguous(), materialising a
full (b, 2, G) copy on every batched call — including output_layout="tuple" and
"nested", which discard it untouched.

It now returns the scatter's native (2, b, G) buffer with col_lse/row_lse as
views into it, matching _bidir_2d, and the public function does the transpose
only in the "padded" branch. Public behaviour is unchanged: "padded" still
returns a contiguous (b, 2, G).

The copy cannot be avoided for "padded" itself — the scatter's batch axis

[…]
