### 2026-07-13 bba0aa8
docs: note why _logsumexp_batched folds instead of using the batch axis
_scatter_logsumexp's batch axis needs a rectangular (*batch, nnz), but the
slices have ragged nnz. Padding them to (b, max_nnz) is correct but measured
2.6x slower at 1.4x the memory with evenly-filled slices, and degrades sharply
with skew (one heavy slice gave a 24x padding factor and 27x the time).

Addresses tvercaut's review comment on PR #89.
