### 2026-07-13 ab53192
docs: say where _logsumexp_batched's unequal-nnz slices come from
The comment said padding "the ragged slices" was slower without saying why the
slices are ragged, three lines under a docstring stating that batched CSR/CSC
require equal nnz per slice. The two read as contradictory.

Batched COO is the layout that permits unequal nnz — one global (3, nnz) index
matrix, so slices can hold any counts — and _logsumexp_batched coalesces every
input to COO, so that case is reachable. Name the layout in the comment.

Addresses tvercaut's review comment on PR #89.
