+++ torchsparsegradutils/sparse_logsumexp.py
@@ -146,7 +146,9 @@ def _logsumexp_batched(input: Tensor, dims, keepdim: bool, include_zeros: bool) 
     at once. Batched inputs go through COO (batched CSR/CSC require equal nnz per
     slice in PyTorch).
     """
-    # Not _scatter_logsumexp's batch axis: padding the ragged slices rectangular measured 2.6x slower.
+    # Not _scatter_logsumexp's batch axis: it needs a rectangular (*batch, nnz), but COO slices can
+    # have unequal nnz (unlike CSR/CSC, which PyTorch forces equal). Padding them rectangular
+    # measured 2.6x slower, and far worse under skew.
     b, nrows, ncols = input.shape
     coo = input if (input.layout == torch.sparse_coo and input.is_coalesced()) else input.to_sparse_coo().coalesce()
     bidx, rows, cols = coo.indices()
