### 2026-07-13 611ce08
refactor: make the padded layout direction-first at both ranks
output_layout="padded" returned (2, G) unbatched but (batch, 2, G) batched, so
out[0] was col_lse in the first case and the first batch slice in the second.
The same argument meant different things depending on input rank.

Return the scatter's native (2, batch, G) buffer instead, and document it. The
direction now leads at both ranks: out[0] and out[1] are the column and row
reductions whatever the input. This is a breaking change to the "padded" shape.

Serving the old shape required a permute + contiguous, so this also drops a c
[…]
