### 2026-07-13 53dbbcd
docs: publish sparse_bidir_logsumexp and flag "padded" as experimental
sparse_bidir_logsumexp was exported in __all__ but never added to
docs/source/api/core.rst, so its docstring rendered nowhere.

The "padded" layout hands back _scatter_logsumexp's native output buffer
without a copy, so its axis order, G = max(rows, cols) group padding and
-inf fill are all artifacts of that buffer. Say so, and point callers who
need a stable layout at "tuple".

Addresses tvercaut's review comment on PR #89.

Suggested-by: tvercaut
