### 2026-07-13 9144b86
docs: align batching terminology and state the batched CSR/CSC constraint
The docstrings described inputs as "2-D" / "batched 3-D (batch, rows, cols)",
which is not the vocabulary the rest of the repo uses. Say unbatched [r, c] and
batched [b, r, c] throughout, and spell the output shapes with the same letters
((b, c), (2, b, G), G = max(r, c)).

Both functions advertise CSR/CSC support, but not that PyTorch requires every
slice of a batched compressed tensor to hold the same number of specified
elements — a ragged batch is COO-only, and the caller hits that at construction,
before reaching us
[…]
