### 2026-07-13 fa37dfa
test: cover CSR and CSC, batched and unbatched, in both logsumexp suites
The docstrings claim COO/CSR/CSC support at both ranks, but CSC only had a
single forward-only smoke test, and every batched test hardcoded .to_sparse_coo()
— so batched CSR/CSC, which reach the reduction by a different route (converted,
as batched compressed tensors expose no per-slice index accessors), were never
exercised.

Add FWD_LAYOUTS (= SPARSE_LAYOUTS + CSC) and drive the forward tests off it,
dropping the one-off test_csc_layout it subsumes. SPARSE_LAYOUTS still backs the
gradient tests: PyTorch has no backward 
[…]
