+++ torchsparsegradutils/tests/test_sparse_bidir_logsumexp.py
@@ -5,9 +5,11 @@ from test_config import DEVICES, INDEX_DTYPES, SPARSE_LAYOUTS, VALUE_DTYPES
 
 # Reuse the sibling module's helpers/fixtures so the two suites stay in lock-step.
 from test_sparse_logsumexp import (
+    FWD_LAYOUTS,
     _assert_close,
     _dense_reference,
     _make_batched_dense,
+    _make_batched_dense_equal_nnz,
     _make_dense,
     _to_layout,
     device_id,
@@ -23,6 +25,16 @@ _NESTED_OK = parse_version(torch.__version__) >= parse_version("2.4")
 
 @pytest.fixture(params=SPARSE_LAYOUTS, ids=[layout_id(x) for x in SPARSE_LAYOUTS])
 def layout(request):
+    return request.param
+
+
+@pytest.fixture(params=FWD_LAYOUTS, ids=[layout_id(x) for x in FWD_LAYOUTS])
+def fwd_layout(request):
+    return request.param
+
+
+@pytest.fixture(params=FWD_LAYOUTS, ids=[layout_id(x) for x in FWD_LAYOUTS])
+def batched_layout(request):
     return request.param
 
 
@@ -46,10 +58,10 @@ def include_zeros(request):
     return request.param
 
 
-def test_matches_two_call_and_dense(layout, device, value_dtype, index_dtype, include_zeros):
+def test_matches_two_call_and_dense(fwd_layout, device, value_dtype, index_dtype, include_zeros):
     """(col_lse, row_lse) equals two sparse_logsumexp calls and the dense reference."""
     dense = _make_dense(device, value_dtype, seed=0)  # 5x4: has an all-zero row and column
-    sp = _to_layout(dense, layout, index_dtype)
+    sp = _to_layout(dense, fwd_layout, index_dtype)
     col_lse, row_lse = sparse_bidir_logsumexp(sp, include_zeros=include_zeros)
     _assert_close(col_lse, sparse_logsumexp(sp, dim=0, include_zeros=include_zeros), value_dtype)
     _assert_close(row_lse, sparse_logsumexp(sp, dim=1, include_zeros=include_zeros), value_dtype)
@@ -57,10 +69,10 @@ def test_matches_two_call_and_dense(layout, device, value_dtype, index_dtype, in
     _assert_close(row_lse, _dense_reference(dense, 1, include_zeros), value_dtype)
 
 
-def test_output_layouts_agree(layout, device, include_zeros):
+def test_output_layouts_agree(fwd_layout, device, include_zeros):
     """padded and nested carry the same values as tuple; padded pads to -inf."""
     dense = _make_dense(device, torch.float64, seed=1)  # 5x4 -> G=5, col_lse padded by one
-    sp = _to_layout(dense, layout)
+    sp = _to_layout(dense, fwd_layout)
     nrows, ncols = dense.shape
     G = max(nrows, ncols)
     col_lse, row_lse = sparse_bidir_logsumexp(sp, include_zeros=include_zeros)
@@ -78,11 +90,11 @@ def test_output_layouts_agree(layout, device, include_zeros):
         _assert_close(parts[1], row_lse, torch.float64)
 
 
-def test_wide_matrix_row_padding(layout, device):
+def test_wide_matrix_row_padding(fwd_layout, device):
     """ncols > nrows exercises the row-side -inf padding; the tall (5x4) fixtures only
     ever hit the column-side pad, so transpose to a wide (4x5) matrix."""
     dense = _make_dense(device, torch.float64, seed=1).t().contiguous()  # (4, 5), wide
-    sp = _to_layout(dense, layout)
+    sp = _to_layout(dense, fwd_layout)
     nrows, ncols = dense.shape  # 4, 5 -> G=5
     G = max(nrows, ncols)
[…]
