+++ torchsparsegradutils/tests/test_sparse_logsumexp.py
@@ -17,7 +17,12 @@ def device_id(device):
 
 
 def layout_id(layout):
-    return "coo" if layout == torch.sparse_coo else "csr"
+    return {torch.sparse_coo: "coo", torch.sparse_csr: "csr", torch.sparse_csc: "csc"}[layout]
+
+
+# SPARSE_LAYOUTS (coo, csr) is the repo-wide matrix, and is what the gradient tests can
+# use: PyTorch has no backward for CSC values. Forward-only tests take the wider set.
+FWD_LAYOUTS = [*SPARSE_LAYOUTS, torch.sparse_csc]
 
 
 def dim_id(dim):
@@ -61,6 +66,11 @@ def layout(request):
     return request.param
 
 
+@pytest.fixture(params=FWD_LAYOUTS, ids=[layout_id(x) for x in FWD_LAYOUTS])
+def fwd_layout(request):
+    return request.param
+
+
 @pytest.fixture(params=DEVICES, ids=[device_id(d) for d in DEVICES])
 def device(request):
     return request.param
@@ -96,30 +106,23 @@ def _make_dense(device, value_dtype, seed):
     return dense.to(device=device, dtype=value_dtype)
 
 
-def test_matches_dense_reference(layout, device, value_dtype, index_dtype, dim, include_zeros):
+def test_matches_dense_reference(fwd_layout, device, value_dtype, index_dtype, dim, include_zeros):
     dense = _make_dense(device, value_dtype, seed=0)
-    sp = _to_layout(dense, layout, index_dtype)
+    sp = _to_layout(dense, fwd_layout, index_dtype)
     out = sparse_logsumexp(sp, dim=dim, include_zeros=include_zeros)
     ref = _dense_reference(dense, dim, include_zeros)
     _assert_close(out, ref, value_dtype)
 
 
-def test_keepdim_shapes(layout, device):
+def test_keepdim_shapes(fwd_layout, device):
     dense = _make_dense(device, torch.float64, seed=1)
-    sp = _to_layout(dense, layout)
+    sp = _to_layout(dense, fwd_layout)
     assert sparse_logsumexp(sp, dim=1, keepdim=True).shape == (5, 1)
     assert sparse_logsumexp(sp, dim=0, keepdim=True).shape == (1, 4)
     assert sparse_logsumexp(sp, dim=[0, 1], keepdim=True).shape == (1, 1)
 
 
-def test_csc_layout(device):
-    """CSC is supported even though it is not in the standard test matrix."""
-    dense = _make_dense(device, torch.float64, seed=2)
-    out = sparse_logsumexp(dense.to_sparse_csc(), dim=1, include_zeros=True)
-    _assert_close(out, torch.logsumexp(dense, dim=1), torch.float64)
-
-
-def test_all_negative_values_stability(layout, device, dim, include_zeros):
+def test_all_negative_values_stability(fwd_layout, device, dim, include_zeros):
     """All-negative values: the stability shift must account for the structural
     zero (value 0) so include_zeros neither overflows to +inf (rows with a zero)
     nor produces 0 * inf -> nan (fully dense rows). Row 0 has a structural zero,
@@ -129,7 +132,7 @@ def test_all_negative_values_stability(layout, device, dim, include_zeros):
         device=device,
         dtype=torch.float64,
     )
-    out = sparse_logsumexp(_to_layout(dense, layout), dim=dim, include_zeros=include_zeros)
+    out = sparse_logsumexp(_to_layout(dense, fwd_layout), dim=dim, include_zeros=include_zeros)
     _assert_close(out, _dense_reference(dense, dim, include_zeros), torch.float64)
     assert not torch.isnan(out).any()
 
[…]
