### 2026-07-13 1da4582
docs: add naming and shape conventions guide
