+++ docs/source/index.rst
@@ -62,6 +62,7 @@ Sparse Multivariate Normal Distributions
    api/index
    benchmarks
    contributing
+   naming
 
 Installation
 ============
