### 2026-07-13 daf770c
docs: clarify terminology and consistency in naming conventions
