### 2026-07-14 dbbf6e2
Add sparse_bidir_logsumexp: simultaneous row- and column-wise log-sum-exp (#89)
* refactor: generalise _scatter_logsumexp to a leading batch dim

Operate on the last axis so values/scatter_index of shape (*batch, nnz)
reduce independently into (*batch, n_groups). The 1-D path is unchanged
(dim=-1 == 0, gather(...,-1,idx) == idx-fancy-index), so sparse_logsumexp's
existing callers and their numerical fixes are untouched; enables a single
batched scatter for the upcoming sparse_bidir_logsumexp.

Suggested-by: tvercaut

* feat: add sparse_bidir_logsumexp (simultaneous row + column LSE)

Computes 
[…]
