### 2026-07-14 f19d7b4
docs: add sparse tensor naming and shape conventions (#90)
* remove specific cpu core alloc

* docs: add naming and shape conventions guide

* docs: clarify terminology and consistency in naming conventions
