+++ docs/source/naming.rst
@@ -1,0 +1,224 @@
+Naming and shape conventions
+============================
+
+This guide defines the terminology used for shapes, batching, sparse layouts,
+and dimensions in ``torchsparsegradutils``. Use it for code, docstrings,
+tests, issues, and pull requests. State the logical object first and then its
+shape: sparse tensor terminology otherwise makes it easy to give different
+concepts the same name.
+
+Core terminology
+----------------
+
+Use **matrix** for a logical object with row and column axes, and use
+**tensor** for the PyTorch object or for a genuinely non-matrix object. For
+the non-hybrid matrix tensors used by most operations in this repository,
+the row and column axes are the final two axes. Hybrid sparse tensors may
+have trailing dense value dimensions.
+
+* An unbatched sparse matrix has shape ``(n_rows, n_cols)``.
+* A batched sparse matrix (a batch of sparse matrices) has shape
+  ``(batch_size, n_rows, n_cols)``.
+* An unbatched square sparse matrix has shape ``(n, n)``; its batched form has
+  shape ``(batch_size, n, n)``.
+* A dense matrix of right-hand sides has shape ``(n_rows, n_rhs)``; its
+  batched form has shape ``(batch_size, n_rows, n_rhs)``.
+* A vector has shape ``(n,)`` and a batch of vectors has shape
+  ``(batch_size, n)``.
+
+``2D``, ``3D``, ``rank 2``, and ``rank 3`` describe only the number of
+tensor axes. They do not identify an axis as a batch, matrix, spatial, or
+dense-value axis. Prefer, for example, “a batched sparse matrix with shape
+``(batch_size, n_rows, n_cols)``” to “a 3D sparse tensor”.
+
+When an API supports at most one batch axis, say so directly. Unless an API
+documents an exception, batch sizes must match exactly and batch broadcasting
+is not supported.
+
+Rank, batch, sparse, and dense dimensions
+------------------------------------------
+
+These terms are not interchangeable:
+
+* **Tensor rank** is ``tensor.ndim``, the total number of logical axes.
+* A **batch dimension** indexes independent problem instances. Repository
+  convention is one leading batch axis.
+* A **sparse dimension** has coordinates represented by sparse indices. For
+  COO, ``sparse_dim()`` is the number of leading indexed dimensions. CSR and
+  CSC always have two sparse matrix dimensions.
+* A **dense dimension of a sparse tensor** is a trailing dimension stored in
+  each specified value. It is not the same as a dense, ``torch.strided``
+  tensor. Most matrix operations here require ``dense_dim() == 0`` unless
+  documented otherwise.
+
+For example, a hybrid sparse matrix with shape
+``(n_rows, n_cols, feature_size)`` can have two sparse dimensions and one
+dense feature dimension. A true three-dimensional sparse object, such as a
+sparse volume ``(depth, height, width)``, is different from a rank-3 tensor
+whose leading axis is a batch of matrices.
+
+COO and compressed-layout batching
+----------------------------------
+
+PyTorch represents batching differently for COO and compressed layouts, so
+documentation must preserve that distinction.
+
+For COO, a tensor of shape ``(batch_size, n_rows, n_cols)`` commonly has three
+sparse dimensions. PyTorch does not label the first one as a batch dimension;
+an API in this package *logically interprets* it as one. COO batch items may
+have different numbers of specified entries, including zero.
+
+For CSR and CSC, PyTorch represents tensors as
+``(*batch_shape, n_rows, n_cols, *dense_shape)``. The normal form supported by
+this package is ``(batch_size, n_rows, n_cols)``. Batched CSR and CSC require
[…]
