### 2026-07-14 ede9f14
docs(spec): add CUDA-rewrite spec
Full design spec for the CUDA-native rewrite: goal + agent execution
model, API map with contract invariants and kernel routing, naming
conventions carried through the migration, 22-commit playbook with
parallel lanes, architecture (two packages, torch.library dispatch,
BatchedCSR), kernel families, build tooling, testing (parity +
legitimacy), and benchmark protocol with dummy sweep charts.

Co-Authored-By: Claude Fable 5 <noreply@anthropic.com>
