+++ spec/images/make_dummy_charts.py
@@ -1,0 +1,163 @@
+"""Generate the dummy sweep charts embedded in ../benchmarks.md §6.
+
+Run:  uv run --with matplotlib python make_dummy_charts.py
+All values are illustrative placeholders — regenerate with real data by
+swapping the DATA blocks once the suite produces CSVs.
+"""
+
+from pathlib import Path
+
+import matplotlib
+
+matplotlib.use("Agg")
+import matplotlib.pyplot as plt
+
+OUT = Path(__file__).parent
+
+# Palette: fixed hue per kernel family (validated, light surface)
+SURFACE = "#fcfcfb"
+TEXT = "#0b0b0b"
+TEXT_2 = "#52514e"
+GRID = "#e7e6e2"
+BLUE_SPMM = "#2a78d6"  # spmm family
+AQUA_SEGLSE = "#1baf7a"  # seglse family
+YELLOW_BASE = "#eda100"  # baseline paths in multi-series charts
+
+
+def sweep_chart(fname, title, subtitle, xlabel, x, y, color, log2_x, ymax):
+    fig, ax = plt.subplots(figsize=(7, 4), dpi=160)
+    fig.patch.set_facecolor(SURFACE)
+    ax.set_facecolor(SURFACE)
+
+    # recessive grid + axes
+    ax.grid(axis="y", color=GRID, linewidth=0.8, zorder=0)
+    for side in ("top", "right"):
+        ax.spines[side].set_visible(False)
+    for side in ("left", "bottom"):
+        ax.spines[side].set_color(GRID)
+    ax.tick_params(colors=TEXT_2, labelsize=9)
+
+    # parity reference — neutral, dashed, labelled directly (no legend needed)
+    ax.axhline(1.0, color=TEXT_2, linewidth=1, linestyle=(0, (4, 4)), zorder=1)
+    ax.annotate("parity 1.0×", xy=(0.99, 1.0), xycoords=("axes fraction", "data"),
+                ha="right", va="bottom", fontsize=8, color=TEXT_2)
+
+    # the one series — 2px line, 8px markers; title carries its identity
+    ax.plot(x, y, color=color, linewidth=2, marker="o", markersize=7,
+            markerfacecolor=color, markeredgecolor=SURFACE, markeredgewidth=1.5,
+            zorder=3)
+
+    # selective direct label: endpoint only, in text ink (not series color)
+    ax.annotate(f"{y[-1]:.2f}×", xy=(x[-1], y[-1]), xytext=(0, 9),
+                textcoords="offset points", ha="center", fontsize=9,
+                color=TEXT, fontweight="bold")
+
+    if log2_x:
+        ax.set_xscale("log", base=2)
+        ax.set_xticks(x)
+        ax.get_xaxis().set_major_formatter(plt.FuncFormatter(lambda v, _: f"{int(v)}"))
+        ax.minorticks_off()
+    else:
+        ax.set_xticks(x)
+
+    ax.set_ylim(0, ymax)
+    ax.set_xlabel(xlabel, fontsize=9, color=TEXT_2)
+    ax.set_ylabel("speedup ×", fontsize=9, color=TEXT_2)
+    ax.set_title(subtitle, fontsize=9, color=TEXT_2, loc="left", pad=4)
+    fig.suptitle(title, fontsize=11, color=TEXT, x=0.125, ha="left", y=0.97)
+
+    ax.annotate("DUMMY DATA", xy=(0.01, 0.02), xycoords="axes fraction",
+                fontsize=8, color=TEXT_2, alpha=0.8, style="italic")
+
[…]
