+++ spec/map.md
@@ -1,0 +1,171 @@
+# MAP
+
+Every public function/class in the package, its current implementation, and its
+CUDA-rewrite classification. Baseline: `main` @ f19d7b4 (v0.2.3). Terminology
+follows `docs/source/naming.rst`. Goal context: [goal.md](goal.md).
+
+## Contract invariants (apply to every public API)
+
+Stated once here; per-op specifics live in the **Contract** column of §1.
+
+1. **Surviving signatures frozen at v0.2.3** — names, parameters, defaults, return
+   types of every API that is not ⚫ RETIRE. Retired APIs are removed outright —
+   this is a breaking release with no migration path or deprecation cycle.
+2. **naming.rst is binding** for all new code, docs, and error messages.
+3. **Gradient guarantee** — gradients w.r.t. a sparse input are sparse, at the
+   input's sparsity pattern, in the input's layout. Never densified.
+4. **Layout parity** — every op accepts what it accepts today (COO and/or CSR);
+   output layouts unchanged.
+5. **Batching** — one leading batch axis, exact match, no broadcasting. Batched
+   COO items may have unequal nse; batched CSR requires equal nse per item.
+   `_nnz()` semantics per layout (CSR: per batch item; COO: whole tensor) preserved.
+6. **Index dtypes** — int32 and int64 supported where supported today; dtype
+   preserved through outputs and gradients (upstream COO int64 coercion excepted).
+7. **Errors** — invalid input raises (never silently accepts); messages state
+   accepted logical shapes and the received shape.
+8. **No legacy code** — the current pure-PyTorch implementations are not carried
+   into the rewrite. Whether they serve as an *external* differential oracle
+   (and what the CPU/non-CUDA story is) is deferred to the testing discussion.
+9. **Coalescing** — COO semantics for uncoalesced input match current behaviour
+   per op; compressed layouts never described as (un)coalesced.
+
+**Legend**
+
+| Tier | Meaning |
+|------|---------|
+| 🔴 KERNEL | Needs a custom CUDA kernel (no vendor primitive covers it) |
+| 🟠 VENDOR | Custom-kernel target with a vendor primitive (cuSPARSE / cuBLAS / cuDSS) as the **baseline to beat**; vendor binding allowed only as bring-up scaffold + benchmark reference |
+| 🟡 COMPOSITE | Host-side orchestration; becomes CUDA-native by calling the 🔴/🟠 ops above it |
+| 🟢 HOST | Stays pure PyTorch/host — no CUDA work needed (setup, generation, stats, glue) |
+| ⚫ RETIRE | Made obsolete by the rewrite (workaround for a gap that native kernels close) |
+
+## 1. Core ops (`torchsparsegradutils/__init__.py` exports)
+
+| API | Location | Current implementation | Tier | CUDA plan | Contract |
+|-----|----------|------------------------|------|-----------|----------|
[…]
