+++ spec/architecture.md
@@ -1,0 +1,153 @@
+# Architecture
+
+Supersedes the 2026-07-13 plan where they differ (that plan kept CPU reference
+implementations, bridges, and block-diag as public API — all dropped per
+[map.md](map.md); its dispatch and format decisions are carried forward here).
+
+## 1. Two packages, one monorepo
+
+| Package | Contents | Ships as |
+|---------|----------|----------|
+| **`torchsparsegradutils`** | Pure Python: public API wrappers, op definitions, dispatch, host-side composites (Krylov loops, distributions, encoder, utils) | `py3-none-any` wheel — no compiler, ever |
+| **`torchsparsegradutils-cuda`** | The kernels: CUDA C++ under `csrc/`, registered into the same ops at import | Prebuilt binaries — mechanism is build.md's decision (HF kernel-builder → Hub is front-runner; stable-ABI `cp310-abi3` wheels per CUDA major the alternative) |
+
+- Dependency direction: `-cuda` depends on the front package, never the reverse.
+- Version handshake: `-cuda` exposes `__backend_api_version__`; the front package
+  probes at import (`_dispatch.py`), refuses on mismatch. Escape hatch:
+  `TSGU_DISABLE_CUDA_BACKEND=1`.
+
+## 2. Dispatch: `torch.library` custom ops
+
+- Every kernel-backed op is a `tsgu::` custom op: `torch.library.custom_op` +
+  `register_fake` + `register_autograd`. This replaces the four
+  `autograd.Function` classes and buys `torch.compile` compatibility and
+  `torch.library.opcheck` testability.
+- **Ops take plain dense tensors** (values / rowptr / col / shape ints), never
+  `torch.sparse_*` tensors — sparse layouts have poor FakeTensor/meta support.
+  Public wrappers unwrap at the boundary and rewrap results in the layout the
+  user passed (COO in → COO grad out, per map.md invariant 3).
+- Fake/meta kernels must be value-independent: output shapes derive from index
+  array lengths and shape args only.
+- The CUDA package registers implementations for the CUDA dispatch key from C++
+  (`STABLE_TORCH_LIBRARY_IMPL(tsgu, CUDA, m)`).
+
+## 3. Canonical internal format: `BatchedCSR`
+
+**[naming.md](naming.md) is the backbone here** — it carries `naming.rst`'s
+conventions through the migration and defines this section's vocabulary
+(*descriptor*, *folded row*, *local column*, `nse_total`/`nse_per_item`, the
+kernel short-name mapping). Every identifier below is used in its naming.md
+sense; new BatchedCSR-related terms get added there, not invented ad hoc.
+
+One frozen descriptor (`_batched.py`) replaces the block-diag hack everywhere,
+with **ragged nse per batch item as first-class** (COO batching's flexibility,
+CSR's kernel-friendliness):
+
+- `values (nse_total,)` · `rowptr (B·n+1,)` absolute over folded rows `b·n+r` ·
+  `col (nse_total,)` **local** columns in `[0, m)` · `shape (B, n, m)`; `B=1`
+  encodes unbatched.
+- Local (unoffset) columns keep int32 viable and let kernels address
+  `Bdense[b, col, :]` directly; batch of an entry recovered as `row_global / n`.
+- Lazy cached members: uncompressed `row_indices` (SDDMM), transposed
+  `BatchedCSC` (gradB, dim=-2 reductions), **SpSM analysis plans** — resolving
+  kernels.md open Q3: the plan cache lives on the descriptor, so its lifetime
+  and invalidation are the descriptor's lifetime; no global cache, no hashing.
[…]

[…]
