### 2026-07-14 659f949
refactor!: kill legacy — drop CuPy/JAX bridges and deprecated APIs
Delete torchsparsegradutils/{cupy,jax}/ and their 4 test files
(test_cupy_bindings.py, test_cupy_sparse_solve.py, test_jax_bindings.py,
test_jax_sparse_solve.py). Remove the cupy/jax/all optional-dependency
extras from pyproject.toml and drop jax[cpu] from requirements-ci.txt.

Delete the deprecated pairwise_voxel_encoder.py (PairwiseVoxelEncoder),
calc_pairwise_coo_indices + the typo alias calc_pariwise_coo_indices in
pairwise_encoder.py, and the __getattr__ lazy-alias machinery in
encoders/__init__.py. Test coverage for the r
[…]
