+++ torchsparsegradutils/benchmarks/sparse_triangular_solve_rand.py
@@ -1,244 +1,0 @@
-#!/usr/bin/env python3
-"""
-Sparse Triangular Solve Benchmark - Random Matrices
-
-This benchmark tests sparse triangular solve operations using randomly
-generated sparse triangular matrices of various sizes.
-"""
-
-import os
-import sys
-
-# Add the parent directory to sys.path to allow importing torchsparsegradutils
-sys.path.insert(0, os.path.join(os.path.dirname(__file__), "..", ".."))
-
-import numpy as np
-import pandas as pd
-import torch
-from benchmark_utils import (
-    format_time,
-    measure_op,
-    print_benchmark_header,
-    print_result_row,
-    print_results_table_header,
-    save_benchmark_results,
-)
-from cupyx.scipy.sparse.linalg._solve import spsolve_triangular
-from tqdm import tqdm
-
-from torchsparsegradutils import sparse_triangular_solve
-from torchsparsegradutils.cupy.cupy_sparse_solve import sparse_solve_c4t
-from torchsparsegradutils.utils import rand_sparse, rand_sparse_tri
-
-# from jax.lax.linalg import triangular_solve  # NOTE: jax doesn't have a sparse triangular solve
-# import torchsparsegradutils.jax as tsgujax
-
-
-REPEATS = 10
-WARMUP_RUNS = 1
-
-# Only run on CUDA
-device = torch.device("cuda")
-assert torch.cuda.is_available(), "This benchmark requires a CUDA GPU"
-
-# problem sizes: (label, N, M, nnz)
-SIZES = [
-    ("small", 2**10, 2**1, 2**11),
-    ("medium", 2**14, 2**2, 2**15),
-    ("large", 2**18, 2**3, 2**19),
-]
-
-UPPER = False
-UNITRIANGULAR = False
-TRANSPOSE = False
-
-INDEX_DTYPES = [torch.int32, torch.int64]
-VALUE_DTYPES = [torch.float32, torch.float64]
-LAYOUTS = [torch.sparse_coo, torch.sparse_csr]
-
-ALGORITHMS = [
-    (
-        "dense.triangular_solve",
-        lambda A, B: torch.triangular_solve(
-            B, A.to_dense(), upper=UPPER, unitriangular=UNITRIANGULAR, transpose=TRANSPOSE
-        ).solution,
-    ),
-    (
-        "torch_triangular_solve",
-        lambda A, B: torch.triangular_solve(
-            B, A, upper=UPPER, unitriangular=UNITRIANGULAR, transpose=TRANSPOSE
-        ).solution,
-    ),
-    (
-        "sparse_triangular_solve",
-        lambda A, B: sparse_triangular_solve(A, B, upper=UPPER, unitriangular=UNITRIANGULAR, transpose=TRANSPOSE),
-    ),
-    (
-        "cupy.spsolve_triangular",
-        lambda A, B: sparse_solve_c4t(
-            A,
-            B,
-            solve=lambda A_inner, b_inner: spsolve_triangular(
-                A_inner, b_inner, lower=not UPPER, unit_diagonal=UNITRIANGULAR
-            ),
-        ),
-    ),
-    # (
-    #     "jax.triangular_solve",
-    #     lambda A, B: tsgujax.sparse_solve_j4t(
-    #         A,
-    #         B,
-    #         solve=lambda A_jax, B_jax: triangular_solve(
-    #             A_jax, B_jax, left_side=True, lower=not UPPER, transpose_a=TRANSPOSE, unit_diagonal=UNITRIANGULAR
-    #         ),
-    #     ),
-    # ),
-]
-
-
-def run_sparse_triangular_solve_benchmark():
-    """Run the sparse triangular solve benchmark suite."""
-
[…]
