+++ torchsparsegradutils/benchmarks/benchmark_suite.py
@@ -14,10 +14,6 @@ from pathlib import Path
 BENCHMARK_SCRIPTS = [
     "sparse_mm_rand.py",
     "sparse_mm_suite.py",
-    "sparse_generic_solve_rand.py",
-    "sparse_generic_solve_suite.py",
-    "sparse_triangular_solve_rand.py",
-    "sparse_triangular_solve_suitesparse.py",
     "batched_sparse_mm_rand.py",
     "sparse_logsumexp_rand.py",
     "sparse_logsumexp_suitesparse.py",
