+++ torchsparsegradutils/jax/jax_sparse_solve.py
@@ -1,258 +1,0 @@
-from typing import Any, Callable, Optional, Tuple
-
-import jax
-import jax.numpy as jnp
-import jax.scipy.sparse.linalg
-import torch
-
-from .jax_bindings import j2t as _j2t, t2j as _t2j, t2j_coo as _t2j_coo, t2j_csr as _t2j_csr
-
-
-def sparse_solve_j4t(
-    A: torch.Tensor,
-    B: torch.Tensor,
-    solve: Optional[Callable[..., Tuple["jax.Array", Any]]] = None,
-    transpose_solve: Optional[Callable[..., Tuple["jax.Array", Any]]] = None,
-    **kwargs: Any,
-) -> torch.Tensor:
-    r"""Solve sparse linear systems using JAX solvers with PyTorch autograd.
-
-    Computes :math:`A x = B` by dispatching to a JAX iterative solver while
-    preserving gradients via a custom autograd function. Supports COO/CSR
-    inputs, single or multiple RHS, and CPU/GPU (device inferred from inputs).
-
-    Parameters
-    ----------
-    A : torch.Tensor
-        Sparse square matrix ``(n,n)`` (COO or CSR).
-    B : torch.Tensor
-        RHS of shape ``(n,)`` or ``(n,k)``.
-    solve : callable, optional
-        JAX solver ``(A,B,**kw)->(x,info)`` (default BiCGSTAB).
-    transpose_solve : callable, optional
-        Solver for transpose system in backward (defaults to same solver on ``A.T``).
-    **kwargs : Any
-        Extra args forwarded to the JAX solver (``tol``, ``atol``, ``maxiter`` ...).
-
-    Returns
-    -------
-    torch.Tensor
-        Solution with same shape as ``B``.
-
-    Raises
-    ------
-    TypeError
-        If layout unsupported.
-    ValueError
-        If shape/dtype constraints violated.
-
-    Notes
-    -----
-    * Auto-enables JAX x64 when inputs are ``float64``.
-    * Backward solves :math:`A^T y = g` (implicit differentiation) and forms
-      sparse gradients w.r.t. non-zero entries only.
-
-    See Also
-    --------
-    torchsparsegradutils.jax.jax_bindings : Conversion helpers.
-    SparseSolveJ4T : Underlying autograd function.
-
-    Examples
-    --------
-    Basic (BiCGSTAB)::
-        >>> import torch
-        >>> from torchsparsegradutils.jax.jax_sparse_solve import sparse_solve_j4t
-        >>> idx = torch.tensor([[0,1,1],[0,0,1]])
-        >>> val = torch.tensor([2.0,-1.0,2.0])
-        >>> A = torch.sparse_coo_tensor(idx, val, (2,2))
-        >>> b = torch.tensor([1.0, 3.0])
-        >>> x = sparse_solve_j4t(A, b)
-        >>> x.shape
-        torch.Size([2])
-
-    CG solver::
-        >>> import jax.scipy.sparse.linalg as jaxla
-        >>> _ = sparse_solve_j4t(A, b, solve=jaxla.cg, tol=1e-10)
-
-    Multi-RHS::
-        >>> B = torch.randn(2,3)
-        >>> X = sparse_solve_j4t(A, B)
-        >>> X.shape
-        torch.Size([2, 3])
-
-    Gradient flow::
-        >>> A.requires_grad_(True)  # doctest: +ELLIPSIS
-        tensor(...)
-        >>> b.requires_grad_(True)  # doctest: +ELLIPSIS
-        tensor(...)
-        >>> x = sparse_solve_j4t(A, b)
-        >>> loss = (x**2).sum(); loss.backward()
-    """
-    # Input validation
[…]
