### 2026-07-14 4735811
build: migrate packaging and CI installs to uv
Dependency-groups (dev/docs) replace optional-dependencies; commit
uv.lock; CI installs via astral-sh/setup-uv + uv sync; delete
requirements-ci.txt (surviving pins folded into dev) and setup.py; add
tox.ini (tox-uv) with the full py310-py313 x torch{min,stable,nightly}
matrix, including the py312-torch-stable alias referenced by
spec/commit.md's verify step. black/isort/flake8 stay in the dev group
and lint CI steps — ruff swap is commit 4.

Co-Authored-By: Claude Fable 5 <noreply@anthropic.com>
