+++ setup.py
@@ -1,5 +1,0 @@
-# Minimal setup.py for backward compatibility
-# All configuration is now in pyproject.toml
-from setuptools import setup
-
-setup()
