+++ requirements-ci.txt
@@ -1,9 +1,0 @@
-numpy
-scipy
-isort==6.0.1
-flake8==7.3.0
-black==26.3.1
-parameterized==0.9.0
-pytest==9.0.3
-pytest-rerunfailures==15.1
-pyyaml==6.0.2
