+++ tox.ini
@@ -1,0 +1,50 @@
+[tox]
+requires =
+    tox>=4.22
+    tox-uv>=1.11
+env_list =
+    py310
+    py311
+    py312
+    py313
+    py{310,311,312,313}-torch-min
+    py312-torch-stable
+    py{310,311,312,313}-torch-nightly
+
+[testenv]
+description = run the test suite against the locked (uv.lock) stable torch
+runner = uv-venv-lock-runner
+dependency_groups = dev
+commands = pytest torchsparsegradutils/tests/ -q
+
+# Alias for py312 with the locked stable torch. commit.md's "tree is never
+# red" verify step names this env explicitly - keep the name stable.
+[testenv:py312-torch-stable]
+description = alias of py312 (locked stable torch); referenced by spec/commit.md
+runner = uv-venv-lock-runner
+dependency_groups = dev
+commands = pytest torchsparsegradutils/tests/ -q
+
+# Smoke-test against the minimum supported torch version, py310-py313.
+# Syncs the rest of the lock normally, then overrides just the torch pin
+# in-venv.
+[testenv:py{310,311,312,313}-torch-min]
+description = smoke test against the minimum supported torch (2.5.*)
+runner = uv-venv-lock-runner
+dependency_groups = dev
+uv_sync_locked = false
+commands_pre =
+    uv pip install --python {envpython} "torch==2.5.*"
+commands = pytest torchsparsegradutils/tests/ -q
+
+# Smoke-test against the torch nightly build, py310-py313. Non-gating:
+# allowed to fail.
+[testenv:py{310,311,312,313}-torch-nightly]
+description = smoke test against torch nightly (non-gating)
+runner = uv-venv-lock-runner
+dependency_groups = dev
+uv_sync_locked = false
+ignore_outcome = true
+commands_pre =
+    uv pip install --python {envpython} --pre torch --index-url https://download.pytorch.org/whl/nightly/cpu
+commands = pytest torchsparsegradutils/tests/ -q
