+++ .github/workflows/docs.yml
@@ -15,20 +15,20 @@ jobs:
     steps:
     - uses: actions/checkout@v4
 
-    - name: Set up Python
-      uses: actions/setup-python@v4
+    - name: Install uv (and Python)
+      uses: astral-sh/setup-uv@v5
       with:
         python-version: '3.10'
+        enable-cache: true
 
     - name: Install dependencies
       run: |
-        python -m pip install --upgrade pip
-        pip install -e .[docs]
+        uv sync --locked --group docs
 
     - name: Build documentation
       run: |
         cd docs
-        make html
+        uv run make html
 
     - name: Check documentation
       run: |
