+++ pyproject.toml
@@ -18,6 +18,7 @@ classifiers = [
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
 ]
 dependencies = [
     "torch>=2.5",
@@ -25,7 +26,14 @@ dependencies = [
     "packaging",
 ]
 
-[project.optional-dependencies]
+[project.urls]
+Homepage = "https://github.com/cai4cai/torchsparsegradutils"
+Documentation = "https://torchsparsegradutils.readthedocs.io"
+Repository = "https://github.com/cai4cai/torchsparsegradutils"
+Issues = "https://github.com/cai4cai/torchsparsegradutils/issues"
+Changelog = "https://github.com/cai4cai/torchsparsegradutils/releases"
+
+[dependency-groups]
 docs = [
     "sphinx>=7.0.0",
     "sphinx-rtd-theme>=1.3.0",
@@ -38,19 +46,17 @@ docs = [
     "sphinx-gallery>=0.13.0",
 ]
 dev = [
-    "pytest",
+    "pytest==9.0.3",
     "pytest-cov",
-    "black",
-    "isort",
-    "flake8",
+    "pytest-rerunfailures==15.1",
+    "black==26.3.1",
+    "isort==6.0.1",
+    "flake8==7.3.0",
+    "parameterized==0.9.0",
+    "pyyaml==6.0.2",
+    "numpy",
+    "scipy",
 ]
-
-[project.urls]
-Homepage = "https://github.com/cai4cai/torchsparsegradutils"
-Documentation = "https://torchsparsegradutils.readthedocs.io"
-Repository = "https://github.com/cai4cai/torchsparsegradutils"
-Issues = "https://github.com/cai4cai/torchsparsegradutils/issues"
-Changelog = "https://github.com/cai4cai/torchsparsegradutils/releases"
 
 [tool.setuptools]
 zip-safe = false
