### 2026-07-14 853551d
style: replace black/isort/flake8 with ruff; add clang-format/clang-tidy
Drops black/isort/flake8 (config, dev-group deps, pre-commit hooks, CI
steps) in favour of ruff for lint + format + import sorting. The .flake8
rule surface (select = B,C,E,F,W,T4,B9; ignore = E203,E266,E501,W503,
F403,F401,E731,E402) ports to [tool.ruff.lint] as it was actually
enforced: B/C/T4/B9 required flake8-bugbear/flake8-2020 plugins that were
never installed, so only E/F/W ever fired; W503 has no ruff equivalent
(ruff never implemented it, and it contradicts ruff-format's style).
`I` (isort) rules replace the sta
[…]
