+++ torchsparsegradutils/distributions/sparse_multivariate_normal.py
@@ -6,7 +6,8 @@ from torch.distributions import constraints
 from torch.distributions.distribution import Distribution
 from torch.distributions.utils import _standard_normal
 
-from torchsparsegradutils import sparse_mm as spmm, sparse_triangular_solve as spts
+from torchsparsegradutils import sparse_mm as spmm
+from torchsparsegradutils import sparse_triangular_solve as spts
 
 # from .contraints import sparse_strictly_lower_triangular
 
@@ -283,7 +284,7 @@ class SparseMultivariateNormal(Distribution):
 
             if scale_tril.dim() < 2:
                 raise ValueError(
-                    "scale_tril matrix must be at least two-dimensional, " "with optional leading batch dimension"
+                    "scale_tril matrix must be at least two-dimensional, with optional leading batch dimension"
                 )
             elif scale_tril.dim() > 3:
                 raise ValueError("scale_tril can only have 1 batch dimension, but has {}".format(scale_tril.dim() - 2))
@@ -306,7 +307,7 @@ class SparseMultivariateNormal(Distribution):
 
             if precision_tril.dim() < 2:
                 raise ValueError(
-                    "precision_tril must be at least two-dimensional, " "with optional leading batch dimensions"
+                    "precision_tril must be at least two-dimensional, with optional leading batch dimensions"
                 )
             elif precision_tril.dim() > 3:
                 raise ValueError(
