### 2026-07-14 6792eff
types: add pyrefly type checking
Add pyrefly (uv add --group dev pyrefly) with [tool.pyrefly] scoped to
torchsparsegradutils/ (tests/ subdir included — 224 errors is well under
the ~300 unmanageable threshold, so no exclusion needed).

Baseline: 224 errors as of this commit (checked via `uv run pyrefly
check`), recorded via pyrefly's native --update-baseline mechanism into
pyrefly_baseline.json. CI (python-package.yml) runs `uv run pyrefly
check --baseline pyrefly_baseline.json`, which reports 0 (all 224
existing errors are baselined) and exits green today, while any new
error introduced by fu
[…]
