+++ pyproject.toml
@@ -54,6 +54,7 @@ dev = [
     "numpy",
     "scipy",
     "ruff==0.15.21",
+    "pyrefly>=1.1.1",
 ]
 
 [tool.setuptools]
@@ -99,6 +100,9 @@ quote-style = "double"
 indent-style = "space"
 line-ending = "auto"
 skip-magic-trailing-comma = false
+
+[tool.pyrefly]
+project-includes = ["torchsparsegradutils/"]
 
 [tool.pytest.ini_options]
 minversion = "6.0"
