### 2026-07-14 a9594d1
test: extract parity oracle; move tests to repo root; drop old benchmark suite
Freezes the pre-rewrite core-op implementations (sparse_mm, triangular/generic
solve, generic lstsq, sparse_logsumexp + bidir, segment_mm/gather_mm) into
tests/oracle/ as parity Oracle A (testing.md), self-contained via a local
_helpers.py copy of the block-diag/stack_csr batching helpers that are
scheduled for retirement. Moves torchsparsegradutils/tests/ -> tests/ (repo
root) so the package no longer ships test code, and deletes the old
torchsparsegradutils/benchmarks/ suite (preserved in git history and the
tagge
[…]
