+++ torchsparsegradutils/benchmarks/sparse_bidir_logsumexp_rand.py
@@ -1,191 +1,0 @@
-#!/usr/bin/env python3
-"""
-Sparse Bidirectional Log-Sum-Exp Benchmark - Random Matrices
-
-Benchmarks ``sparse_bidir_logsumexp`` (both row- and column-wise reductions in a
-single traversal) against the two-call baseline
-``(sparse_logsumexp(A, dim=0), sparse_logsumexp(A, dim=1))`` on randomly generated
-sparse matrices of various sizes, sparsity patterns, layouts and dtypes. Forward
-and backward time and peak memory are measured; the baseline is what the single-
-pass primitive exists to beat. Both algorithms are driven through a single tensor
-(``output_layout="padded"`` / ``torch.cat``) so ``measure_op``'s ``.sum().backward()``
-exercises both reductions.
-"""
-
-import os
-import sys
-
-# Add the parent directory to sys.path to allow importing torchsparsegradutils
-sys.path.insert(0, os.path.join(os.path.dirname(__file__), "..", ".."))
-
-import numpy as np
-import torch
-from benchmark_utils import (
-    measure_op,
-    print_benchmark_header,
-    print_result_row,
-    print_results_table_header,
-    save_benchmark_results,
-)
-from tqdm import tqdm
-
-from torchsparsegradutils import sparse_bidir_logsumexp, sparse_logsumexp
-from torchsparsegradutils.utils import rand_sparse
-
-REPEATS = 100
-WARMUP_RUNS = 10
-
-# Only run on CUDA
-device = torch.device("cuda")
-assert torch.cuda.is_available(), "This benchmark requires a CUDA GPU"
-
-# problem sizes: (label, N, M, nnz)
-SIZES = [
-    ("small", 2**10, 2**10, 2**12),
-    ("medium", 2**12, 2**12, 2**14),
-    ("large", 2**14, 2**14, 2**16),
-    ("xlarge", 2**16, 2**16, 2**18),
-    ("million", 2**20, 2**20, 2**22),
-]
-
-INDEX_DTYPES = [torch.int32, torch.int64]
-VALUE_DTYPES = [torch.float32, torch.float64]
-LAYOUTS = [torch.sparse_coo, torch.sparse_csr]
-
-# Both ops return a single tensor so measure_op's out.sum().backward() drives both
-# reductions. B is an unused placeholder for measure_op's binary signature.
-ALGORITHMS = [
-    ("sparse_bidir_logsumexp", lambda A, B: sparse_bidir_logsumexp(A, output_layout="padded")),
-    ("two_call_baseline", lambda A, B: torch.cat([sparse_logsumexp(A, dim=0), sparse_logsumexp(A, dim=1)])),
-]
-
-
-def run_sparse_bidir_logsumexp_benchmark():
-    """Run the sparse bidirectional log-sum-exp benchmark suite."""
-
-    print_benchmark_header("Sparse Bidirectional Log-Sum-Exp Benchmark - Random Matrices")
-
-    records = []
-
-    for size_label, N, M, nnz in tqdm(SIZES, desc="Problem sizes"):
-        print(f"\n🔍 Testing size: {size_label} (N={N}, M={M}, nnz={nnz})")
-
-        A_shape = (N, M)
-        B = torch.zeros(1, device=device)  # unused placeholder for measure_op's binary signature
-
-        for idx_dt in tqdm(INDEX_DTYPES, desc="Index dtypes", leave=False):
-            for val_dt in tqdm(VALUE_DTYPES, desc="Value dtypes", leave=False):
-                for layout in tqdm(LAYOUTS, desc="Layouts", leave=False):
-                    layout_name = "coo" if layout == torch.sparse_coo else "csr"
-                    print(f"\n  📊 Configuration: idx_dtype={idx_dt}, val_dtype={val_dt}, layout={layout_name}")
-
[…]
