+++ tests/oracle/sparse_solve.py
@@ -1,0 +1,523 @@
+# extracted from f19d7b4 — parity Oracle A, never shipped; do not edit
+import warnings
+from typing import Callable, Optional, cast
+
+import torch
+
+from tests.oracle._helpers import sparse_block_diag, sparse_block_diag_split, stack_csr
+from torchsparsegradutils.utils import convert_coo_to_csr
+
+
+def sparse_triangular_solve(
+    A: torch.Tensor,
+    B: torch.Tensor,
+    upper: bool = True,
+    unitriangular: bool = False,
+    transpose: bool = False,
+) -> torch.Tensor:
+    r"""Sparse triangular solve with memory-efficient sparse gradients.
+
+     Solves the triangular system :math:`\mathbf{A}\,\mathbf{x} = \mathbf{B}` (or
+     :math:`\mathbf{A}^{\top}\,\mathbf{x} = \mathbf{B}` if ``transpose=True``), where
+     :math:`\mathbf{A} \in \mathbb{R}^{m\times m}` is sparse triangular (COO/CSR) and
+     :math:`\mathbf{B} \in \mathbb{R}^{m\times p}` is dense. Gradients preserve the sparsity
+     pattern of :math:`\mathbf{A}` by evaluating only at its nonzero entries. Supports
+     unbatched 2D and batched 3D inputs; COO inputs are converted to CSR internally for the
+     factor solve.
+
+     Let the upstream gradient be :math:`\mathbf{G} = \frac{\partial \mathcal{L}}{\partial \mathbf{x}}`
+     for a scalar objective :math:`\mathcal{L}` and solution :math:`\mathbf{x}`. The dense-form
+     gradients are
+
+     Gradient with respect to B (dense):
+
+     .. math::
+         \frac{\partial \mathcal{L}}{\partial \mathbf{B}} \;=\; \mathbf{A}^{-\top} \, \mathbf{G},
+
+     and for ``transpose=True`` replace :math:`\mathbf{A}` by :math:`\mathbf{A}^{\top}` so that
+     :math:`\frac{\partial \mathcal{L}}{\partial \mathbf{B}} = \left(\mathbf{A}^{\top}\right)^{-\top} \mathbf{G} = \mathbf{A}^{-1} \mathbf{G}`.
+
+     Gradient with respect to A (sparse):
+
+     .. math::
+         \frac{\partial \mathcal{L}}{\partial \mathbf{A}} \;=\; -\big(\mathbf{A}^{-\top} \, \mathbf{G}\big)\, \mathbf{x}^{\top},
+
+     and only entries at the nonzeros of :math:`\mathbf{A}` are evaluated. Equivalently,
+     for a nonzero :math:`\mathbf{A}_{ij}` the contribution is
+
+     .. math::
+         \bigg[\frac{\partial \mathcal{L}}{\partial \mathbf{A}}\bigg]_{ij}
+         \;=\; -\, \big(\mathbf{A}^{-\top} \, \mathbf{G}\big)_{i,:} \,\cdot\, \mathbf{x}_{j,:},
+
+     where the dot denotes a row-wise inner product across the :math:`p` right-hand sides.
+
+    Parameters
+    ----------
+    A : torch.Tensor, sparse COO or CSR, shape ``(m, m)`` or ``(b, m, m)``
+        Sparse triangular coefficient matrix. Must be square per batch. All tensors must
+        be on the same device.
+    B : torch.Tensor, dense (strided), shape ``(m, p)`` or ``(b, m, p)``
+        Right-hand side. ``B.shape[-2]`` must equal ``A.shape[-2]`` (``m``).
+    upper : bool, optional
+        If ``True`` (default), treat ``A`` as upper-triangular; else lower-triangular.
+    unitriangular : bool, optional
[…]
