+++ torchsparsegradutils/benchmarks/__init__.py
@@ -1,10 +1,0 @@
-"""
-Benchmarks for torchsparsegradutils sparse operations.
-
-This package contains benchmarking scripts for comparing different
-sparse linear algebra operations and solvers.
-"""
-
-__all__ = [
-    "benchmark_utils",
-]
