### 2026-07-14 6b1fc3e
refactor: restructure front package to architecture.md §5 hierarchy
Move each public op's implementation into ops/ (matmul, triangular_solve,
generic_solve, lstsq, logsumexp, indexed_matmul), moving the old
autograd.Function bodies into private _legacy_* functions marked for
Phase-3 kernel deletion. Move the Krylov solvers into solvers/ and split
utils/utils.py into utils/convert.py (kept, re-exported) and
utils/_block_diag.py (marked ⚫ RETIRE for commit 21). Fold the dead
distributions/constraints.py stub away. Public API and torchsparsegradutils
.utils re-exports are unchanged; only test imp
[…]
