+++ torchsparsegradutils/ops/triangular_solve.py
@@ -1,5 +1,4 @@
-import warnings
-from typing import Callable, Optional, cast
+from typing import cast
 
 import torch
 
@@ -144,6 +143,13 @@ def sparse_triangular_solve(
     if A.dim() == 3 and A.size(0) != B.size(0):
         raise ValueError("If batched, A and B must have the same batch size")
 
+    return _legacy_sparse_triangular_solve(A, B, upper, unitriangular, transpose)
+
+
+# deleted by its kernel commit (spec/commit.md Phase 3)
+def _legacy_sparse_triangular_solve(
+    A: torch.Tensor, B: torch.Tensor, upper: bool, unitriangular: bool, transpose: bool
+) -> torch.Tensor:
     return cast(torch.Tensor, SparseTriangularSolve.apply(A, B, upper, unitriangular, transpose))
 
 
@@ -250,272 +256,5 @@ class SparseTriangularSolve(torch.autograd.Function):
                 gradA = stack_csr([*gradA])
 
             gradB = gradB.view(ctx.B_shape)
-
-        return gradA, gradB, None, None, None
-
-
-def sparse_generic_solve(
-    A: torch.Tensor,
-    B: torch.Tensor,
-    solve: Optional[Callable[..., torch.Tensor]] = None,
-    transpose_solve: Optional[Callable[..., torch.Tensor]] = None,
-    **kwargs,
-) -> torch.Tensor:
-    r"""Sparse linear solve with iterative methods and sparse-aware gradients.
-
-     Solves :math:`\mathbf{A}\,\mathbf{x} = \mathbf{B}` with sparse :math:`\mathbf{A} \in \mathbb{R}^{n\times n}`
-     (COO/CSR) and dense :math:`\mathbf{B} \in \mathbb{R}^{n\times p}` using iterative methods, while
-     preserving sparsity in :math:`\frac{\partial \mathcal{L}}{\partial \mathbf{A}}`. Supports single
-     (vector) and multiple (matrix) right-hand sides and works with non-differentiable solvers via
-     the implicit function theorem.
-
-     Let :math:`\mathbf{G} = \frac{\partial \mathcal{L}}{\partial \mathbf{x}}` be the upstream gradient and
-     :math:`\mathbf{x}` the solution. The dense-form gradients are
-
-     Gradient with respect to B (dense):
-
-     .. math::
-         \frac{\partial \mathcal{L}}{\partial \mathbf{B}} \;=\; \mathbf{A}^{-\top} \, \mathbf{G}
-         \;\equiv\; \mathbf{G}_B.
-
-     Gradient with respect to A (sparse):
-
-     .. math::
-         \frac{\partial \mathcal{L}}{\partial \mathbf{A}} \;=\; -\, \mathbf{G}_B\, \mathbf{x}^{\top}.
-
-     We evaluate only the entries corresponding to nonzeros of :math:`\mathbf{A}`, yielding a
-     sparse gradient tensor with memory proportional to ``nnz(A)``. Equivalently, for a nonzero
-     :math:`\mathbf{A}_{ij}` the contribution is
-
-     .. math::
-         \bigg[\frac{\partial \mathcal{L}}{\partial \mathbf{A}}\bigg]_{ij}
-         \;=\; -\, (\mathbf{G}_B)_{i,:} \,\cdot\, \mathbf{x}_{j,:}.
-
-    Parameters
-    ----------
-    A : torch.Tensor, sparse COO or CSR, shape ``(n, n)``
-        Sparse square coefficient matrix. Must be invertible (or suitable) for the
-        chosen solver. All tensors must be on the same device.
-    B : torch.Tensor, dense (strided), shape ``(n,)`` or ``(n, k)``
[…]
