### 2026-07-14 121b141
feat: BatchedCSR descriptor (_batched.py)
Implements architecture.md §3's canonical internal sparse format exactly,
with naming.md §2 vocabulary (descriptor, folded row, local column,
nse_total/nse_per_item, B=1 encodes unbatched). BatchedCSR and BatchedCSC
are frozen dataclasses with lazy cached row_indices/col_indices,
transposed (CSR<->CSC), and a plans dict placeholder for commit 16's SpSM
plan cache.

from_torch() accepts 2D CSR (zero-copy), 2D/3D COO (coalesced internally
via utils.convert's building blocks), torch batched CSR (equal nse), and
list-of-CSR (ragged-native); to_torch(like=.
[…]
