+++ torchsparsegradutils/_batched.py
@@ -1,0 +1,459 @@
+"""BatchedCSR / BatchedCSC descriptors — the canonical internal sparse format.
+
+Architecture: architecture.md §3. Vocabulary: naming.md §2 (descriptor, folded
+row, local column, ``nse_total``/``nse_per_item``, ``B = 1`` encodes
+unbatched). This module is private: nothing outside the front package
+re-exports it, and no op wires it in yet — that starts in later commits
+(``torch.library`` op schemas take plain dense tensors, per architecture §2;
+the descriptor is how those tensors get assembled/disassembled from user
+layouts).
+
+One frozen descriptor replaces the block-diag batching hack everywhere, with
+ragged ``nse`` per batch item as first-class:
+
+- ``values (nse_total,)`` · ``rowptr (B·n+1,)`` absolute over folded rows
+  ``row_global = b · n + r`` · ``col (nse_total,)`` **local** columns in
+  ``[0, m)`` · ``shape (B, n, m)``; ``B = 1`` encodes unbatched.
+- Local (unoffset) columns keep int32 viable and let kernels address
+  ``Bdense[b, col, :]`` directly; the batch of an entry recovers as
+  ``row_global // n``.
+"""
+
+from __future__ import annotations
+
+import functools
+from dataclasses import dataclass
+from typing import List, Sequence, Tuple, Union
+
+import torch
+
+from torchsparsegradutils.utils.convert import _compress_row_indices, _sort_coo_indices
+
+# int32 index arrays can address at most 2**31 - 1 positions; the eligibility
+# bound (architecture.md §3 / naming.md §2) checks nse_total, rowptr length,
+# and n_cols all stay strictly under this.
+_INT32_BOUND = 2**31
+
+_ACCEPTED_LOGICAL_FORMS = (
+    "a 2D sparse matrix (n_rows, n_cols) in COO or CSR layout, a batched sparse "
+    "matrix (batch_size, n_rows, n_cols) as 3D COO (leading sparse dim logically "
+    "interpreted as batch, ragged nse per item allowed) or as batched CSR (equal "
+    "nse per item), or a list of 2D CSR sparse matrices (ragged nse per item "
+    "allowed)"
+)
+
+
+def _invalid_source(received: str) -> ValueError:
+    """naming.md §1 error template: accepted logical forms + received shape/layout."""
+    return ValueError(f"BatchedCSR.from_torch/to_torch accepts {_ACCEPTED_LOGICAL_FORMS}; got {received}.")
+
+
+def _resolve_index_dtype(source_dtype: torch.dtype, *, nse_total: int, rowptr_len: int, n_cols: int) -> torch.dtype:
+    """Apply the int32 storage rule (architecture.md §3): keep int32 only when
+    the source already used int32 *and* max(nse_total, rowptr_len, n_cols) is
+    still below 2**31 — int64 sources are never downcast, and eligible int32
+    sources are never silently upcast to int64.
+    """
+    if source_dtype == torch.int64:
+        return torch.int64
+    if source_dtype != torch.int32:
+        raise TypeError(f"Unsupported index dtype {source_dtype}; expected torch.int32 or torch.int64.")
+    bound = max(nse_total, rowptr_len, n_cols)
+    return torch.int32 if bound < _INT32_BOUND else torch.int64
+
+
+def _fold_coo_to_csr(
+    batch: torch.Tensor,
+    row_local: torch.Tensor,
+    col_local: torch.Tensor,
+    values: torch.Tensor,
+    *,
+    batch_size: int,
+    n_rows: int,
+) -> Tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
+    """Sort scattered (batch, row_local, col_local) coordinates and compress
+    them into a folded CSR ``(rowptr, col, values)`` triple, reusing
[…]
