+++ tests/test_batched.py
@@ -1,0 +1,492 @@
+"""Tests for the BatchedCSR / BatchedCSC descriptors (torchsparsegradutils/_batched.py).
+
+Covers: every from_torch() input form + to_torch(like=...) round-trip, folded-row
+math (row_global = b * n_rows + r), ragged batches (incl. an empty batch item),
+B=1 unbatched, the int32 eligibility rule, and naming.md §1 error-message
+conformance. Property-based round-trip exactness over random shapes/nse/dtypes
+via hypothesis (plain @given — profiles arrive commit 11).
+
+Note: torch's ``sparse_coo_tensor``/``.coalesce()`` always coerce indices to
+int64 regardless of the dtype passed in (map.md invariant 6's "upstream COO
+int64 coercion excepted"). So genuine int32-index fixtures below are built
+directly as CSR (``torch.sparse_csr_tensor``), never via a COO detour.
+"""
+
+import itertools
+
+import pytest
+import torch
+from hypothesis import given, settings
+from hypothesis import strategies as st
+from test_config import INDEX_DTYPES, VALUE_DTYPES
+
+from torchsparsegradutils._batched import BatchedCSC, BatchedCSR, _resolve_index_dtype
+from torchsparsegradutils.utils.convert import stack_csr
+
+# --------------------------------------------------------------------------- #
+# Fixture builders — independent of _batched.py, so round-trip tests aren't
+# circularly validated against the code under test.
+# --------------------------------------------------------------------------- #
+
+
+def _make_2d_coo(coords, n_rows, n_cols, value_dtype):
+    """A 2D COO tensor. Indices are always int64 (upstream coercion)."""
+    if coords:
+        idx = torch.tensor(coords, dtype=torch.int64).t().contiguous()
+        vals = torch.arange(1, len(coords) + 1, dtype=value_dtype)
+    else:
+        idx = torch.empty((2, 0), dtype=torch.int64)
+        vals = torch.empty((0,), dtype=value_dtype)
+    return torch.sparse_coo_tensor(idx, vals, (n_rows, n_cols)).coalesce()
+
+
+def _make_batched_coo(coords, batch_size, n_rows, n_cols, value_dtype):
+    """A 3D (batched) COO tensor. Indices are always int64 (upstream coercion)."""
+    if coords:
+        idx = torch.tensor(coords, dtype=torch.int64).t().contiguous()
+        vals = torch.arange(1, len(coords) + 1, dtype=value_dtype)
+    else:
+        idx = torch.empty((3, 0), dtype=torch.int64)
+        vals = torch.empty((0,), dtype=value_dtype)
+    return torch.sparse_coo_tensor(idx, vals, (batch_size, n_rows, n_cols)).coalesce()
+
+
+def _make_2d_csr(coords, n_rows, n_cols, index_dtype, value_dtype):
+    """A 2D CSR tensor built directly from (row, col) coords, with a genuinely
+    controlled index dtype (int32 survives — no COO detour)."""
+    coords = sorted(set(coords))
+    crow = [0] * (n_rows + 1)
+    for r, _c in coords:
+        crow[r + 1] += 1
+    for i in range(1, n_rows + 1):
+        crow[i] += crow[i - 1]
+    crow_t = torch.tensor(crow, dtype=index_dtype)
+    col_t = torch.tensor([c for _r, c in coords], dtype=index_dtype)
+    vals = torch.arange(1, len(coords) + 1, dtype=value_dtype)
+    return torch.sparse_csr_tensor(crow_t, col_t, vals, (n_rows, n_cols))
+
[…]
