### 2026-07-14 d3847a8
feat: tsgu:: op schemas, fake kernels, autograd registration, dispatch probe
Adds torchsparsegradutils/_dispatch.py (tsgu library namespace, CUDA backend
probe + __backend_api_version__ handshake, TSGU_DISABLE_CUDA_BACKEND=1
escape hatch, backend_available()), and all nine tsgu:: op definitions from
map.md's Kernel-routing table (spmm/sddmm in ops/matmul.py, spsm in
ops/triangular_solve.py, seglse/seglse_bwd/seglse_bidir/seglse_bidir_bwd in
ops/logsumexp.py, grouped_gemm in ops/indexed_matmul.py, coo2csr in
utils/convert.py). Schemas take plain dense tensors (BatchedCSR's
vals/rowptr/col plus 
[…]
