+++ torchsparsegradutils/ops/triangular_solve.py
@@ -1,8 +1,132 @@
 from typing import cast
 
 import torch
+from torch import Tensor
 
 from torchsparsegradutils.utils import convert_coo_to_csr, sparse_block_diag, sparse_block_diag_split, stack_csr
+
+# ---------------------------------------------------------------------------
+# tsgu::spsm — op schema, fake kernel, autograd registration.
+# spec/commit.md Phase 1 #9; routing verbatim from spec/map.md "Kernel
+# routing". Schema takes plain dense tensors only (architecture.md §2). The
+# `sparse_triangular_solve` wrapper below still calls
+# `_legacy_sparse_triangular_solve`; nothing wires this op in until commit 16
+# (spec/commit.md Phase 3), which also adds the real SpSM analysis-plan
+# cache (architecture.md §3).
+#
+# No CUDA/CPU implementation is registered in this commit — the op exists
+# only as schema + fake (meta) kernel and raises NotImplementedError if
+# actually invoked. Its `register_autograd` backward below therefore also
+# references torch.ops.tsgu.spsm / torch.ops.tsgu.sddmm with no real
+# implementation yet; nothing exercises those code paths until later.
+# ---------------------------------------------------------------------------
+
+
+@torch.library.custom_op("tsgu::spsm", mutates_args=())
+def spsm(
+    vals: Tensor,
+    rowptr: Tensor,
+    col: Tensor,
+    rhs: Tensor,
+    B: int,
+    n: int,
+    upper: bool,
+    unitriangular: bool,
+    transpose: bool,
+) -> Tensor:
+    r"""Batched sparse triangular solve (map.md: ``sparse_triangular_solve``
+    forward; also serves its own ``gradB`` via the "transposed plan" —
+    calling this same op again with ``transpose`` flipped, map.md routing;
+    the real analysis-plan cache reuse, architecture.md §3, lands with the
+    kernel in commit 16).
+
+    Solves, per batch item ``b``, :math:`A[b]\,x[b] = \mathrm{rhs}[b]` (or
+    :math:`A[b]^\top\,x[b] = \mathrm{rhs}[b]` when ``transpose=True``) for
+    triangular ``A[b]``.
+
+    Parameters
+    ----------
+    vals : Tensor, shape ``(nse_total,)``
+        Stored values of the triangular matrix (naming.md §2 ``vals``).
+    rowptr : Tensor, shape ``(B * n + 1,)``
+        Absolute CSR pointer over folded rows (naming.md §2 ``rowptr``).
+    col : Tensor, shape ``(nse_total,)``
+        Local column indices in ``[0, n)`` (naming.md §2 ``col``) — ``A`` is
+        square, ``n_rows == n_cols == n``.
+    rhs : Tensor, shape ``(B, n, p)``
+        Right-hand side.
+    B, n : int
+        ``batch_size``, ``n_rows`` (== ``n_cols``).
+    upper, unitriangular, transpose : bool
+        Kept verbatim from the public ``sparse_triangular_solve`` contract
+        (map.md invariant 1).
+
+    Returns
+    -------
+    Tensor, shape ``(B, n, p)``
+        Solution ``x``, same shape as ``rhs``.
+    """
+    raise NotImplementedError(
+        "tsgu::spsm has no implementation registered yet — lands in spec/commit.md Phase 3 (commit 16)."
+    )
+
+
+@spsm.register_fake
+def _spsm_fake(
+    vals: Tensor,
+    rowptr: Tensor,
+    col: Tensor,
+    rhs: Tensor,
+    B: int,
+    n: int,
+    upper: bool,
+    unitriangular: bool,
+    transpose: bool,
[…]
