### 2026-07-14 f4e31b8
build: cuda package scaffold (kernel-builder) + smoke kernel
Scaffolds the torchsparsegradutils-cuda package (architecture.md §5) via HF
kernel-builder, Hub-only distribution (build.md): build.toml (stable-ABI,
edition 5) + flake.nix (github:huggingface/kernels), csrc/registration.cpp,
csrc/common/{batched_csr,dispatch,reduce,stream}.cuh, and one trivial
tsgu::_smoke op (STABLE_TORCH_LIBRARY_IMPL(tsgu, CUDA, m)) wired end-to-end:
schema + fake kernel added to torchsparsegradutils/_dispatch.py (private,
guarded), CUDA impl launches a real kernel via the four common headers. One
dummy NVBench ta
[…]
