### 2026-07-14 db9e20e
docs(spec): record kernel-builder layout facts from commit 10
torch-ext/ is a hard kernel-builder requirement for the Python glue;
NVBench targets are standalone CMake by design (no builder integration).
Also records torch stable-ABI variant naming in use.

Co-Authored-By: Claude Fable 5 <noreply@anthropic.com>
