### 2026-07-14 25e213c
test+bench: gate runner and benchmark harness
Testing side: tests/_tolerances.py (the one per-dtype tolerance policy —
f64 tight rel, f32 rel with reduction-length allowance, ULP bounds for
bf16/fp16); hypothesis profiles (pr=50, nightly=1000, deadline=None in CI,
default local=100) registered in tests/conftest.py; the six-stage gate
runner (tests/gpu_gates.py + gate1..gate6 pytest markers + `tox -e gpu`)
running stages in order, fail-fast, treating an empty marker as a pass.
Stage 1 (smoke) and stage 3 (parity) are populated: the tsgu::_smoke op and
an Oracle-A-vs-Oracle-B self-consistency ch
[…]
