### 2026-07-15 5a45403
feat(cuda): tsgu::seglse + seglse_bwd — sparse_logsumexp live
Commit 12 of the spec/commit.md playbook (kernel-commit template T1-T6),
Family 2 "Segmented logsumexp" (kernels.md) — the pipe-cleaner kernel.

T1 cuda/csrc/kernels/logsumexp/seglse.cu: forward = one warp per folded
segment (rowptr gives contiguous segments for free), per-lane online
log-sum-exp accumulation (common/reduce.cuh OnlineLogSumExp) merged by a
warp shuffle-combine — single pass over vals, no atomics, deterministic.
include_zeros folds m - nse_per_segment structural zeros in as one extra
accumulator at lane 0. Backward =
[…]
