+++ benchmarks/bench_seglse.py
@@ -1,0 +1,155 @@
+"""Op-level acceptance-bar benchmark for tsgu::seglse (spec/commit.md Phase 3
+commit 12; spec/benchmarks.md §3: seglse baselines = pytorch_scatter's
+scatter_logsumexp on equivalent index arrays, and the old pure-PyTorch path
+(tests/oracle's frozen Oracle A)).
+
+    uv run python -m benchmarks.bench_seglse
+
+Runs benchmarks.md §1's protocol (do_bench CUDA-event windowing, memory
+reset/peak/workspace) for one representative CSR configuration on the
+synthetic tier (benchmarks.md §2), and writes one JSON per baseline under
+benchmarks/results/ (backend="custom" -- a real tsgu:: kernel exists to
+attribute "ours" to, unlike the oracle-only rows benchmarks/run.py's
+docstring describes for pre-kernel commits).
+
+pytorch_scatter is optional: if it isn't importable, that row is skipped and
+the run says so loudly instead of fabricating a baseline number.
+"""
+
+from __future__ import annotations
+
+import torch
+
+from benchmarks import harness, memory
+from benchmarks.env import fingerprint
+from benchmarks.results import Result, write_result
+from tests.oracle import oracle_sparse_logsumexp
+
+try:
+    import torch_scatter
+
+    _HAS_TORCH_SCATTER = True
+except ImportError:
+    _HAS_TORCH_SCATTER = False
+
+
+# Synthetic CSR config (benchmarks.md §2 "synthetic batched/ragged" tier,
+# unbatched slice): DLMC-shaped sparsity (~90% sparse), n_rows x n_cols with
+# a p sweep not applicable to seglse (no rhs width -- p column is null per
+# the acceptance-bar table's own seglse rows in benchmarks.md §5).
+_N, _M = 4096, 4096
+_DENSITY = 0.10  # ~90% sparse, DLMC transformer regime
+
+
+def _make_csr(device, value_dtype, index_dtype, seed=0):
+    torch.manual_seed(seed)
+    dense = torch.randn(_N, _M, dtype=value_dtype, device=device)
+    mask = torch.rand(_N, _M, device=device) < _DENSITY
+    dense = dense * mask
+    csr = dense.to_sparse_csr()
+    crow = csr.crow_indices().to(index_dtype)
+    col = csr.col_indices().to(index_dtype)
+    vals = csr.values().clone().requires_grad_(True)
+    return dense, crow, col, vals
+
+
+def _bench_row(*, baseline_name, baseline_ms, ours_ms, bar, matrix, nse, fp, mem=None, backward=False):
+    speedup = baseline_ms / ours_ms if (baseline_ms and ours_ms) else None
+    bar_met = speedup is not None and speedup >= 1.0
+    variant = f"csr·B=1·f32/i32{'·bwd' if backward else ''}"
+    result = Result.build(
+        op="seglse",
+        family="segmented-logsumexp",
+        backend="custom",
+        variant=variant,
+        matrix=matrix,
+        n=_N,
+        m=_M,
+        nse=nse,
+        p=None,
+        fp=fp,
+        baseline_name=baseline_name,
+        baseline_ms=baseline_ms,
+        ours_ms=ours_ms,
+        speedup=speedup,
+        peak_fwd_mb=mem.peak_fwd_mb if mem else None,
+        workspace_mb=mem.workspace_mb if mem else None,
+        bar=bar,
+        bar_met=bar_met,
+    )
+    path = write_result(result)
[…]
