### 2026-07-15 1774a61
feat(cuda): tsgu::seglse_bidir + bwd — sparse_bidir_logsumexp live
Commit 13 of the spec/commit.md playbook (kernel-commit template T1-T6),
Family 2 "Segmented logsumexp" bidirectional variant (kernels.md).

T1 cuda/csrc/kernels/logsumexp/seglse_bidir.cu: single traversal of
(rowptr, col, vals) updating row and column accumulators together
(kernels.md: "one read of values/indices instead of two"). Row reduction:
identical warp-per-folded-row online-LSE as seglse.cu (no atomics,
deterministic). Column reduction: each row's entries land in scattered
columns, so this is kernels.md's "one-pass onl
[…]
