+++ tests/test_sparse_bidir_logsumexp.py
@@ -1,9 +1,14 @@
 import pytest
 import torch
 from packaging.version import parse as parse_version
-from test_config import DEVICES, INDEX_DTYPES, SPARSE_LAYOUTS, VALUE_DTYPES
+from test_config import INDEX_DTYPES, SPARSE_LAYOUTS, VALUE_DTYPES
 
-# Reuse the sibling module's helpers/fixtures so the two suites stay in lock-step.
+# Reuse the sibling module's helpers/fixtures so the two suites stay in
+# lock-step. `device` in particular: spec/commit.md Phase 3 commit 13 wires
+# sparse_bidir_logsumexp to tsgu::seglse_bidir, which is CUDA-only
+# (architecture.md §4 "CUDA-required at runtime", same story as commit 12's
+# tsgu::seglse) -- every test in this module now needs the same CPU-skip
+# fixture test_sparse_logsumexp.py uses, not a plain DEVICES-parametrized one.
 from test_sparse_logsumexp import (
     FWD_LAYOUTS,
     _assert_close,
@@ -12,32 +17,16 @@ from test_sparse_logsumexp import (
     _make_batched_dense_equal_nnz,
     _make_dense,
     _to_layout,
+    device,  # noqa: F401  (re-exported as this module's `device` fixture)
     device_id,
     dtype_id,
     layout_id,
 )
 
 from torchsparsegradutils import sparse_bidir_logsumexp, sparse_logsumexp
-from torchsparsegradutils._dispatch import backend_available
 from torchsparsegradutils.ops.logsumexp import _bidir_2d, _bidir_batched
 
 _NESTED_OK = parse_version(torch.__version__) >= parse_version("2.4")
-
-
-def _skip_if_cpu_needs_seglse(device):
-    """The five tests below compare sparse_bidir_logsumexp (still its
-    pure-PyTorch _legacy_* body -- untouched by spec/commit.md Phase 3 commit
-    12) against sparse_logsumexp as a two-call reference. sparse_logsumexp
-    itself is CUDA-only as of commit 12 (tsgu::seglse, architecture.md §4),
-    so that comparison can't run on cpu -- skip cleanly rather than hit
-    sparse_logsumexp's NotImplementedError, matching test_sparse_logsumexp.py's
-    device fixture."""
-    if not (device.type == "cuda" and backend_available()):
-        pytest.skip(
-            "two-call reference uses sparse_logsumexp, which is CUDA-only as of "
-            "spec/commit.md Phase 3 commit 12 -- no CUDA device and/or loaded "
-            "torchsparsegradutils_cuda backend available here."
-        )
 
 
 @pytest.fixture(params=SPARSE_LAYOUTS, ids=[layout_id(x) for x in SPARSE_LAYOUTS])
@@ -52,11 +41,6 @@ def fwd_layout(request):
 
 @pytest.fixture(params=FWD_LAYOUTS, ids=[layout_id(x) for x in FWD_LAYOUTS])
 def batched_layout(request):
-    return request.param
-
-
-@pytest.fixture(params=DEVICES, ids=[device_id(d) for d in DEVICES])
-def device(request):
     return request.param
 
 
@@ -77,7 +61,6 @@ def include_zeros(request):
 
 def test_matches_two_call_and_dense(fwd_layout, device, value_dtype, index_dtype, include_zeros):
     """(col_lse, row_lse) equals two sparse_logsumexp calls and the dense reference."""
-    _skip_if_cpu_needs_seglse(device)
     dense = _make_dense(device, value_dtype, seed=0)  # 5x4: has an all-zero row and column
     sp = _to_layout(dense, fwd_layout, index_dtype)
     col_lse, row_lse = sparse_bidir_logsumexp(sp, include_zeros=include_zeros)
[…]
