+++ cuda/bench/bench_seglse_bidir.cu
@@ -1,0 +1,214 @@
+// NVBench target for tsgu::seglse_bidir (spec/commit.md Phase 3 commit 13,
+// T3; kernels.md Family 2). Axes are the kernel's tuning knobs
+// (cuda/csrc/kernels/logsumexp/seglse_bidir.cu): number of rows, average
+// nse per row (sparsity), and n_cols (column-axis width, which drives
+// spinlock contention on the atomic column path). Standalone CMake +
+// FetchContent, same pattern as bench_seglse.cu (deliberately independent
+// of the kernel-builder Nix build).
+
+#include <nvbench/nvbench.cuh>
+
+#include <thrust/device_vector.h>
+#include <thrust/host_vector.h>
+
+#include <cmath>
+#include <random>
+
+namespace {
+
+constexpr int kWarpSize = 32;
+constexpr int kWarpsPerBlock = 8;
+constexpr int kThreadsPerBlock = kWarpSize * kWarpsPerBlock;
+constexpr int kFlatThreads = 256;
+
+struct OnlineLSE {
+  float max_val;
+  float sum_exp;
+
+  __device__ __forceinline__ static OnlineLSE identity() { return OnlineLSE{-INFINITY, 0.0f}; }
+
+  __device__ __forceinline__ void update(float value) {
+    if (value <= max_val) {
+      sum_exp += expf(value - max_val);
+      return;
+    }
+    sum_exp = sum_exp * expf(max_val - value) + 1.0f;
+    max_val = value;
+  }
+
+  __device__ __forceinline__ static OnlineLSE combine(OnlineLSE a, OnlineLSE b) {
+    if (a.max_val == b.max_val) {
+      return OnlineLSE{a.max_val, a.sum_exp + b.sum_exp};
+    }
+    if (a.max_val > b.max_val) {
+      return OnlineLSE{a.max_val, a.sum_exp + (b.sum_exp * expf(b.max_val - a.max_val))};
+    }
+    return OnlineLSE{b.max_val, b.sum_exp + (a.sum_exp * expf(a.max_val - b.max_val))};
+  }
+
+  __device__ __forceinline__ float log_sum_exp() const { return max_val + logf(sum_exp); }
+};
+
+__device__ __forceinline__ void atomic_update_lse(int *lock, float *max_arr, float *sum_arr, int32_t *count_arr,
+                                                    int64_t idx, float value) {
+  while (atomicCAS(&lock[idx], 0, 1) != 0) {
+  }
+  __threadfence();
+  float m = max_arr[idx];
+  float s = sum_arr[idx];
+  if (value <= m) {
+    s += expf(value - m);
+  } else {
+    s = s * expf(m - value) + 1.0f;
+    m = value;
+  }
+  max_arr[idx] = m;
+  sum_arr[idx] = s;
+  count_arr[idx] += 1;
+  __threadfence();
+  atomicExch(&lock[idx], 0);
+}
+
+__global__ void bidir_init_kernel(float *padded, int64_t padded_n, float *col_max, float *col_sum,
+                                   int32_t *col_count, int *lock, int64_t col_n) {
+  int64_t idx = static_cast<int64_t>(blockIdx.x) * blockDim.x + threadIdx.x;
+  if (idx < padded_n) {
+    padded[idx] = -INFINITY;
+  }
+  if (idx < col_n) {
+    col_max[idx] = -INFINITY;
+    col_sum[idx] = 0.0f;
+    col_count[idx] = 0;
+    lock[idx] = 0;
+  }
+}
+
+__global__ void bidir_fwd_kernel(float *padded, float const *vals, int32_t const *rowptr, int32_t const *col,
[…]
