+++ cuda/build.toml
@@ -37,6 +37,7 @@ depends = ["torch"]
 src = [
   "csrc/kernels/_smoke/_smoke.cu",
   "csrc/kernels/logsumexp/seglse.cu",
+  "csrc/kernels/logsumexp/seglse_bidir.cu",
   "csrc/common/batched_csr.cuh",
   "csrc/common/dispatch.cuh",
   "csrc/common/reduce.cuh",

[…]
