### 2026-07-15 dc7eb31
perf(cuda): seglse_bidir — replace per-column spinlock with two-phase atomics
Commit 13 shipped correct but failed 2 of 3 benchmarks.md §3 bars (0.85x vs
the old pure-PyTorch bidir, 0.93x vs two tsgu::seglse calls against the
>=1.5x fusion bar): the column accumulator serialised ~nse/m contending
warps per column behind a CAS spinlock.

The online-log-sum-exp recurrence has no atomic form, but its result
decomposes into two operations that do: traversal folds each entry into
col_max via lock-free float/double atomicMax (sign-aware integer
reinterpretation) alongside the unchanged atomics-free 
[…]
