+++ cuda/csrc/kernels/logsumexp/seglse_bidir.cu
@@ -23,25 +23,36 @@
 // contiguity to exploit), so this is exactly the "one-pass online, via
 // atomics" candidate kernels.md's own forward section describes for the
 // single-axis case — applied here to the column axis only, while the row
-// axis takes the atomics-free path in the very same kernel launch. Each
-// entry's value is folded into its column's running (max, sum-of-shifted-
-// exp) accumulator under a per-column spinlock (there is no atomic
-// generalisation of the online-log-sum-exp recurrence itself, only of the
-// critical section guarding it). This is the non-deterministic path
-// kernels.md's "Determinism policy" flags for atomics-based kernels: run
-// order across rows/warps affects float rounding in the column sums (not
-// their mathematical value). No deterministic column fallback is
-// implemented in this commit (open per kernels.md's resolution: a kernel
-// "may" ship an atomic fast path; this one does, documented here rather
-// than silently).
+// axis takes the atomics-free path in the very same kernel launch. The
+// online-log-sum-exp recurrence has no atomic generalisation, but its
+// *result* decomposes into two operations that do: the running max is a
+// plain max (lock-free float/double atomicMax via the sign-aware integer
+// reinterpretation trick below), and once the final max is known the sum
+// of shifted exponentials is a plain sum (native atomicAdd). So the column
+// side is two phases: the traversal kernel folds each entry into
+// `col_max` (and its entry counter) with hardware atomics, then a second
+// entry-sweep kernel accumulates `exp(v - col_max[c])` into `col_sum`.
+// Shifting by the *final* max is numerically at least as good as the
+// spinlocked running-max rescale this replaces (which this design
+// benchmarked far ahead of: per-column CAS spinlocks serialised ~nse/m
+// contending warps per column and lost the benchmarks.md fusion bar).
+// This is the non-deterministic path kernels.md's "Determinism policy"
+// flags for atomics-based kernels: atomicAdd order across warps affects
+// float rounding in the column sums (not their mathematical value). No
+// deterministic column fallback is implemented in this commit (open per
+// kernels.md's resolution: a kernel "may" ship an atomic fast path; this
+// one does, documented here rather than silently).
 //
 // include_zeros: the row side folds its structural-zero count (m minus the
 // row's specified-entry count) in directly, same as seglse. The column
 // side's structural-zero count (n minus that column's specified-entry
 // count) isn't known until every row has been visited, so it is folded in
-// by a small second kernel after the traversal — still only one pass over
-// `vals` (kernels.md's target), the second kernel only ever touches the
-// O(B*m) column accumulators, never `vals` again.
+// by the finalize kernel, which only touches the O(B*m) column
+// accumulators. Total traffic over `vals`/`col` is two sweeps (max phase +
+// sum phase) — more than kernels.md's single-traversal ideal but with no
+// sort and no locks, which is where the fusion win over two seglse calls
+// actually materialises (the dim-0 seglse path must argsort+gather nse
+// entries; this reads them in place).
 //
[…]
