### 2026-07-15 1e040d2
chore: prepare v0.2.4 release
