### 2026-07-16 ee1575c
feat(cuda): tsgu::sddmm
Commit 14 of the spec/commit.md playbook (kernel-commit template T1-T6),
Family 1 "SDDMM (the shared backward)" (kernels.md). Kernel only -- no
wrapper switch (map.md's own commit-14 entry: "nothing routes to it alone");
tested directly against both oracles, a gradcheck probe, and NVBench.

T1 cuda/csrc/kernels/sddmm/sddmm.cu: out[k] = dot(g[row(k),:], mat[col(k),:])
for every specified entry k, one warp per entry (not per row, unlike the
seglse family -- the reduction axis here is p, so each lane strides its
slice of p and a warp-shuffle sum finishes it). The schema (r
[…]
