+++ cuda/csrc/kernels/sddmm/sddmm.cu
@@ -1,0 +1,199 @@
+// tsgu::sddmm — Family 1 "SDDMM (the shared backward)" (kernels.md Family 1;
+// spec/commit.md Phase 3 commit 14).
+//
+// Sampled dense-dense matmul at a fixed sparsity pattern:
+// `out[k] = dot(g[b(k), row(k), :], mat[b(k), col[k], :])` for every
+// specified entry `k` of the pattern given by `rowptr`/`col` — never a dense
+// materialisation, and `out` reuses the pattern's own index arrays (no index
+// allocation, per kernels.md). Serves the shared backward of four ops
+// (`sparse_mm`, `sparse_triangular_solve`, `sparse_generic_solve`,
+// `sparse_generic_lstsq` — map.md routing); the schema itself (op def, fake
+// kernel) already landed in torchsparsegradutils/ops/matmul.py (commit 9)
+// and is not touched here — this file supplies the one CUDA implementation
+// that schema has been missing.
+//
+// Parallelisation (kernels.md Family 1: "one warp per specified entry,
+// warp-reduce over p"): one warp per `k` in `[0, nse_total)`, not one warp
+// per row as in the seglse family — the reduction axis here is `p` (the
+// dense inner dimension), not the segment's member count, so each lane
+// strides its slice of `p` and a warp-shuffle sum-reduction (reduce.cuh's
+// warp_reduce_sum) finishes it. Shared-memory G-row tiling for large `p`
+// (kernels.md's "optional optimisation") is not pursued here — the plain
+// warp path is exercised by the benchmark (bench_sddmm.cu) up to p=512 and
+// is not "clearly poor" at that width, so the optional path is skipped per
+// the family section's own guidance.
+//
+// Row lookup: the schema gives each entry only `col[k]`, not its row —
+// `rowptr` is monotonic non-decreasing over folded rows (naming.md §2:
+// `row_global = b * n + r`), so `row(k)` is recovered with a per-warp binary
+// search (`find_row` below, run once by lane 0, broadcast via warp shuffle)
+// rather than a second precomputed index array — the same technique CSR
+// SDDMM/SpMV implementations generally use to avoid uncompressing `rowptr`
+// into a full per-entry row array. `batch(k) = row(k) / n`, matching
+// common/batched_csr.cuh's `BatchedCSRView::batch_of`.
+//
+// Fusion: `negate` (kernels.md: "negate-and-scale folded in for the solve
+// backwards, avoid a second pass over nse") is applied as a sign flip on the
+// warp's own reduced value before the single write to `out[k]` — no second
+// kernel, no extra pass over `vals`/`col` (there is no `vals` here at all;
+// this op samples `g`/`mat`, not the pattern's stored values).
+//
+// No atomics anywhere (each warp owns exactly one output element) →
+// deterministic by construction; no separate path is needed under
+// `torch.use_deterministic_algorithms(True)` (kernels.md open Q1, same
+// resolution as seglse.cu).
+//
+// f32/f64 x i32/i64 via common/dispatch.cuh, matching every other kernel in
+// this codebase (architecture.md §3).
+//
+// Dense-operand contiguity: `g`/`mat` are read through flat strided-free
+// pointers inside the kernel (row-major `(B, n/m, p)`), so a non-contiguous
[…]
