+++ tests/test_gate4_gradcheck_sddmm.py
@@ -1,0 +1,135 @@
+"""Gate 4 — gradcheck for tsgu::sddmm via a probe op (spec/testing.md "Gates
+& ordering": "gradcheck / gradgradcheck (f64)"; spec/commit.md Phase 3
+commit 14).
+
+``tsgu::sddmm`` itself gets no ``register_autograd`` in this commit
+(ops/matmul.py's commit-9 note: it is only ever used as a backward
+primitive by other ops, never differentiated through directly -- its own
+higher-order-gradient support is "decided at its own kernel commit", i.e.
+here). ``torch.autograd.gradcheck`` needs *some* backward to check against
+a numerical Jacobian, so this module wraps the real kernel call in a small,
+test-local ``torch.autograd.Function`` (:class:`_SddmmProbe`) whose
+``backward`` is a hand-derived, dense-torch-ops implementation of the
+sampled dot product's own adjoint (test-only code, never shipped) --
+*not* another call into the kernel. gradcheck perturbs ``g``/``mat`` and
+re-invokes the probe's ``forward`` (the real CUDA kernel) each time, so a
+passing gradcheck proves the **forward** kernel's output is numerically
+consistent with the analytic adjoint anyone differentiating through this
+sampling pattern would write -- exactly what future commits (15-17) need
+when they call ``tsgu::sddmm`` from inside their own ``register_autograd``.
+"""
+
+from __future__ import annotations
+
+import pytest
+import torch
+
+import torchsparsegradutils  # noqa: F401  (import side-effect: registers tsgu:: ops)
+from tests._sddmm_helpers import make_batched_pattern, random_masks
+from torchsparsegradutils._dispatch import backend_available
+
+pytestmark = pytest.mark.gate4
+
+_SKIP_REASON = (
+    "tsgu::sddmm has no CPU implementation (CUDA-only, architecture.md §4) -- "
+    "needs both a CUDA device and a loaded, version-matched backend."
+)
+requires_cuda_backend = pytest.mark.skipif(not (torch.cuda.is_available() and backend_available()), reason=_SKIP_REASON)
+
+
+class _SddmmProbe(torch.autograd.Function):
+    """Test-only autograd wrapper around ``torch.ops.tsgu.sddmm``. The
+    backward below is a from-scratch, dense-torch-ops derivation of
+    ``out[k] = sign * dot(g[row(k), :], mat[col(k), :])``'s adjoint w.r.t.
+    ``g``/``mat`` -- it does not call back into the kernel or reuse any of
+    its logic, so a gradcheck pass is not circular."""
+
+    @staticmethod
+    def forward(ctx, rowptr, col, g, mat, B, n, m, negate):
+        ctx.save_for_backward(rowptr, col, g, mat)
+        ctx.B, ctx.n, ctx.m, ctx.negate = B, n, m, negate
+        return torch.ops.tsgu.sddmm(rowptr, col, g, mat, B, n, m, negate)
+
+    @staticmethod
+    def backward(ctx, grad_output):
+        rowptr, col, g, mat = ctx.saved_tensors
+        B, n, m, negate = ctx.B, ctx.n, ctx.m, ctx.negate
+        sign = -1.0 if negate else 1.0
+
+        row_g = torch.repeat_interleave(
+            torch.arange(B * n, device=rowptr.device, dtype=torch.int64), (rowptr[1:] - rowptr[:-1]).long()
+        )
+        batch = row_g // n
+        local_row = row_g % n
+        col_long = col.long()
+
[…]
