### 2026-07-16 f69990d
perf(cuda): sddmm — row-tiled G reuse path; unbatched cuSPARSE bar now met
Commit 14 shipped correct but failed its unbatched perf bar (kernels.md
Family 1 baseline cusparseSDDMM): 11.81ms vs 8.66ms = 0.73x on the
4096^2 / nse=1.68M / p=128 acceptance row. The warp-per-entry design was
already at its no-reuse roofline (~150 of ~176 GB/s) — it re-reads
g[row,:] once per specified entry (~410 entries/row at that density), so
the fix is kernels.md's own optional path: "tile G rows through shared
memory".

sddmm_row_kernel: one warp per FOLDED ROW, g[row,:] staged once into
dynamic shared memory, 
[…]
