+++ cuda/bench/bench_sddmm.cu
@@ -1,11 +1,13 @@
 // NVBench target for tsgu::sddmm (spec/commit.md Phase 3 commit 14, T3;
 // kernels.md Family 1). Axes are the kernel's tuning knobs
-// (cuda/csrc/kernels/sddmm/sddmm.cu): total specified entries (`nse_total`,
-// one warp launched per entry), the dense inner dimension `p` (each warp's
-// own reduction length), and batch size `B` (affects the per-warp `rowptr`
-// binary-search depth via `total_rows = B * n`). Standalone CMake +
-// FetchContent, same pattern as bench_seglse.cu (deliberately independent of
-// the kernel-builder Nix build).
+// (cuda/csrc/kernels/sddmm/sddmm.cu): total specified entries (`nse_total`),
+// the dense inner dimension `p` (each warp's reduction length), batch size
+// `B` (affects `total_rows = B * n`), and `path` — 0 = warp-per-entry
+// fallback, 1 = row-tiled hot path (G row staged in shared memory; the
+// launcher's production heuristic picks it when avg entries/row >= 2 and the
+// staging buffer fits 48KB). Standalone CMake + FetchContent, same pattern
+// as bench_seglse.cu (deliberately independent of the kernel-builder Nix
+// build).
 //
 // A vendor-baseline (cuSPARSE cusparseSDDMM, CSR, unbatched) row is *not*
 // duplicated here: benchmarks.md's acceptance layer is the op-level Python
@@ -88,6 +90,51 @@ __global__ void sddmm_kernel(float *__restrict__ out, int32_t const *__restrict_
   }
 }
 
+// Row-tiled hot path — mirrors csrc/kernels/sddmm/sddmm.cu's
+// sddmm_row_kernel: one warp per folded row, G row staged once through
+// dynamic shared memory, one warp-reduced dot per entry of the row.
+__global__ void sddmm_row_kernel(float *__restrict__ out, int32_t const *__restrict__ rowptr,
+                                  int32_t const *__restrict__ col, float const *__restrict__ g,
+                                  float const *__restrict__ mat, int64_t total_rows, int64_t n, int64_t m,
+                                  int64_t p) {
+  extern __shared__ float g_smem_block[];
+
+  int64_t warp_in_block = threadIdx.x / kWarpSize;
+  int64_t lane = threadIdx.x % kWarpSize;
+  int64_t row_g = static_cast<int64_t>(blockIdx.x) * kWarpsPerBlock + warp_in_block;
+  if (row_g >= total_rows) {
+    return;
+  }
+
+  int64_t start = static_cast<int64_t>(rowptr[row_g]);
+  int64_t end = static_cast<int64_t>(rowptr[row_g + 1]);
+  if (start == end) {
+    return;
+  }
+
+  float *g_smem = g_smem_block + warp_in_block * p;
+  float const *g_row = g + row_g * p;
+  for (int64_t j = lane; j < p; j += kWarpSize) {
+    g_smem[j] = g_row[j];
+  }
+  __syncwarp();
+
+  int64_t b = row_g / n;
+  float const *mat_b = mat + b * m * p;
+
+  for (int64_t k = start; k < end; ++k) {
+    float const *mat_row = mat_b + static_cast<int64_t>(col[k]) * p;
+    float partial = 0.0f;
+    for (int64_t j = lane; j < p; j += kWarpSize) {
+      partial += g_smem[j] * mat_row[j];
+    }
+    partial = warp_reduce_sum(partial);
+    if (lane == 0) {
+      out[k] = partial;
[…]
