+++ cuda/csrc/kernels/sddmm/sddmm.cu
@@ -12,16 +12,36 @@
 // and is not touched here — this file supplies the one CUDA implementation
 // that schema has been missing.
 //
-// Parallelisation (kernels.md Family 1: "one warp per specified entry,
-// warp-reduce over p"): one warp per `k` in `[0, nse_total)`, not one warp
-// per row as in the seglse family — the reduction axis here is `p` (the
-// dense inner dimension), not the segment's member count, so each lane
-// strides its slice of `p` and a warp-shuffle sum-reduction (reduce.cuh's
-// warp_reduce_sum) finishes it. Shared-memory G-row tiling for large `p`
-// (kernels.md's "optional optimisation") is not pursued here — the plain
-// warp path is exercised by the benchmark (bench_sddmm.cu) up to p=512 and
-// is not "clearly poor" at that width, so the optional path is skipped per
-// the family section's own guidance.
+// Parallelisation — two kernels, chosen host-side by the launcher:
+//
+// 1. `sddmm_row_kernel` (the hot path): one warp per FOLDED ROW. The warp
+//    stages `g[row, :]` into shared memory once, then walks the row's
+//    entries doing one warp-reduced dot against `mat[col[k], :]` per entry.
+//    This is kernels.md Family 1's "tile G rows through shared memory for
+//    large p" — with ~hundreds of entries per row (the benchmark corpus,
+//    and any real gradient pattern of similar density), the warp-per-entry
+//    design re-reads the same G row once PER ENTRY, which put commit 14's
+//    first cut at 0.73x of cuSPARSE (11.81ms vs 8.66ms, 4096^2/1.68M/p=128):
+//    it was already at the no-reuse roofline (~150 GB/s of ~176 GB/s peak),
+//    so the only way forward was to stop paying for G. Row-tiling reads G
+//    exactly once per row (n*p total instead of nse*p) — a ~2x traffic cut
+//    at that density. Row lookup also becomes free (row = warp id), killing
+//    the per-entry binary search.
+//
+// 2. `sddmm_kernel` (fallback): one warp per SPECIFIED ENTRY with a per-warp
+//    binary search for the row — kernels.md's "good for small p" baseline
+//    design, kept for the patterns where row-tiling has nothing to reuse or
+//    can't run: near-diagonal patterns (avg < 2 entries/row — the G-row
+//    staging would be pure overhead and warp-per-row underparallelises),
+//    and p too large for the shared-memory budget (row of f64 at
+//    kWarpsPerBlock rows/block must fit the 48KB default smem window).
+//    It is also perfectly load-balanced, which the row kernel is not —
+//    acceptable because skewed-row load balancing is spmm's problem
+//    (merge-path, kernels.md Family 3), not the backward's.
+//
+// Both kernels write each output element from exactly one warp — no atomics,
+// both deterministic, so the heuristic switch is invisible to the
+// determinism contract (gate 5 run-twice holds regardless of path).
 //
 // Row lookup: the schema gives each entry only `col[k]`, not its row —
 // `rowptr` is monotonic non-decreasing over folded rows (naming.md §2:
@@ -132,6 +152,56 @@ __global__ void sddmm_kernel(scalar_t *__restrict__ out, index_t const *__restri
   }
 }
 
+// Row-tiled hot path (module comment §1): one warp per folded row, G row
[…]
