### 2026-07-16 3522b9e
feat(cuda): tsgu::spmm — sparse_mm live
Commit 15 of the spec/commit.md playbook (kernel-commit template T1-T6),
Family 3 "Vendor-baseline forwards" SpMM row (kernels.md), architecture.md
§5's own annotation: "warp-per-folded-row, column-tiled". sparse_mm forward
now dispatches to tsgu::spmm; its backward (gradA via tsgu::sddmm at A's own
pattern, gradB via tsgu::spmm on the cached CSC transpose) was already wired
in ops/matmul.py's register_autograd since commit 9 — this commit supplies
the missing CUDA implementation both calls needed.

T1 cuda/csrc/kernels/spmm/spmm.cu: one warp per folded 
[…]
