### 2026-07-16 a28fd95
perf(cuda): spmm — templated column tiles + shuffle-staged entries; all bars now met
Commit 15 shipped correct but failed all three acceptance bars: 0.74x vs
cuSPARSE SpMM (p=128), 0.65x vs SpMV (p=1), 0.63x batched vs block-diag+
cuSPARSE. Two causes, both instruction-issue waste in the hot loop:

1. Fixed-width unrolls predicated at runtime: the wide path always ran a
   16-tile (512-column) unrolled FMA loop with per-tile bounds guards, so
   at p=128 only 4 of 16 iterations did work and at p=32 (the batched
   row) 1 of 16; the narrow path ran 32 guarded iterations even at p=1.
2. Serial s
[…]
