+++ cuda/csrc/kernels/spmm/spmm.cu
@@ -28,32 +28,39 @@
 // atomics or a second reduction pass. Warp-per-row sidesteps that entirely by
 // construction, not as a follow-up perf fix.
 //
-// Two kernels, chosen host-side by `p`:
+// Two kernels, chosen host-side by `p`, both templated on the column-tile
+// count so the hot loop's unroll width matches the actual `p` at COMPILE
+// time (this file's first cut instantiated one fixed 16-tile/32-column
+// unroll and predicated the unused iterations away at runtime — at p=128
+// only 4 of 16 wide-path iterations did work and at p=1 only 1 of 32
+// narrow-path iterations did, a 3-30x instruction-issue overhead per entry
+// that put every acceptance row at 0.63-0.74x of its baseline; the
+// templated dispatch below removes the dead issue slots outright):
 //
-// 1. `spmm_wide_kernel` (p >= kWarpSize): each lane owns a strided slice of
-//    the `p` output columns (`j = lane, lane + 32, lane + 64, ...`), and
-//    walks the row's `nse_row` entries ONCE, accumulating
-//    `acc[t] += vals[k] * dense[b, col[k], lane + t*32]` for each of its
-//    owned columns into small per-lane registers (`kMaxRegTiles` = 16,
-//    covering p up to 512 -- benchmarks.md's own p sweep ceiling -- in one
-//    pass over the row; wider p re-walks the row once per extra 512-wide
-//    tile, re-reading only the cheap `col`/`vals` arrays, never re-reading
-//    `out` or repeating any `dense` read). Empty rows (`start == end`) fall
-//    straight through to the unconditional zero-write below -- naming.md/
-//    this commit's own instruction: "out is dense -- every row must be
-//    initialized", never skipped like sddmm's nse-aligned output is.
+// 1. `spmm_wide_kernel<kTiles>` (p > 16; kTiles = ceil(p/32) rounded
+//    up to {1,2,4,8,16}): each lane owns a strided slice of the `p` output
+//    columns (`j = lane, lane + t*32`), and walks the row's entries ONCE,
+//    accumulating `acc[t] += vals[k] * dense[b, col[k], j]` in registers.
+//    The row's `col`/`vals` entries are staged CHUNK-WISE through warp
+//    shuffles: each lane loads one entry of a 32-entry chunk (coalesced
+//    reads of both arrays), then the chunk is replayed to all lanes via
+//    `__shfl_sync` broadcast — no lane ever issues the serial per-entry
+//    scalar loads the first cut paid for. p beyond kTiles*32 (i.e. > 512)
+//    re-walks the row once per extra 512-wide tile in the kTiles=16
+//    instantiation, re-reading only `col`/`vals`, never re-reading `out` or
+//    repeating any `dense` read. Empty rows fall through to the
+//    unconditional zero-write (out is dense — every row initialized, never
+//    skipped like sddmm's nse-aligned output).
 //
-// 2. `spmm_narrow_kernel` (p < kWarpSize, the SpMV-shaped regime, `p = 1`
-//    included): column-tiling one lane per output column would waste
-//    31/32 lanes when p=1 -- kernels.md/this commit's own guidance ("must
-//    not waste a full warp's lanes on one column"). Instead this path
-//    parallelises over the row's ENTRIES (the classic CSR "vector" SpMV
-//    kernel): each lane strides over a disjoint subset of the row's
[…]
