+++ cuda/bench/bench_spmm.cu
@@ -38,8 +38,11 @@ __device__ __forceinline__ float warp_reduce_sum(float val) {
   return val;
 }
 
-// Mirrors csrc/kernels/spmm/spmm.cu's spmm_wide_kernel (p >= kWarpSize):
-// column-tiled, one warp per folded row, row's entries walked once per tile.
+// Mirrors csrc/kernels/spmm/spmm.cu's spmm_wide_kernel (p > 16): column-
+// tiled, one warp per folded row, entries staged in 32-entry chunks via
+// warp-shuffle broadcast, kTiles = compile-time unroll width chosen by the
+// dispatcher below ({1,2,4,8,16}, smallest with kTiles*32 >= p).
+template <int kTiles>
 __global__ void spmm_wide_kernel(float *__restrict__ out, int32_t const *__restrict__ rowptr,
                                   int32_t const *__restrict__ col, float const *__restrict__ vals,
                                   float const *__restrict__ dense, int64_t total_rows, int64_t n, int64_t m,
@@ -57,33 +60,49 @@ __global__ void spmm_wide_kernel(float *__restrict__ out, int32_t const *__restr
 
   float *out_row = out + row_g * p;
   float const *dense_b = dense + b * m * p;
-  int64_t const tile_width = static_cast<int64_t>(kWarpSize) * kMaxRegTiles;
+  int64_t const tile_width = static_cast<int64_t>(kWarpSize) * kTiles;
 
   for (int64_t col_base = 0; col_base < p; col_base += tile_width) {
-    float acc[kMaxRegTiles];
+    float acc[kTiles];
 #pragma unroll
-    for (int t = 0; t < kMaxRegTiles; ++t) acc[t] = 0.0f;
+    for (int t = 0; t < kTiles; ++t) acc[t] = 0.0f;
 
-    for (int64_t k = start; k < end; ++k) {
-      float v = vals[k];
-      float const *mat_row = dense_b + static_cast<int64_t>(col[k]) * p + col_base;
+    bool valid[kTiles];
 #pragma unroll
-      for (int t = 0; t < kMaxRegTiles; ++t) {
-        int64_t j = static_cast<int64_t>(lane) + static_cast<int64_t>(t) * kWarpSize;
-        if (col_base + j < p) acc[t] += v * mat_row[j];
+    for (int t = 0; t < kTiles; ++t) {
+      valid[t] = col_base + static_cast<int64_t>(lane) + static_cast<int64_t>(t) * kWarpSize < p;
+    }
+
+    for (int64_t chunk = start; chunk < end; chunk += kWarpSize) {
+      int64_t kk = chunk + lane;
+      float v_lane = kk < end ? vals[kk] : 0.0f;
+      int64_t c_lane = kk < end ? static_cast<int64_t>(col[kk]) : 0;
+      int chunk_len = static_cast<int>(min(static_cast<int64_t>(kWarpSize), end - chunk));
+
+      for (int i = 0; i < chunk_len; ++i) {
+        float v = __shfl_sync(0xffffffffU, v_lane, i);
+        int64_t c = __shfl_sync(0xffffffffU, c_lane, i);
+        float const *mat_row = dense_b + c * p + col_base + lane;
+#pragma unroll
+        for (int t = 0; t < kTiles; ++t) {
+          if (valid[t]) acc[t] += v * mat_row[static_cast<int64_t>(t) * kWarpSize];
+        }
       }
     }
 
 #pragma unroll
-    for (int t = 0; t < kMaxRegTiles; ++t) {
[…]
