### 2026-07-16 e62183d
feat(cuda): tsgu::spsm + descriptor plan cache — sparse_triangular_solve live
Commit 16 of the spec/commit.md playbook (kernel-commit template T1-T6),
Family 3 "Vendor-baseline forwards" Triangular SpSM row (kernels.md). Custom
provenance (not vendor-scaffold): v1 is a from-scratch, level-scheduled
SpTRSV — no cuSPARSE handle anywhere in this file (advisor decision recorded
in this commit's brief: linking cuSPARSE into the stable-ABI extension is an
unforced build risk; the post-migration backlog's "custom SpTRSV v2 only if
warm SpSM loses" is a later rework, not this one).

T1 cuda/csrc/kerne
[…]
