+++ cuda/csrc/kernels/spsm/spsm.cu
@@ -1,0 +1,238 @@
+// tsgu::spsm — Family 3 "Vendor-baseline forwards", Triangular SpSM row
+// (kernels.md Family 3; spec/commit.md Phase 3 commit 16). Provenance:
+// **custom**, not vendor-scaffold — goal.md's scaffold rule permits a
+// cuSPARSE-wrapping v1, but this implementation is a from-scratch,
+// level-scheduled SpTRSV; no cuSPARSE handle is opened anywhere in this
+// file (advisor decision recorded in this commit's brief: linking cuSPARSE
+// into the stable-ABI extension is an unforced build risk, and the
+// post-migration backlog's "custom SpTRSV v2 only if warm SpSM loses"
+// applies to a LATER rework, not this one).
+//
+// Solves, per folded row/owner (naming.md §2), a triangular system: for
+// `!transpose`, `A[b] x[b] = rhs[b]`; for `transpose`, `A[b]^T x[b] =
+// rhs[b]` — both with `A` triangular (`upper`/`unitriangular` flags kept
+// verbatim, map.md contract). Also serves this same op's own `gradB`
+// (map.md routing: "transposed plan" — the same op with `transpose`
+// flipped, torchsparsegradutils/ops/triangular_solve.py's
+// `_spsm_backward`, commit 9).
+//
+// Design (kernels.md Family 3 SpSM row, expanded to implementation level in
+// this commit's brief) — two phases:
+//
+// 1. ANALYSIS (plan.h/plan.cpp): host-side, cached, NOT the hot path. Every
+//    call reaches this same C++ entry point (the frozen schema has no plan-
+//    tensor argument — plan.h's own module comment has the full "why"), so
+//    the plan cache lives HERE, keyed on rowptr/col tensor identity,
+//    realizing architecture.md §3's "the plan cache lives on the
+//    descriptor" via the fact that a Python BatchedCSR holds the same
+//    rowptr/col tensors for its whole lifetime. Produces a level schedule:
+//    `level[owner] = 1 + max(level[dep])` over the owner's strict off-
+//    diagonal dependencies, rows/columns grouped into `row_order` by level.
+//
+// 2. SOLVE (this file): one kernel LAUNCH PER LEVEL (kernels.md: "one
+//    launch per level is the simple, correct v1"), all on the current
+//    stream — CUDA's own same-stream launch ordering is what guarantees a
+//    level's dependencies (all in strictly earlier levels, by the level
+//    recurrence) are fully written before this level's kernel begins, with
+//    NO host sync anywhere in this function (the analysis's one host
+//    round-trip already happened, cached, before any solve kernel ever
+//    launches — genuinely amortised on a cache hit). Within a level, ONE
+//    WARP PER OWNER (mirroring spmm.cu's "one warp per folded row" — same
+//    reasoning: every output row is owned by exactly one warp, so there
+//    are NO ATOMICS anywhere in this kernel -> deterministic by
+//    construction, same resolution as every other kernel in this codebase
+//    for kernels.md's open Q1 (no separate path needed under
+//    torch.use_deterministic_algorithms(True)) -- determinism additionally
+//    depends on the level ORDER being fixed across calls, which it is: the
+//    plan is built once and reused, and even a freshly-rebuilt plan for an
+//    unchanged pattern recomputes the identical level assignment (the
[…]
