+++ tests/test_gate4_gradcheck_spsm.py
@@ -1,0 +1,195 @@
+"""Gate 4 — gradcheck + gradgradcheck for ``sparse_triangular_solve`` (the
+public wrapper, spec/testing.md "Gates & ordering": "gradcheck /
+gradgradcheck (f64)"; spec/commit.md Phase 3 commit 16).
+
+Gradcheck: ``tsgu::spsm`` already has a real ``register_autograd``
+(torchsparsegradutils/ops/triangular_solve.py, commit 9; transpose-indexing
+resolved commit 16), composed with PyTorch's own sparse-tensor ``values()``
+adjoint at the wrapper boundary (same pattern as ``sparse_mm``'s
+``test_gate4_gradcheck_spmm.py``) -- ``torch.autograd.gradcheck`` runs
+directly on ``sparse_triangular_solve``.
+
+Gradgradcheck (PR #85: "solve ops support higher-order" -- verified here,
+not assumed):
+
+- **w.r.t. B (dense) alone: SUPPORTED and verified below.** ``gradB`` is
+  computed by calling ``torch.ops.tsgu.spsm`` again (module docstring:
+  "the same op with `transpose` flipped") -- itself a ``register_autograd``
+  op, so this recursive call stays fully in the autograd graph to any
+  order.
+- **w.r.t. A's sparse values: NOT supported, same as the pre-existing
+  ``sparse_mm`` limitation (commit 15) -- documented, not silently
+  papered over.** ``gradA`` is computed via ``torch.ops.tsgu.sddmm``
+  (negate epilogue), and ``tsgu::sddmm`` has **no ``register_autograd``**
+  (torchsparsegradutils/ops/matmul.py, commit 9's own note: "in this
+  commit it is only ever used as a backward primitive... its own
+  higher-order-gradient support is decided at its own kernel commit" --
+  commit 14 shipped ``tsgu::sddmm`` without one, and no later commit
+  (15, 16) has added it; that is Family-1's own decision to make, not
+  this commit's). The practical effect, verified directly below via
+  ``create_graph=True`` (``torch.autograd.gradgradcheck`` itself cannot
+  even run this comparison -- PyTorch's gradcheck machinery raises
+  "Sparse output is not supported at gradcheck yet" on the sparse gradA
+  output before ever reaching the second-order check): the first-order
+  ``gradA`` tensor comes back with ``requires_grad=False`` (no
+  ``grad_fn``), so ``torch.autograd.grad`` on any expression combining it
+  cleanly raises ``"...does not require grad and does not have a
+  grad_fn"`` -- a clear, catchable failure, never a silently wrong zero
+  second derivative.
+
+  The **old (Oracle A) implementation had the identical limitation for a
+  different reason**: ``SparseTriangularSolve.forward`` calls
+  ``A.detach()``/``B.detach()`` before the solve and only ``save_for_backward``s
+  the *detached* ``A``/``x`` -- so any use of those saved tensors inside
+  ``backward`` (index_select, elementwise ops) has no graph connection back
+  to the original ``A``/``B`` either, and double-backward through the
+  legacy path was never possible. This commit does not regress anything
+  that used to work; it inherits a documented, pre-existing gap.
+"""
+
+from __future__ import annotations
+
+import pytest
+import torch
+
+import torchsparsegradutils  # noqa: F401  (import side-effect: registers tsgu:: ops)
+from torchsparsegradutils import sparse_triangular_solve
+from torchsparsegradutils._dispatch import backend_available
+from torchsparsegradutils.utils import rand_sparse_tri
+
+pytestmark = pytest.mark.gate4
+
+_SKIP_REASON = (
[…]
