### 2026-07-16 0a679bb
perf(cuda): spsm — sync-free single-launch solve; cold cuSPARSE bar now met
Commit 16 shipped correct but failed its cold bar (0.84x vs cuSPARSE
SpSM) and only tied warm (1.04x). The cost was structural: one kernel
launch per level, and the acceptance row (n=4096, bandwidth-8 banded) has
one LEVEL PER ROW — the "solve" was 4096 back-to-back ~3.6us launches,
~15ms of launch latency wrapping ~100us of arithmetic.

Fix, three layers, all in spsm_syncfree_kernel (replaces
spsm_level_kernel outright):

1. Sync-free single launch (standard sync-free SpTRSV shape): a fixed
   persistent grid; warps c
[…]
