+++ cuda/csrc/kernels/spsm/spsm.cu
@@ -29,35 +29,48 @@
 //    `level[owner] = 1 + max(level[dep])` over the owner's strict off-
 //    diagonal dependencies, rows/columns grouped into `row_order` by level.
 //
-// 2. SOLVE (this file): one kernel LAUNCH PER LEVEL (kernels.md: "one
-//    launch per level is the simple, correct v1"), all on the current
-//    stream — CUDA's own same-stream launch ordering is what guarantees a
-//    level's dependencies (all in strictly earlier levels, by the level
-//    recurrence) are fully written before this level's kernel begins, with
-//    NO host sync anywhere in this function (the analysis's one host
-//    round-trip already happened, cached, before any solve kernel ever
-//    launches — genuinely amortised on a cache hit). Within a level, ONE
-//    WARP PER OWNER (mirroring spmm.cu's "one warp per folded row" — same
-//    reasoning: every output row is owned by exactly one warp, so there
-//    are NO ATOMICS anywhere in this kernel -> deterministic by
-//    construction, same resolution as every other kernel in this codebase
-//    for kernels.md's open Q1 (no separate path needed under
-//    torch.use_deterministic_algorithms(True)) -- determinism additionally
-//    depends on the level ORDER being fixed across calls, which it is: the
-//    plan is built once and reused, and even a freshly-rebuilt plan for an
-//    unchanged pattern recomputes the identical level assignment (the
-//    analysis has no randomness or races of its own).
+// 2. SOLVE (this file): ONE kernel launch total — a SYNC-FREE persistent-
+//    warp solve (the standard sync-free SpTRSV technique, à la Liu et al.),
+//    replacing commit 16's one-launch-per-level v1. The per-level shape was
+//    correct but its cost was the LEVEL COUNT: a banded matrix (the
+//    acceptance row: n=4096, bandwidth 8) has one level per row, so the
+//    "solve" was 4096 back-to-back kernel launches at ~3.6us each — ~15ms
+//    of pure launch latency for ~100us of arithmetic, which is why warm
+//    only tied cuSPARSE (1.04x) and cold lost (0.84x).
 //
-// The solve loop is intentionally the SIMPLEST correct shape (one register
+//    Shape: a fixed grid of persistent warps; each warp repeatedly claims
+//    the next unprocessed position in `row_order` via an atomic ticket
+//    counter, waits (spin on a per-row done-flag) for the claimed row's
+//    dependencies to be solved, solves it, publishes its done-flag. Two
+//    properties make this safe and exact:
+//
+//    - NO DEADLOCK, regardless of block scheduling: `row_order` is
+//      topological (level-sorted, plan.cpp), so a row's dependencies sit
+//      strictly earlier in `row_order` and their tickets were claimed
+//      before this warp claimed its own. Inductively the minimal
+//      unfinished ticket always has every dependency finished, so some
+//      resident warp always progresses — no reliance on CUDA scheduling
+//      blocks in blockIdx order (the assumption published sync-free
+//      SpTRSVs lean on; the ticket counter removes it).
+//
+//    - STILL DETERMINISTIC: the flag atomics only gate VISIBILITY (when a
+//      row may be read), never participate in a value. Each x row is
+//      written by exactly one warp, and each lane's accumulator walks the
+//      row's entries in the same fixed order every run — bitwise-identical
[…]
