+++ cuda/bench/bench_spsm.cu
@@ -1,11 +1,11 @@
 // NVBench target for tsgu::spsm (spec/commit.md Phase 3 commit 16, T3;
 // kernels.md Family 3 SpSM row). Axes are the SOLVE kernel's tuning knobs
-// (cuda/csrc/kernels/spsm/spsm.cu's spsm_level_kernel): matrix size `n`,
-// the dense rhs width `p`, and `n_levels` — the level-schedule DEPTH, this
-// kernel's dominant cost axis (module comment in spsm.cu: level count/
-// launch latency, not per-row FMA throughput, is what drives SpSM time; a
-// deep chain means many small launches, a shallow/wide schedule means few
-// large ones).
+// (cuda/csrc/kernels/spsm/spsm.cu's spsm_syncfree_kernel): matrix size `n`,
+// the dense rhs width `p`, and `n_levels` — the dependency-chain DEPTH,
+// this kernel's dominant cost axis (module comment in spsm.cu: the serial
+// chain's per-hop latency, not per-row FMA throughput, is what drives SpSM
+// time; the sync-free single launch replaced commit 16's one-launch-per-
+// level v1 whose cost was launch latency times the level count).
 //
 // This target benches the SOLVE kernel directly, with a level schedule
 // already known (built host-side, by construction, from a synthetic
@@ -36,39 +36,97 @@ constexpr int kWarpSize = 32;
 constexpr int kWarpsPerBlock = 8;
 constexpr int kThreadsPerBlock = kWarpSize * kWarpsPerBlock;
 
-// Mirrors csrc/kernels/spsm/spsm.cu's spsm_level_kernel exactly (float
+// Mirrors csrc/kernels/spsm/spsm.cu's spsm_syncfree_kernel exactly (float
 // values, int64_t plan indices — the plan's arrays are always int64
-// regardless of the caller's rowptr/col dtype, per plan.h).
-__global__ void spsm_level_kernel(float *__restrict__ x, int64_t const *__restrict__ eff_ptr,
-                                   int64_t const *__restrict__ eff_dep, int64_t const *__restrict__ eff_val_idx,
-                                   int64_t const *__restrict__ diag_val_idx, int64_t const *__restrict__ row_order,
-                                   int64_t level_start, int64_t level_size, float const *__restrict__ vals,
-                                   float const *__restrict__ rhs, int64_t p, bool unitriangular) {
-  int64_t warp_in_block = threadIdx.x / kWarpSize;
-  int64_t lane = threadIdx.x % kWarpSize;
-  int64_t idx = static_cast<int64_t>(blockIdx.x) * kWarpsPerBlock + warp_in_block;
-  if (idx >= level_size) return;
-  int64_t owner = row_order[level_start + idx];
+// regardless of the caller's rowptr/col dtype, per plan.h): persistent
+// warps claim kRowsPerTicket-row chunks of row_order via an atomic ticket,
+// spin on per-row done-flags for cross-warp dependencies, entry-group ×
+// column lane split for p <= 16.
+constexpr int kRowsPerTicket = 8;
 
-  int64_t start = eff_ptr[owner];
-  int64_t end = eff_ptr[owner + 1];
+__global__ void spsm_syncfree_kernel(float *__restrict__ x, int64_t const *__restrict__ eff_ptr,
+                                     int64_t const *__restrict__ eff_dep, int64_t const *__restrict__ eff_val_idx,
+                                     int64_t const *__restrict__ diag_val_idx, int64_t const *__restrict__ row_order,
[…]
