### 2026-07-17 d8349f0
build(bench): cap compile parallelism by available RAM — OOM guard
NVBench TUs peak 0.7-2.8 GB each in cc1plus; an uncapped wide build (22
jobs on this 20-core/30 GB laptop) held 18.3 GB and OOM'd the machine to
a hard power-off (2026-07-16, kernel log). Derive a Ninja job pool from
available physical RAM at ~3 GB/job (clamped to [2, nproc], override via
-DTSGU_BENCH_COMPILE_JOBS), apply it to compile and link edges, document
-G Ninja as the build generator, and warn when the generator ignores
pools. On this machine the cap resolves to 6 (~5 GB peak).

Co-Authored-By: Claude Fable 5 <noreply@a
[…]
