+++ cuda/bench/CMakeLists.txt
@@ -7,13 +7,43 @@
 # Deliberately standalone from the kernel-builder/Nix build (build.toml only
 # builds the torch extension itself) — NVBench benchmarks measure kernel
 # throughput directly and gain nothing from going through torch::Tensor /
-# Python; build directly:
-#   cmake -S cuda/bench -B cuda/bench/build -DCMAKE_CUDA_ARCHITECTURES=86
+# Python; build directly (Ninja, so the RAM guard below is enforced):
+#   cmake -S cuda/bench -B cuda/bench/build -G Ninja -DCMAKE_CUDA_ARCHITECTURES=86
 #   cmake --build cuda/bench/build
-# (needs CMake >= 3.30 — NVBench's own floor; use `nix shell nixpkgs#cmake`
-# if the system cmake is older, as it is on this machine's Ubuntu 22.04).
+# (needs CMake >= 3.30 — NVBench's own floor; use `nix shell nixpkgs#cmake
+# nixpkgs#ninja` if the system cmake is older, as it is on this machine's
+# Ubuntu 22.04).
 cmake_minimum_required(VERSION 3.30)
 project(tsgu_cuda_bench LANGUAGES CXX CUDA)
+
+# Compile-parallelism guard. NVBench TUs are template-heavy: each cc1plus
+# peaks at ~0.7–2.8 GB RSS, and Ninja's default of nproc+2 jobs (22 on a
+# 20-core box) wants ~18 GB at once — enough to OOM a 30 GB laptop running a
+# desktop session (observed 2026-07-16: global OOM, hard power-off). Bound
+# the pool from *available* RAM at ~3 GB per job so a cold, wide build stays
+# inside memory no matter how `cmake --build` is invoked. Job pools are a
+# Ninja feature; the Makefile generator ignores them but defaults to -j1
+# anyway. Override for a dedicated build box with -DTSGU_BENCH_COMPILE_JOBS=N.
+if(NOT DEFINED TSGU_BENCH_COMPILE_JOBS)
+  cmake_host_system_information(RESULT _tsgu_avail_mb QUERY AVAILABLE_PHYSICAL_MEMORY)
+  cmake_host_system_information(RESULT _tsgu_ncores QUERY NUMBER_OF_LOGICAL_CORES)
+  math(EXPR TSGU_BENCH_COMPILE_JOBS "${_tsgu_avail_mb} / 3072")
+  if(TSGU_BENCH_COMPILE_JOBS LESS 2)
+    set(TSGU_BENCH_COMPILE_JOBS 2)
+  endif()
+  if(TSGU_BENCH_COMPILE_JOBS GREATER ${_tsgu_ncores})
+    set(TSGU_BENCH_COMPILE_JOBS ${_tsgu_ncores})
+  endif()
+endif()
+message(STATUS "tsgu bench: capping concurrent compile jobs at ${TSGU_BENCH_COMPILE_JOBS}")
+set_property(GLOBAL PROPERTY JOB_POOLS "compile_pool=${TSGU_BENCH_COMPILE_JOBS}")
+set(CMAKE_JOB_POOL_COMPILE compile_pool)
+set(CMAKE_JOB_POOL_LINK compile_pool)
+if(NOT CMAKE_GENERATOR MATCHES "Ninja")
+  message(WARNING "tsgu bench: generator is '${CMAKE_GENERATOR}', which ignores "
+    "job pools — the RAM guard above is NOT enforced. Configure with -G Ninja, "
+    "or never build with more than -j ${TSGU_BENCH_COMPILE_JOBS}.")
+endif()
 
 set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CUDA_STANDARD 17)
