### 2026-07-17 5d1e7f3
test(bench): fix VRAM-fallback assertion on CUDA machines
test_corpus_vram_fallback_and_scaling asserted the no-CUDA fallback
constant unconditionally, so it failed on any machine with a GPU (real
VRAM of the dev laptop's RTX A1000 reports ~3.68 GiB != the 4.0 GiB
fallback). Assert the real-VRAM path when CUDA is present and the
fallback only when it is not.

Co-Authored-By: Claude Fable 5 <noreply@anthropic.com>
