### 2026-07-17 53e084e
style: restore ruff-clean tree — format drift + fixture-shadowing ignore
`uv run ruff check .` / `ruff format --check .` (CI's exact lint steps)
were red at HEAD: six files had drifted from ruff's formatting, and
test_sparse_bidir_logsumexp.py's pytest fixture re-export pattern (the
shared `device` fixture imported, then test parameters of the same name)
trips F811. Reformat the drifted files and add a documented per-file
F811 ignore — the shadowing is how pytest fixtures compose, not a
redefinition.

Co-Authored-By: Claude Fable 5 <noreply@anthropic.com>
