### 2026-07-17 110c2b7
refactor: batched Krylov solvers — sparse_generic_solve + lstsq live
spec/commit.md Phase 3 commit 17. No new kernels.

solvers/: all four solvers (linear_cg, minres, bicgstab, lsmr) now
iterate in the canonical batched (batch_size, n, n_rhs) shape with
per-batch/per-column convergence masks (converged or broken-down
columns freeze; bicgstab's per-column Python recursion and lstsq's
per-column lsmr loop are gone). A shared solvers/_matvec.py adapter
canonicalises the right-hand side and the operator: sparse COO/CSR/CSC
operators route every matvec through tsgu::spmm, with the transpose
matvec 
[…]
