+++ torchsparsegradutils/solvers/_matvec.py
@@ -1,0 +1,252 @@
+"""Shared matvec adapter for the batched Krylov solvers (spec/commit.md
+Phase 3 commit 17; map.md §2: solver hot path is SpMV/SpMM).
+
+Every solver in ``solvers/`` canonicalises its right-hand side to one batched
+dense tensor of shape ``(batch_size, n, n_rhs)`` (``B = 1`` encodes unbatched
+— naming.md §2) and drives its iteration through a :class:`BatchedOperator`
+built from whatever the caller passed as the operator:
+
+- a **sparse matrix** (COO/CSR tensor, 2D or batched 3D) routes its matvec
+  through ``tsgu::spmm`` when the CUDA backend can take it (map.md kernel
+  routing: composites call ``tsgu::spmm``; ``spmv`` = ``spmm`` with
+  ``p = 1``). The transpose matvec uses the descriptor's lazily **cached
+  BatchedCSC** (``BatchedCSR.transposed``, architecture.md §3) — built once
+  per operator, reused every iteration. Off-CUDA (or for value dtypes the
+  kernels don't cover) it falls back to plain torch sparse matmul semantics,
+  exactly what the solvers did before this commit.
+- a **dense matrix** uses ``.matmul`` (broadcasts over the batch axis).
+- a **callable** is honoured as-is (map.md contract: pluggable user solvers
+  and closures keep working). The adapter feeds it operands shaped the way
+  the caller's own right-hand side was shaped — a user closure written for
+  ``(n, n_rhs)`` operands never sees the internal batch axis.
+- a **``BatchedCSR`` descriptor** is accepted directly (package-internal
+  callers that already hold one skip the re-extraction).
+
+This module is private to ``solvers/``; nothing outside the front package
+imports it.
+"""
+
+from __future__ import annotations
+
+from typing import Callable, Optional, Tuple, Union
+
+import torch
+
+from torchsparsegradutils._batched import BatchedCSR
+from torchsparsegradutils._dispatch import backend_available
+
+# Value dtypes the CUDA kernels are templated over (architecture.md §3:
+# "kernels templated over {f32,f64} x {i32,i64}").
+_KERNEL_VALUE_DTYPES = (torch.float32, torch.float64)
+
+_SPARSE_LAYOUTS = (torch.sparse_coo, torch.sparse_csr)
+
+OperatorLike = Union[torch.Tensor, BatchedCSR, Callable[[torch.Tensor], torch.Tensor]]
+
+
+def as_batched_rhs(rhs: torch.Tensor) -> Tuple[torch.Tensor, bool, bool]:
+    """Canonicalise a right-hand side to ``(batch_size, n, n_rhs)``.
+
+    Returns ``(rhs_batched, was_vector, was_batched)`` where ``was_vector``
+    records a 1D ``(n,)`` input and ``was_batched`` a 3D
+    ``(batch_size, n, n_rhs)`` input; 2D ``(n, n_rhs)`` input has both False.
+    Use :func:`restore_rhs_shape` to undo.
+    """
+    if rhs.ndim == 1:
+        return rhs.unsqueeze(0).unsqueeze(-1), True, False
+    if rhs.ndim == 2:
+        return rhs.unsqueeze(0), False, False
+    if rhs.ndim == 3:
+        return rhs, False, True
+    raise ValueError(
+        "rhs must be a vector (n,), a matrix of right-hand sides (n, n_rhs), or a "
+        f"batched matrix (batch_size, n, n_rhs); got shape {tuple(rhs.shape)}."
+    )
+
+
+def restore_rhs_shape(result: torch.Tensor, was_vector: bool, was_batched: bool) -> torch.Tensor:
[…]
