### 2026-07-17 b033ea3
feat(cuda): tsgu::grouped_gemm — segment_mm + gather_mm live
spec/commit.md Phase 3 commit 18 (kernel-commit template T1-T6).

Kernel (cuda/csrc/kernels/grouped_gemm/grouped_gemm.cu): one kernel pair
behind the commit-9 schema. Gather mode (reduce=False) is a 32x32-tiled
GEMM with the gather fused into the prologue via an adaptive per-block
vote — a block whose row-tile shares one group (the segment_mm interior)
stages the b k-tile in shared memory and runs the fully tiled inner
loop; mixed tiles (gather_mm's arbitrary idx, segment boundaries) read b
per-row from global through L2, never mater
[…]
