+++ tests/test_gate2_opcheck_grouped_gemm.py
@@ -1,0 +1,131 @@
+"""Gate 2 — opcheck for tsgu::grouped_gemm (spec/testing.md "Gates &
+ordering": "opcheck (all ops x representative inputs)"; spec/commit.md Phase
+3 commit 18).
+
+tsgu::grouped_gemm gets its real CUDA implementation in this commit AND is
+wired into the public ``segment_mm``/``gather_mm`` wrappers with a real
+``register_autograd`` (torchsparsegradutils/ops/indexed_matmul.py, landed in
+commit 9) -- so opcheck here exercises schema, fake-kernel, AND
+autograd-registration consistency for real, including ``requires_grad=True``
+cases (testing.md Pillar 2: "requires_grad on/off").
+
+Mode-specific notes:
+
+- Gather mode (``reduce=False``) accepts arbitrary ``idx``; samples below
+  include repeated, out-of-order indices.
+- Scatter-reduce mode (``reduce=True``) requires ``idx`` sorted
+  non-decreasing (deterministic ordered accumulation, no atomics -- the
+  op's documented contract), so every scatter sample uses sorted ``idx``.
+  Scatter samples run with ``requires_grad=False`` only: reduce=True is a
+  gradient leaf by design (its registered backward raises
+  NotImplementedError, see ``_grouped_gemm_backward``), and opcheck's
+  ``test_aot_dispatch_dynamic`` differentiates any input that requires
+  grad, which would exercise exactly that deliberate raise rather than a
+  real adjoint.
+"""
+
+from __future__ import annotations
+
+import pytest
+import torch
+
+import torchsparsegradutils  # noqa: F401  (import side-effect: registers tsgu:: ops)
+from torchsparsegradutils._dispatch import backend_available
+
+pytestmark = pytest.mark.gate2
+
+_SKIP_REASON = (
+    "tsgu::grouped_gemm has no CPU implementation (CUDA-only, architecture.md §4) -- "
+    "needs both a CUDA device and a loaded, version-matched backend."
+)
+
+
+def _grouped_gemm_cuda_ready() -> bool:
+    # Identical in effect to the other gate files' machine-level condition
+    # once commit 18's kernel is registered; the extra dispatch-key check
+    # keeps this file skipping (not erroring) when the front-package wiring
+    # is present but the cuda/ tree has not yet been rebuilt with the
+    # grouped_gemm kernel (commit 18 is developed in two concurrent lanes).
+    return (
+        torch.cuda.is_available()
+        and backend_available()
+        and torch._C._dispatch_has_kernel_for_dispatch_key("tsgu::grouped_gemm", "CUDA")
+    )
+
+
+requires_cuda_backend = pytest.mark.skipif(not _grouped_gemm_cuda_ready(), reason=_SKIP_REASON)
+
+INDEX_DTYPES = (torch.int32, torch.int64)
+VALUE_DTYPES = (torch.float32, torch.float64)
+
+# Small representative sizes (hard resource guard: tiny tensors only).
+_N, _D1, _D2, _R = 6, 4, 3, 3
+
+
+def _gather_inputs(index_dtype, value_dtype, *, requires_grad=False):
+    a = torch.randn(_N, _D1, dtype=value_dtype, device="cuda", requires_grad=requires_grad)
+    b = torch.randn(_R, _D1, _D2, dtype=value_dtype, device="cuda", requires_grad=requires_grad)
+    # Arbitrary order with repeats -- gather mode's contract allows it.
[…]
