### 2026-07-17 2b0ac3d
perf(cuda): grouped_gemm — register tiles + LDS.128 frags; bar narrowed, still open
The commit-18 geometry (32x32 block tile, 4 outputs/thread) ran at ~6%
of f32 peak: 4 FMAs per 5 shared loads is shared-bandwidth-bound. Both
kernels are templated over (BM, BN, BK, TM, TN) with two configs the
launcher selects at runtime: 64x64x16 with 4x4 register tiles (default;
the only config for f64's gather path) and 128x128x16 with 8x8 register
tiles (f32 wide shapes, D >= 128). The scatter (gradB) kernel further
switches to contiguous per-thread fragments read as 16-byte vectors
(LDS.128, float4/double
[…]
