+++ cuda/csrc/kernels/grouped_gemm/grouped_gemm.cu
@@ -18,15 +18,18 @@
 //    per-row group index with shape `(N,)`; `out[i, :] = a[i, :] @
 //    b[idx[i], :, :]`, shape `(N, D2)`.
 //
-//    Classic tiled GEMM: each block owns a 32(rows)x32(cols) output tile
-//    (256 threads as a (32, 8) layout, 4 output rows per thread), walking
-//    the shared inner dimension D1 in 32-wide chunks with the `a` tile
-//    staged through shared memory every chunk. The gather is fused into the
-//    GEMM prologue via an ADAPTIVE per-block `b`-operand path, decided by a
-//    runtime block vote after the tile's 32 `idx` values are staged:
+//    Classic register-blocked tiled GEMM: each block owns a 64(rows)x64(cols)
+//    output tile (256 threads as a logical (16, 16) layout, each owning a
+//    4x4 register tile strided 16 apart), walking the shared inner dimension
+//    D1 in 16-wide chunks with the `a` tile staged (transposed) through
+//    shared memory every chunk — 16 FMAs per 8 shared loads, the geometry
+//    the first cut's 32x32/4-per-thread shape lacked (see the constants
+//    block). The gather is fused into the GEMM prologue via an ADAPTIVE
+//    per-block `b`-operand path, decided by a runtime block vote after the
+//    tile's 64 `idx` values are staged:
 //
 //    - UNIFORM tile (all valid rows share one group — always true inside a
-//      `segment_mm` segment longer than the 32-row tile, and for tail rows
+//      `segment_mm` segment longer than the 64-row tile, and for tail rows
 //      past N, which are excluded from the vote): the `b` k-tile is staged
 //      through shared memory too, and the inner loop is the fully-tiled GEMM
 //      loop — cuBLAS-shaped arithmetic intensity, no per-row indirection
@@ -60,19 +63,20 @@
 //    out-of-place rows from their groups. Debug builds may check host-side
 //    in the Python wrapper; the kernel trusts the contract.
 //
-//    Grid `(num_groups, ceil(D1/32), ceil(D2/32))`: each block owns one
-//    group's 32x32 tile of the `(D1, D2)` outer-product accumulator, binary-
+//    Grid `(num_groups, ceil(D1/64), ceil(D2/64))`: each block owns one
+//    group's 64x64 tile of the `(D1, D2)` outer-product accumulator, binary-
 //    searches `idx` for its group's row range once, then accumulates over
-//    rows `lo..hi` IN ASCENDING ORDER, staging 32-row chunks of `a` and `b`
-//    through shared memory and accumulating outer products in registers.
+//    rows `lo..hi` IN ASCENDING ORDER, staging 16-row chunks of `a` and `b`
+//    through shared memory and accumulating outer products in 4x4 register
+//    tiles (same thread geometry as gather mode).
 //    Sequential ordered accumulation, exactly one block owning every output
 //    element, no atomics → bitwise deterministic. Empty groups (and
 //    `N == 0`) fall through to the unconditional zero-write — the register
 //    accumulator starts at zero and every in-range output element is written
 //    exactly once, so `new_empty` output is safe.
 //
-// Shared-memory budget: two 32x33 tiles (the +1 pads bank conflicts away)
[…]
