+++ cuda/bench/bench_grouped_gemm.cu
@@ -38,10 +38,12 @@
 
 namespace {
 
-constexpr int kTile = 32;
-constexpr int kBlockRows = 8;
-constexpr int kThreadsPerBlock = kTile * kBlockRows;
-constexpr int kRowsPerThread = kTile / kBlockRows;
+constexpr int kBM = 64;                // block-tile rows (gather: output rows; scatter: D1 extent)
+constexpr int kBN = 64;                // block-tile cols (D2 extent)
+constexpr int kBK = 16;                // inner-dimension chunk (gather: D1; scatter: row chunk)
+constexpr int kThreadsPerBlock = 256;  // logical (16, 16)
+constexpr int kTM = 4;                 // register-tile rows per thread (kBM / 16)
+constexpr int kTN = 4;                 // register-tile cols per thread (kBN / 16)
 
 // --- Mirrors of csrc/kernels/grouped_gemm/grouped_gemm.cu (f32/i32) --------
 
@@ -73,82 +75,127 @@ __device__ __forceinline__ int64_t upper_bound_idx(int32_t const* __restrict__ i
   return lo;
 }
 
-// Mirrors grouped_gemm_gather_kernel: 32x32 output tile per block, (32, 8)
-// threads x 4 rows each, adaptive uniform/mixed b-operand path via block vote.
+// Mirror of grouped_gemm_gather_kernel (f32/i32) — keep in sync with
+// csrc/kernels/grouped_gemm/grouped_gemm.cu.
+template <int BM, int BN, int BK, int TM, int TN>
 __global__ void grouped_gemm_gather_kernel(float* __restrict__ out, float const* __restrict__ a,
                                            float const* __restrict__ b, int32_t const* __restrict__ idx, int64_t N,
                                            int64_t D1, int64_t D2) {
-  __shared__ float a_s[kTile][kTile + 1];
-  __shared__ float b_s[kTile][kTile + 1];
-  __shared__ int64_t idx_s[kTile];
+  // a staged transposed (a_s[kk][row]) so the inner loop's per-thread row
+  // reads are broadcast/conflict-free; +1 pads away bank conflicts on the
+  // transposed store.
+  __shared__ float a_s[BK][BM + 1];
+  __shared__ float b_s[BK][BN + 1];
+  __shared__ int64_t idx_s[BM];
   __shared__ int uniform_s;
 
-  int64_t const row_base = static_cast<int64_t>(blockIdx.x) * kTile;
-  int64_t const col_base = static_cast<int64_t>(blockIdx.y) * kTile;
-  int const tid = static_cast<int>(threadIdx.y) * kTile + static_cast<int>(threadIdx.x);
+  int64_t const row_base = static_cast<int64_t>(blockIdx.x) * BM;
+  int64_t const col_base = static_cast<int64_t>(blockIdx.y) * BN;
+  int const tid = static_cast<int>(threadIdx.x);
+  constexpr int LX = BN / TN;  // column-lane count (thread grid is (LY, LX) = 256)
+  constexpr int LY = BM / TM;  // row-lane count
+  int const tx = tid % LX;     // register-tile column lane
+  int const ty = tid / LX;     // register-tile row lane
 
+  // Stage the tile's idx values once, then block-vote uniformity: rows past
+  // N carry the sentinel -1 and never break the vote (a tail tile of one
+  // long segment still takes the uniform path). row_base < N always holds
+  // (grid is sized from N), so idx_s[0] is a real group.
[…]
