### 2026-07-17 6a0f349
feat(cuda): tsgu::coo2csr — convert utils live; BatchedCSR internal switch
spec/commit.md Phase 3 commit 19 (kernel-commit template T1-T6).

Kernel (cuda/csrc/kernels/convert/coo2csr.cu): fused sort+compress of
batched COO coordinates into folded CSR. Fold+iota, then two-pass
STABLE cub::DeviceRadixSort (pass 1 col full-width, pass 2 folded
row_global bit-limited to ceil(log2(B*n))) — the exact (row_global,
col) lexicographic stable order _fold_coo_to_csr defines — then a
one-thread-per-entry lower-bound compress into rowptr. No atomics,
bitwise deterministic; O(nse) workspaces via the torch a
[…]
