+++ benchmarks/bench_coo2csr.py
@@ -1,0 +1,167 @@
+"""Op-level acceptance-bar benchmark for tsgu::coo2csr (spec/commit.md Phase
+3 commit 19; map.md §3: ``convert_coo_to_csr_indices_values`` is 🟠 VENDOR,
+"Native sort+compress kernel, baseline cuSPARSE Xcoo2csr").
+
+    uv run python -m benchmarks.bench_coo2csr
+
+Runs benchmarks.md §1's protocol (do_bench CUDA-event windowing, memory
+reset/peak/workspace) over an nse x B x index-dtype sweep and writes one
+JSON per row under benchmarks/results/ (backend="custom").
+
+Baseline recorded here: the **pure-torch fallback path** this kernel
+replaces — a two-pass stable ``torch.argsort`` (secondary key ``col``, then
+primary key ``row_global``; each argsort is itself a CUB-backed device
+radix sort) followed by a bincount+cumsum compress. That composition is
+exactly what ``convert_coo_to_csr_indices_values`` / ``BatchedCSR``'s COO
+paths execute when the kernel is absent, so the row's speedup is the
+end-user-visible win of commit 19's internal switch. (The raw cuSPARSE
+``Xcoo2csr`` primitive is not reachable from Python without a bespoke
+binding — it only does the compress step, not the sort — so the vendor
+head-to-head lives in the kernel lane's NVBench target, not here.)
+
+VRAM budget (benchmarks/corpus.peak_budget_gb — this machine has a 4 GB
+card): configurations whose estimated peak exceeds the budget are skipped
+WITH a printed note (benchmarks.md: no silent caps).
+"""
+
+from __future__ import annotations
+
+import torch
+
+from benchmarks import harness, memory
+from benchmarks.corpus import dtype_short, peak_budget_gb
+from benchmarks.env import fingerprint
+from benchmarks.results import Result, write_result
+
+# Sweep axes (this commit's T3/T5 instructions).
+SWEEP_NSE = (100_000, 1_000_000, 5_000_000)
+SWEEP_B = (1, 16)
+INDEX_DTYPES = (torch.int32, torch.int64)
+
+# Fixed logical matrix shape per item; nse sweeps independently (duplicate
+# coordinates may occur at the highest densities — the op's contract keeps
+# them, no dedup, so they are a legitimate sort workload).
+_N = _M = 4096
+_SEED = 20260717
+
+
+def _estimate_peak_bytes(nse: int, B: int, index_dtype: torch.dtype) -> int:
+    """Rough peak-bytes model: 3 input coordinate arrays + 3 outputs
+    (rowptr, col_sorted, permutation) + the baseline's argsort temporaries
+    (each stable argsort materialises keys + an int64 permutation). A
+    documented fudge factor stands in for allocator overhead, like
+    benchmarks/corpus.estimate_peak_bytes does for the sparse ops."""
+    idx = torch.tensor([], dtype=index_dtype).element_size()
+    coords = 3 * nse * idx
+    outputs = (B * _N + 1) * idx + 2 * nse * idx
+    sort_scratch = 4 * nse * 8  # int64 row_global/keys + two argsort permutations
+    return int(2.0 * (coords + outputs + sort_scratch))
+
+
+def _make_inputs(nse: int, B: int, index_dtype: torch.dtype, device: torch.device):
+    gen = torch.Generator(device="cpu").manual_seed(_SEED)
+    batch = torch.randint(0, B, (nse,), generator=gen).to(index_dtype).to(device)
+    row = torch.randint(0, _N, (nse,), generator=gen).to(index_dtype).to(device)
+    col = torch.randint(0, _M, (nse,), generator=gen).to(index_dtype).to(device)
[…]
