### 2026-07-17 39ee77d
test: composite revalidation + e2e benchmarks
spec/commit.md Phase 4 commit 20.

Composite revalidation: the distributions/encoder statistical suites
(Hotelling T2 / Nagao) re-run green on the CUDA-native ops — 1500+
tests incl. 223 CUDA-parameterised, zero composite code changes needed
for correctness, as predicted.

One product change, forced by the pinned regression: the encoder-CSR
backward blow-up (old README) reproduced at 33x COO peak_bwd and
root-caused to torch's own sparse_csr_tensor CONSTRUCTOR autograd,
which densifies the incoming gradient at (n_rows, n_cols) scale
(measured: an O
[…]
