+++ spec/images/make_real_charts.py
@@ -1,0 +1,288 @@
+"""Generate the real-data sweep charts embedded in ../benchmarks.md §6.
+
+Run:  uv run --with matplotlib python spec/images/make_real_charts.py
+
+Successor of the retired make_dummy_charts.py (git history; same palette,
+spine/grid treatment, parity line, endpoint direct-labels — no redesign).
+Every point is read
+from a persisted result JSON under benchmarks/results/; nothing is
+invented. Slots whose §6 sweep (p-sweep over {1..512}, B-sweep, seglse
+density-sweep) has no sweep JSONs yet are filled with the closest sweep
+the real data supports — see the chart docstrings and benchmarks.md §6.
+"""
+
+import json
+import math
+from pathlib import Path
+
+import matplotlib
+
+matplotlib.use("Agg")
+import matplotlib.pyplot as plt  # noqa: E402  (backend must be set first)
+
+OUT = Path(__file__).parent
+RESULTS = OUT.parent.parent / "benchmarks" / "results"
+
+# Palette: fixed hue per kernel family (validated, light surface) — same
+# assignments as benchmarks/viz.py FAMILY_COLORS; e2e hue is new, never
+# recycled from another family.
+SURFACE = "#fcfcfb"
+TEXT = "#0b0b0b"
+TEXT_2 = "#52514e"
+GRID = "#e7e6e2"
+BLUE_SPMM = "#2a78d6"  # spmm + sddmm families
+AQUA_SEGLSE = "#1baf7a"  # seglse family
+RED_SPSM = "#c0392b"  # spsm family
+PURPLE_GG = "#8e5fd6"  # grouped_gemm family
+YELLOW_BASE = "#eda100"  # coo2csr + baseline paths in multi-series charts
+MAGENTA_E2E = "#c9528e"  # e2e composites (rsample, CG)
+
+
+def load(name):
+    return json.loads((RESULTS / name).read_text())
+
+
+def geomean(xs):
+    return math.exp(sum(math.log(x) for x in xs) / len(xs))
+
+
+def styled_axes(figsize=(7, 4)):
+    fig, ax = plt.subplots(figsize=figsize, dpi=160)
+    fig.patch.set_facecolor(SURFACE)
+    ax.set_facecolor(SURFACE)
+    ax.grid(axis="y", color=GRID, linewidth=0.8, zorder=0)
+    for side in ("top", "right"):
+        ax.spines[side].set_visible(False)
+    for side in ("left", "bottom"):
+        ax.spines[side].set_color(GRID)
+    ax.tick_params(colors=TEXT_2, labelsize=9)
+    return fig, ax
+
+
+def source_note(ax, text, loc="left", y=0.02):
+    x, ha = (0.01, "left") if loc == "left" else (0.99, "right")
+    ax.annotate(text, xy=(x, y), xycoords="axes fraction", ha=ha,
+                fontsize=7, color=TEXT_2, alpha=0.8, style="italic")
+
+
+def sweep_chart(fname, title, subtitle, xlabel, x, y, color, log2_x, ymax, source,
+                source_y=0.02):
+    fig, ax = styled_axes()
+
+    # parity reference — neutral, dashed, labelled directly (no legend needed)
+    ax.axhline(1.0, color=TEXT_2, linewidth=1, linestyle=(0, (4, 4)), zorder=1)
+    ax.annotate("parity 1.0×", xy=(0.99, 1.0), xycoords=("axes fraction", "data"),
+                ha="right", va="bottom", fontsize=8, color=TEXT_2)
+
+    # the one series — 2px line, 8px markers; title carries its identity
[…]

[…]
