### 2026-07-17 0e0cf41
refactor!: final purge — no legacy left
spec/commit.md Phase 4 commit 21.

Deletes sparse_block_diag / sparse_block_diag_split
(utils/_block_diag.py) and their exports and unit tests — the block-
diagonal batching workaround existed to fake batched SpMM/solve, and
the native batched kernels (BatchedCSR + tsgu:: ops, commits 12-19)
removed every consumer; the frozen test-only oracle keeps its own copy
for parity baselines, as designed. The four old autograd.Function
classes died in their kernel commits; nothing remains.

Scrubs every historical `_legacy`/`block_diag` mention from package
commen
[…]
