+++ torchsparsegradutils/utils/_block_diag.py
@@ -1,323 +1,0 @@
-# ⚫ RETIRE — deleted by commit 21 (spec/commit.md Phase 4); only _legacy_* op code and its tests may use this
-from typing import Tuple
-
-import torch
-
-
-def sparse_block_diag(*sparse_tensors: torch.Tensor) -> torch.Tensor:
-    """
-    Construct a block-diagonal sparse matrix from COO/CSR inputs.
-
-    Builds a block-diagonal sparse tensor from a sequence of **2D** sparse tensors
-    that are all in the same layout (either COO or CSR). The result has blocks
-    placed on the diagonal and zeros elsewhere, analogous to :func:`torch.block_diag`
-    but for sparse inputs.
-
-    Parameters
-    ----------
-    *sparse_tensors : torch.Tensor
-        Variable number of 2D sparse tensors (all COO or all CSR). Each tensor must
-        have exactly 2 sparse dimensions and 0 dense dimensions.
-
-    Returns
-    -------
-    torch.Tensor
-        A sparse tensor in the same layout as the inputs with shape
-        ``(sum_i n_i, sum_i m_i)``, where each block ``i`` has shape ``(n_i, m_i)``.
-
-    Raises
-    ------
-    TypeError
-        If any input is not a :class:`torch.Tensor`.
-    ValueError
-        If no tensors are provided; if layouts are mixed; or if any input does not
-        have exactly 2 sparse dims and 0 dense dims.
-
-    Notes
-    -----
-    The resulting block structure is
-
-    .. code-block:: text
-
-        [A₁  0   0  ... 0 ]
-        [0   A₂  0  ... 0 ]
-        [0   0   A₃ ... 0 ]
-        [⋮   ⋮   ⋮  ⋱  ⋮ ]
-        [0   0   0  ... Aₙ]
-
-    Offsets for each block are computed using **cumulative** row/column sizes of
-    all preceding blocks (not simply ``i * size``), so inputs may have different
-    shapes.
-
-    Examples
-    --------
-    COO inputs:
-
-    >>> import torch
-    >>> from torchsparsegradutils.utils import sparse_block_diag
-    >>> A = torch.sparse_coo_tensor(torch.tensor([[0, 1], [0, 1]]), torch.tensor([1., 2.]), size=(2, 2))
-    >>> B = torch.sparse_coo_tensor(torch.tensor([[0], [0]]), torch.tensor([3.]), size=(1, 1))
-    >>> C = sparse_block_diag(A, B)
-    >>> C.shape
-    torch.Size([3, 3])
-    >>> C.layout
-    torch.sparse_coo
-
-    CSR inputs:
-
-    >>> A_csr = A.to_sparse_csr()
-    >>> B_csr = B.to_sparse_csr()
-    >>> D = sparse_block_diag(A_csr, B_csr)
-    >>> D.layout
-    torch.sparse_csr
-
-    See Also
-    --------
-    torch.block_diag : Dense block-diagonal construction for dense inputs.
-    stack_csr : Stack CSR matrices along a new batch dimension.
-    """
-    # ---- validation ----
-    for i, t in enumerate(sparse_tensors):
-        if not isinstance(t, torch.Tensor):
-            raise TypeError(f"TypeError: expected Tensor as element {i} in argument 0, but got {type(t).__name__}")
-
-    if len(sparse_tensors) == 0:
-        raise ValueError("At least one sparse tensor must be provided.")
-
-    if all(t.layout == torch.sparse_coo for t in sparse_tensors):
-        layout = torch.sparse_coo
-    elif all(t.layout == torch.sparse_csr for t in sparse_tensors):
-        layout = torch.sparse_csr
-    else:
[…]
