### 2026-07-17 f3356cb
docs: rewrite README + docs for the CUDA-native package
spec/commit.md Phase 4 commit 22 — the final migration commit.

README rewritten around the shipped architecture: two-package model
(pure-Python front package + torchsparsegradutils_cuda backend,
CUDA-required at runtime with the TSGU_DISABLE_CUDA_BACKEND escape
hatch and version handshake), the nine tsgu:: ops behind map.md §1's
public surface, memory-sparse gradients at the input's pattern,
batched-first BatchedCSR internals, torch.library registration.
Removed APIs (CuPy/JAX bridges, block-diag helpers, the four
autograd.Function class
[…]
