+++ docs/source/api/generated/torchsparsegradutils.sparse_bidir_logsumexp.rst
@@ -1,0 +1,6 @@
+﻿torchsparsegradutils.sparse\_bidir\_logsumexp
+=============================================
+
+.. currentmodule:: torchsparsegradutils
+
+.. autofunction:: sparse_bidir_logsumexp
