+++ docs/source/index.rst
@@ -9,38 +9,50 @@ torchsparsegradutils Documentation
    :target: https://github.com/cai4cai/torchsparsegradutils?tab=Apache-2.0-1-ov-file#readme
    :alt: License
 
-.. image:: https://img.shields.io/badge/code%20style-black-000000.svg
-   :target: https://github.com/psf/black
-   :alt: Code Style: Black
+**torchsparsegradutils** provides sparse autograd utilities for PyTorch,
+implemented on native CUDA kernels (the ``tsgu::`` op set). Every operation
+computes gradients with respect to a sparse input at that input's sparsity
+pattern, in that input's layout — never by materialising a dense gradient.
+The package fills long-standing gaps in PyTorch's sparse ecosystem: sparse
+gradients for matmul and linear solves, and a native sparse ``logsumexp``.
 
-**torchsparsegradutils** is a comprehensive collection of utility functions to work with PyTorch sparse tensors,
-ensuring memory efficiency and supporting various sparsity-preserving tensor operations with automatic differentiation.
-This package addresses fundamental gaps in PyTorch's sparse tensor ecosystem, providing essential operations that
-preserve sparsity in gradients during backpropagation.
+Key Features
+============
 
-🚀 Key Features
-===============
+Native CUDA kernels
+-------------------
+
+- All core ops route to custom CUDA kernels registered as ``torch.library``
+  custom ops — ``torch.compile``-compatible and ``opcheck``-tested.
+- Batched-first: one leading batch axis, with ragged specified-entry counts
+  per batch item supported through an internal ``BatchedCSR`` descriptor.
+- CUDA-required at runtime: the compiled backend package
+  ``torchsparsegradutils_cuda`` registers the kernels at import (see
+  :doc:`installation`).
 
 Core Sparse Operations with Sparse Gradient Support
 ----------------------------------------------------
 
-**Memory-Efficient Sparse Matrix Multiplication**
-
-- :func:`~torchsparsegradutils.sparse_mm`: Memory-efficient sparse matrix multiplication with batch support
-- Preserves sparsity in gradients during backpropagation
-- Workaround for `PyTorch issue #41128 <https://github.com/pytorch/pytorch/issues/41128>`_
-- Supports both COO and CSR formats with optional batching
-
-**Sparse Linear System Solvers**
-
-- :func:`~torchsparsegradutils.sparse_triangular_solve`: Sparse triangular solver with batch support
-- :func:`~torchsparsegradutils.sparse_generic_solve`: Generic sparse linear solver with pluggable backends
-- :func:`~torchsparsegradutils.sparse_generic_lstsq`: Generic sparse linear least-squares solver
+- :func:`~torchsparsegradutils.sparse_mm`: sparse × dense matrix
+  multiplication with batch support and sparse gradients
+- :func:`~torchsparsegradutils.sparse_logsumexp`: sparse-aware
+  ``log-sum-exp`` reduction mirroring ``torch.logsumexp``
+- :func:`~torchsparsegradutils.sparse_bidir_logsumexp`: row and column
+  ``log-sum-exp`` in one fused kernel traversal
+- :func:`~torchsparsegradutils.sparse_triangular_solve`: sparse triangular
+  solver with batch support
+- :func:`~torchsparsegradutils.sparse_generic_solve`: generic sparse linear
+  solver with pluggable iterative backends
+- :func:`~torchsparsegradutils.sparse_generic_lstsq`: generic sparse linear
+  least-squares solver
+- :func:`~torchsparsegradutils.segment_mm` /
[…]
