+++ docs/source/installation.rst
@@ -1,86 +1,108 @@
 Installation
 ============
 
-PyPI Installation
-------------------
+torchsparsegradutils ships as two packages: the pure-Python front package
+(``torchsparsegradutils`` — public API, op schemas, dispatch, host-side
+composites) and the CUDA backend (``torchsparsegradutils_cuda`` — the
+compiled ``tsgu::`` kernels). The front package installs anywhere; **the
+backend, an NVIDIA GPU, and CUDA are required to actually run the ops**.
 
-The easiest way to install torchsparsegradutils is from PyPI:
+Front package
+-------------
+
+Install from PyPI:
 
 .. code-block:: bash
 
    pip install torchsparsegradutils
 
-This will install the package with its core dependencies.
+This installs a pure-Python wheel with its core dependencies — no compiler
+involved.
+
+CUDA backend (required at runtime)
+----------------------------------
+
+The core ops are CUDA-only by design: no CPU implementation ships, and
+calling an op without a backend (or with CPU tensors) raises a clear error.
+That error is intentional, not an installation failure.
+
+The backend is built with `Hugging Face kernel-builder
+<https://huggingface.co/docs/kernels>`_ from the ``cuda/`` directory of the
+repository and distributed through the Hugging Face Hub (kernel-builder's Nix
+build produces the binaries; its ``build-and-upload`` flow publishes them as
+a Hub kernel repository). Prebuilt backend wheels are a post-migration item
+and do not exist yet — until then, either pull the published kernel from the
+Hub or build ``cuda/`` locally with kernel-builder.
+
+At import, the front package probes for ``torchsparsegradutils_cuda`` and
+validates a version handshake (``__backend_api_version__``); mismatched
+packages refuse to wire together rather than failing inside a kernel.
+
+Import-only workflows
+~~~~~~~~~~~~~~~~~~~~~
+
+For workflows that only need to *import* the package — docs builds, CPU-only
+CI that collects tests, introspection — set:
+
+.. code-block:: bash
+
+   export TSGU_DISABLE_CUDA_BACKEND=1
+
+This skips the backend probe. ``import torchsparsegradutils`` then works
+without a GPU; calling any kernel-backed op still raises.
 
 Requirements
 ------------
 
-Core Requirements
-~~~~~~~~~~~~~~~~~
-
 - Python >= 3.10
 - PyTorch >= 2.5
+- NVIDIA GPU + CUDA for anything beyond importing the package
 
 Development Installation
 ------------------------
 
-For development or to get the latest features:
+The repository is managed with `uv <https://docs.astral.sh/uv/>`_:
 
 .. code-block:: bash
 
    git clone https://github.com/cai4cai/torchsparsegradutils.git
    cd torchsparsegradutils
-   pip install -e .
-
-To install development dependencies (via `uv <https://docs.astral.sh/uv/>`_):
-
-.. code-block:: bash
-
    uv sync --group dev
 
 Verification
 ------------
 
-To verify your installation:
+To verify the front package and backend are wired together:
 
 .. code-block:: python
 
-   import torchsparsegradutils as tsgu
-   print(tsgu.__version__)
-
-   # Test basic functionality
    import torch
+   import torchsparsegradutils
+   from torchsparsegradutils._dispatch import backend_available
+
+   print(backend_available())  # True iff a compatible CUDA backend is loaded
+
+   # Exercise an op end-to-end (requires the backend and a CUDA device)
    from torchsparsegradutils import sparse_mm
 
-   # Create a simple sparse matrix
    indices = torch.tensor([[0, 1], [1, 0]])
[…]
