### 2026-07-17 741fc96
docs: kernel best-practices research + perf-phase playbook
kernel_best_practices.md: sourced optimization playbook per kernel family
(33-agent web research + 8-agent PTX round; hardware table verified by
device query; two claims settled by local SASS probes). PERF_PLAN.md:
12-commit perf-phase playbook on branch cuda-rewrite-ptx — Opus workers,
advisor-reviewed, SASS-audit + benchmark gates per commit.

Co-Authored-By: Claude Fable 5 <noreply@anthropic.com>
