+++ kernel_best_practices.md
@@ -1,0 +1,444 @@
+# Kernel Best Practices
+
+Optimization playbook for the tsgu CUDA kernels, to be applied **after** the migration
+lands: each section is a priority-ordered list of techniques for one kernel family from
+[spec/kernels.md](spec/kernels.md), with sources and decision rules for where each
+technique wins. Every claim below traces to a cited source or to the verified hardware
+table; unsourced design reasoning is explicitly marked as such.
+
+Provenance: assembled 2026-07-17 from a 33-agent web-research run (20 searchers × 3
+angles per kernel, per-kernel synthesis, completeness critique + gap-fill), plus an
+8-agent curated-source round on the PTX/microarchitecture layer (WebFetch-only,
+fetch failures flagged inline rather than filled from memory) with two claims
+settled by local SASS probes on the dev GPU. Spec
+acceptance bars (beat cuSPARSE / pytorch_scatter, deterministic-path requirement,
+no dense materialisation, O(nse) workspace) are taken as fixed constraints, not
+re-litigated here.
+
+## Target hardware (verified by direct device query, 2026-07-17)
+
+All sizing decisions below key off these numbers, queried from the actual dev GPU via
+`torch.cuda.get_device_properties` (they override any conflicting figure in cited sources):
+
+| Property | RTX A1000 Laptop GPU (GA107, sm_86) |
+|---|---|
+| SMs | **16** |
+| L2 cache | **1.5 MB** |
+| DRAM bandwidth | **~176 GB/s** (128-bit bus, 5.5 GHz memory clock, DDR) |
+| Shared memory | **99 KB max per block (opt-in), 100 KB carveout per SM** (128 KB unified L1+smem) |
+| Registers | 64 K × 32-bit per SM (256 KB) |
+| Max resident warps/SM | **48** (1536 threads) |
+| fp64 rate | **1/64** of fp32 (GA10x: 2 FP64 units per SM) — correctness-only |
+| Tensor cores | yes (fp16/bf16 → fp32 accumulate); `cp.async` supported |
+
+Beware of sources quoting A100 (sm_80) numbers — 164 KB shared memory, 40 MB L2, fast fp64
+do **not** apply here. A laptop part also throttles: lock clocks before benchmarking
+(cross-cutting §10).
+
+## SDDMM (sampled dense-dense matmul)
+
+**What dominates performance.** SDDMM at our shapes is memory-bound gather: every specified entry (i,j) reads two dense rows G[i,:] and B[j,:] of length p and writes one scalar, so arithmetic intensity is ~1 FLOP per 4 bytes read and the ~192 GB/s DRAM of the A1000 is the wall. The three levers, in order, are (1) making the row gathers coalesced and vectorized, (2) not re-reading rows that many entries share (cache/shared-memory reuse — critical with a small L2), and (3) keeping enough independent reductions in flight that the memory latency is hidden despite the dot-product's serial tail. For small p the reduction itself is nearly free and everything is traffic; only at large p does reduction parallelism (splitting p across more lanes/warps) become a first-order concern. fp32 is the performance path; fp64 (1/64 rate on GA10x sm_86) will additionally be FMA-limited at larger p, which makes the traffic optimizations below matter even more since there is no compute headroom to hide behind.
+
+### Best practices, in priority order
+
[…]
