+++ PERF_PLAN.md
@@ -1,0 +1,107 @@
+# Performance Phase Playbook — cuda-rewrite-ptx
+
+Successor to the 22-commit migration playbook. Migration is complete; this phase
+implements [kernel_best_practices.md](kernel_best_practices.md) end to end.
+Source of truth for *what* each commit does is that document's sections; this file
+is the *who/when/where/gates*.
+
+## Execution model
+
+**Advisor: Fable (this session's model).** The advisor writes no kernel code by
+default. The advisor: assembles each worker prompt (best-practices section +
+spec contract refs + acceptance bar + baseline numbers), reviews every diff and
+SASS-audit output, runs the gates, owns the benchmark ledger, and decides
+pass/retry/escalate.
+
+**Workers: Opus agents (`model: opus`), one commit each.** Every worker prompt
+carries, verbatim:
+
+- the relevant `kernel_best_practices.md` section(s) and hardware table,
+- the acceptance bar and the baseline benchmark rows it must beat,
+- the standing warning (recurring mistake, seen twice in the migration): **no
+  fixed-width predicated unrolls, no serial per-entry scalar loads** — vectorize
+  loads, verify in SASS,
+- the gate list (below) the worker must run before reporting done.
+
+**Escalation rule (amended from migration precedent):** if a worker misses its
+perf bar, the advisor diagnoses (SASS + Nsight, not guesswork) and re-briefs a
+fresh Opus worker with the diagnosis. Only after a second miss does the advisor
+consider intervening directly — this phase is Opus-does-everything by design.
+
+## Worktree
+
+All work happens in a new worktree **`cuda-rewrite-ptx`**, branched off
+`cuda-rewrite`. **Not created yet — creation is Step 0 at kickoff:**
+
+```bash
+# from the main checkout
+git -C .worktrees/cuda-rewrite add kernel_best_practices.md PERF_PLAN.md
+git -C .worktrees/cuda-rewrite commit -m "docs: kernel best-practices research + perf-phase playbook"
+git worktree add .worktrees/cuda-rewrite-ptx -b cuda-rewrite-ptx cuda-rewrite
+```
+
+The docs commit must land on `cuda-rewrite` **before** the worktree is created —
+both files are currently untracked, and untracked files do not follow a new
+worktree. `cuda-rewrite` then stays frozen for the maintainer RFC; all perf
+commits go to `cuda-rewrite-ptx`.
+
+Machine constraints carry over unchanged: nix builds `--max-jobs 1 --cores 6`,
+repoint both `.pth` files after every `nix build`, disk is ~90% full (a worktree
+checkout is cheap but nix build artifacts are not — prune `result` symlinks and
+old store paths between kernels if needed), GPU clocks cannot be locked (no
+root) so every benchmark records the variance protocol from cross-cutting §10.
+
+## Standing gates (every commit)
+
+1. `tox -e gpu` green, including the run-twice determinism tests.
+2. SASS audit script (commit 1) clean for the touched kernels: no `LDL`/`STL`
+   spills, expected `LDG.E.128`/`LDGSTS` present, no `BAR` in spin loops.
+3. Benchmark row(s) recorded with backend provenance (`custom` vs
+   `vendor-scaffold`); result JSONs never committed.
+4. A perf commit that regresses any other kernel's benchmark does not land.
+5. Deterministic path exists and is exercised for every touched op
+   (raise-don't-fallback, per spec).
+
+## Commit sequence
+
+Ordering logic: verification tooling first (nothing lands unverified), then the
+open acceptance bar (grouped_gemm — the only release blocker), then cheap
[…]
