### 2026-07-28 edadb5d
Fix README rand_sparse_tri quickstart example (#92)
