### 2026-07-28 68968df
Clarify README sparsity comment
