### 2026-07-28 8e2f890
Fix unit triangular documentation
