### 2026-07-29 6f25990
Sparsity comment made clear
