### 2026-07-29 b96fc72
Fix README rand_sparse_tri QuickStart example (#92)  (#93)
* Fix README rand_sparse_tri quickstart example (#92)

* Clarify README sparsity comment

* Fix unit triangular documentation

* Sparsity comment made clear
